### 2024-04-05 #6584 @editorialbot
[PRE REVIEW]: CoastSeg: an accessible and extendable hub for satellite-derived-shoreline (SDS) detection and mapping
**Submitting author:** @2320sharon (<a href="http://orcid.org/0000-0001-6513-9132">Sharon Fitzpatrick</a>)
**Repository:** https://github.com/SatelliteShorelines/CoastSeg
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.0
