### 2024-04-05 8931945
#240 preview images button now account for months selected
