### 2024-04-05 bd47c99
#240 fix load_settings so months_list is not filtered out
