### 2024-04-05 dc86aa6
#240 remove prints in settings_UI
