### 2024-04-05 bb664ab
#240 update dependencies to coastsat-package>=0.1.61
