### 2024-04-06 31894cd
#240 add an exception to preview images if start date is after end date
