### 2024-04-08 ef4f6c2
remove loading all the months by default from load_settings in common.py
