### 2024-04-08 c15b52f
update to use coastsat-package>=0.1.62
