### 2024-04-08 7e945ab
v1.2.1  closes #240 #234 #226 Make downloads resumeable and add monthly filter and control dates & satellites for extracted shorelines
