+++ docs/index.md
@@ -1,0 +1,373 @@
+# CoastSeg
+
+<!--  [![image](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/2320sharon/CoastSeg/blob/main/coastseg_for_google_colab.ipynb) -->
+
+[![image](https://img.shields.io/pypi/v/coastseg.svg?color=%23ec3dc8)](https://pypi.python.org/pypi/coastseg)
+</br>
+[![image](https://github.com/SatelliteShorelines/CoastSeg/actions/workflows/pip_install_e.yml/badge.svg)](https://github.com/SatelliteShorelines/CoastSeg/actions)
+
+[![Last Commit](https://img.shields.io/github/last-commit/SatelliteShorelines/CoastSeg)](https://github.com/Doodleverse/segmentation_gym/commits/main)
+[![Maintenance](https://img.shields.io/badge/Maintained%3F-yes-green.svg)](https://github.com/SatelliteShorelines/CoastSeg/graphs/commit-activity)
+[![Wiki](https://img.shields.io/badge/wiki-documentation-forestgreen)](https://github.com/SatelliteShorelines/CoastSeg/wiki)
+![GitHub](https://img.shields.io/github/license/Doodleverse/segmentation_gym)
+[![Wiki](https://img.shields.io/badge/discussion-active-forestgreen)](https://github.com/SatelliteShorelines/CoastSeg/discussions)
+
+![Python](https://img.shields.io/badge/python-3670A0?style=for-the-badge&logo=python&logoColor=ffdd54)
+![TensorFlow](https://img.shields.io/badge/TensorFlow-%23FF6F00.svg?style=for-the-badge&logo=TensorFlow&logoColor=white)
+![Keras](https://img.shields.io/badge/Keras-%23D00000.svg?style=for-the-badge&logo=Keras&logoColor=white)
+
+<img src="https://user-images.githubusercontent.com/61564689/212394936-263ec9fc-fb82-45b8-bc79-bc57dafdae73.gif" width="350" height="350">
+
+### Table of Contents
+
+- [Installation Instructions](#installation-instructions)
+- [Getting Started](#getting-started)
+- [CoastSeg Update Guide](#coastseg-update-guide)
+
+### Useful Links
+
+- [Wiki](https://github.com/SatelliteShorelines/CoastSeg/wiki)
+- [Discussion](https://github.com/SatelliteShorelines/CoastSeg/discussions)
+- [Contribution Guide](https://github.com/Doodleverse/CoastSeg/wiki/13.-%5BADVANCED%5D-Contribution-Guide)
+
+## What is CoastSeg?
+
+Coastseg stands for Coastal Segmentation, it is an interactive jupyter notebook for downloading satellite imagery with [CoastSat](https://github.com/kvos/CoastSat) and applying Doodleverse/[Zoo](https://github.com/Doodleverse/segmentation_zoo) image segmentation models to satellite imagery. CoastSeg provides an interactive interface for drawing Regions of Interest (ROIs) on a map, downloading satellite imagery, loading geojson files, extracting shorelines from satellite imagery, and much more. In a nutshell, it is ...
+
+- ... an easy way to download time-series of cloud-masked and post-processed satellite imagery anywhere in the world
+- ... a mapping extension for [CoastSat](https://github.com/kvos/CoastSat) using [Segmentation Zoo](https://github.com/Doodleverse/segmentation_zoo) models.
+- ... an interactive interface to download satellite imagery using CoastSat from Google Earth Engine
+- ... an interactive interface for extracting shorelines from satellite imagery
+- ... an interactive interface to apply segmentation models to satellite imagery
+
+In more detail, CoastSeg represents the following advances in satellite-derived shoreline workflows:
+
[…]
