+++ mkdocs.yml
@@ -1,0 +1,21 @@
+site_name: CoastSeg
+site_url: https://SatelliteShorelines.github.io/CoastSeg/
+
+theme:
+  name: material
+  palette:
+    # Palette toggle for light mode
+    - scheme: default
+      toggle:
+        icon: material/brightness-7
+        name: Switch to dark mode
+
+    # Palette toggle for dark mode
+    - scheme: slate
+      toggle:
+        icon: material/brightness-4
+        name: Switch to light mode
+
+nav:
+  - Home: index.md
+  - About: about.md
