+++ docs/install.md
@@ -1,0 +1,239 @@
+## Regular users on Windows and Linux
+
+Please refer to the [installation guide](https://github.com/Doodleverse/CoastSeg#installation-instructions) on the README
+
+## Mac users
+
+CoastSeg requires Tensorflow (TF), which doesn't play nicely with Mac. We advise you to use either Linux or Windows, if you can. We cannot troubleshoot Mac installations, but we can offer the following advice:
+
+- TF for mac has its own instructions: https://developer.apple.com/metal/tensorflow-plugin/
+- New Mac silicon runs TF, (and has its own TF branch), but the old intel Mac chips might not work with parts of TF.
+- We are not sure if TF is compatible with M2 macs
+- Our continuous integration tests check only the 'latest' version of Mac OS.
+- If you get a working installation on Mac, please let us know, and we can edit these pages to communicate better advice. Thanks in advance
+
+## Users working over secure network
+
+First, some notes:
+
+- ⚠️ The CoastSeg team is aware of issues with installing CoastSeg's conda dependencies ⚠️
+- ⚠️The following instructions are valid as of 2023-09-12. If this proposed solution fails for you, please create a new [Issue](https://github.com/Doodleverse/CoastSeg/issues) ⚠️)
+- ⚠️ Where possible, we advise you install the initial Coastseg conda environment not over a secure network. If and _ONLY_ if that is not possible, please follow the instructions posted below carefully
+- If you are running these commands on a secure network verify you are connected to your VPN before running the following commands
+- if you get an SSL error while running any of the `conda install` commands try the following command
+- ONLY RUN THE PIP INSTALL COMMANDS AFTER CREATING A PIP.INI file [Go to Phase 1: Create a pip.ini File](#phase-1-create-a-pipini-file)
+
+### Quick Start
+
+Second, you must temporarily (for a few minutes) disable SSL verification. This should be done only once, when setting up the initial conda environment. If you find yourself having to do this for other instructions listed below and elsewhere, something is wrong and you should either start again from scratch, or raise an Issue if there is a legitimate bug. Coastseg developers are not liable for any unintended effects of ignoring the advice posted here, i.e. leaving `ssl_verify` to be `False` outside of the specific and unique case explained here.
+
+```
+conda config --set ssl_verify False
+```
+
+Now, you can install the rest of the dependencies. You must do this immediately.
+
+```
+conda install -c conda-forge geopandas jupyterlab -y
+pip install coastseg
+pip uninstall h5py -y
+conda install -c conda-forge h5py -y
+```
+
+Finally, re-enable SSL verification. You _must_ do this _immediately_ after installing the conda packages above. If any of the above commands fail for whatever reason, you must also run this command to re-enable SSL. Coastseg developers are not liable for any unintended effects of ignoring the advice posted here, i.e. leaving `ssl_verify` to be `False` outside of the specific and unique case explained here.
+
+```
+conda config --set ssl_verify True
+```
+
+### Detailed Guide
+
+#### Phase 0: Allow Git to work over a secure network
+
+1. Configure git to use ssl
+
+- Make sure to replace `"C:/Users/ulastname/Documents/DOIRootCA2.cer"` with your user name
+
+```
[…]
