+++ .github/workflows/ci.yml
@@ -19,8 +19,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install mkdocs
-          pip install mkdocs-material
+          pip install -r requirements_docs.txt
 
       - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v4

[…]
