+++ scripts/apply_tidal_correction.py
@@ -11,7 +11,6 @@ import geopandas as gpd
 import numpy as np
 import pandas as pd
 import pyproj
-import pyTMD.constants
 import pyTMD.io
 import pyTMD.io.model
 import pyTMD.predict
