### 2024-04-10 aeab264
fix typo in ci.yml to release website on refs/heads/deploy_site
