+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v4
         # Using the depol
-        if: github.ref == 'refs/heads/depoly_site'
+        if: github.ref == 'refs/heads/deploy_site'
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./docs
