### 2024-04-10 180d88f
update ci.yml to deploy with force
