### 2024-04-10 c93faa9
fix typo in how-to-download
