+++ .github/workflows/ci.yml
@@ -1,33 +1,28 @@
-name: Deploy MkDocs
-
+name: ci
 on:
   push:
     branches:
-      - deploy_site # Or your default branch
-
+      - main
+permissions:
+  contents: write
 jobs:
   deploy:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-
-      - name: Setup Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.x"
-
-      - name: Install dependencies
+      - name: Configure Git Credentials
         run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements_docs.txt
-
-      - name: Build and Deploy to GitHub Pages
-        run: mkdocs gh-deploy --force
-
-      # - name: Deploy to GitHub Pages
-      #   uses: peaceiris/actions-gh-pages@v4
-      #   # Using the depol
-      #   if: github.ref == 'refs/heads/deploy_site'
-      #   with:
-      #     github_token: ${{ secrets.GITHUB_TOKEN }}
-      #     publish_dir: ./docs
+          git config user.name github-actions[bot]
+          git config user.email 41898282+github-actions[bot]@users.noreply.github.com
+      - uses: actions/setup-python@v5
+        with:
+          python-version: 3.x
+      - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV
+      - uses: actions/cache@v4
+        with:
+          key: mkdocs-material-${{ env.cache_id }}
+          path: .cache
+          restore-keys: |
+            mkdocs-material-
+      - run: pip install mkdocs-material
+      - run: mkdocs gh-deploy --force
