### 2024-04-11 f8c87f6
update to use dark mode by default and add how to extract shoreline area guide
