+++ docs/how-to-extract-shoreline-area.md
@@ -1,0 +1,30 @@
+# How to Use the Shoreline Extraction Area
+
+The shoreline extraction area is used to keep only the extracted shorelines within the region of the Shoreline Extraction Area.
+You can draw multiple Shoreline Extraction Area polygons if your shoreline is complex or if you have multiple ROIs.
+
+## Steps
+
+1. Navigate to `Draw Controls` and click `Shoreline Extraction Area`
+   </br>
+
+![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/a8aa2126-a706-46fd-9460-ef3b9297db13)
+
+2. Click either the polygon or rectangle drawing tool on the map
+
+![draw_controls_shape](https://github.com/Doodleverse/CoastSeg/assets/61564689/5acba879-a527-45ad-981f-a2a7956fb47e) 3. Draw the Shoreline Extraction Area on the map (it will be shown in pink)
+
+- draw as many of these as you like
+
+![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/215d4f1b-9ec2-43c9-9c7c-f75dd2757c82)
+
+4. (Optional) If you make a mistake navigate to `Remove Feature from Map` and select `Shoreline Extraction Area` from the dropdown to remove the Shoreline Extraction Area from the map
+
+![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/cfc00702-cfff-4c30-b129-988681db1cc0)
+
+5. (Optional) To switch back to drawing bounding boxes navigate to `Draw Controls` and click `Bounding Box`
+6. Click Extract Shorelines
+
+- The Shoreline Extraction Area you drew will now be used to extract only the shorelines in this region
+
+![shoreline_extraction_demo_screenshot](https://github.com/Doodleverse/CoastSeg/assets/61564689/7b925a9a-c812-48ce-bbc4-2902ba4845a2)
