### 2024-04-11 5f6cd21
remove debug from index.md
