+++ .github/workflows/ci.yml
@@ -2,7 +2,7 @@ name: ci
 on:
   push:
     branches:
-      - main
+      - deploy_site
 permissions:
   contents: write
 jobs:
