+++ docs/about.md
@@ -1,1 +1,0 @@
-# Welcome to About

[…]
