### 2024-04-12 7db3d1b
update install guide
