### 2024-04-15 09b1506
refactor download tide model into 2 tutorials
