+++ docs/How-to-Tidally-Correct.md
@@ -1,0 +1,182 @@
+# ⚠️ Before you begin ⚠️
+
+Make sure you have downloaded the FES2014 tide model following the guide 'How to Download the Tide Model'
+
+## ⚠️ Important: Read Before Using Tide Correction Button
+
+If the tide model was NOT downloaded to `CoastSeg/tide_model` then the tide correction button will NOT work. The tide correction button will try to load the tide model from `<CoastSeg location>/tide_model` and an error will occur. Instead follow the instructions in at [Option 2: Tide Correction Script](#option-2-tide-correction-script) to use the script `apply_tidal_correction.py`
+
+## Option 1: Tide Correction Button Method
+
+Ever since coastseg 1.1.0 released you can now use the `correct tides` button in coastseg to automatically correct the timeseries. This button will automatically find which files need to be tidally corrected and apply the tidal correction for you.
+
+⚠️ <b>Important</b> ⚠️
+
+If the tide model was NOT downloaded to `CoastSeg/tide_model` then the tide correction button will NOT work. The tide correction button will try to load the tide model from `<CoastSeg location>/tide_model` and an error will occur. Instead follow the instructions in at [Option 2: Tide Correction Script](#option-2-tide-correction-script)
+
+### Steps
+
+**WARNING**
+You must have downloaded the tide model to the `tide_model` folder within coastseg for this button to work correctly.
+
+1.**Enter the beach slope**
+
+- This is the beach slope in meters for the ROI you have selected.
+
+  2.**Enter the beach elevation**
+
+- This is the beach elevation in meters relative to MSL(Mean Sea Level) for the ROI you have selected.
+
+  3.**Select the ROIs' ids to tidally correct**
+
+  4.**Click compute tides and wait a few minutes.**
+
+If you encounter any issues or have any questions please submit an issue at [CoastSeg Issues](https://github.com/Doodleverse/CoastSeg/issues/new/choose)
+
+![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/c72ee626-86fd-46ae-98d0-0921b0082d8b)
+
+## Option 2: Tide Correction Script
+
+⚠️ If the tide model was NOT downloaded to `CoastSeg/tide_model`, then you must use the tide correction script.
+
+### Example 1: Apply Tide Correction to a session and provide tide model location
+
+1.**Choose a session**
+
+- In this example I will be using session `dnz3_extract_shorelines_10_yr` and using ROI `ID_dnz3_datetime12-22-23__09_10_44`
+- Example: `C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44`
+
+  2.**Get the location of the config_gdf.geojson file**
+
+- This parameter is used after the `-C `
+- Example: `-C "C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44\config_gdf.geojson"`
+
+  3.**Get the location of the transect_time_series.csv file**
+
+- This parameter is used after the `-T `
+- Example: `-T "C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44\transect_time_series.csv"`
+
+  4.**Get the location of the tide_model**
+
+- This parameter is used after the `-M`
+- Example: `-M "C:\development\doodleverse\coastseg\CoastSeg\tide_model"`
+
[…]
