+++ mkdocs.yml
@@ -61,6 +61,7 @@ nav:
       - How to Use Extract Shoreline Area: how-to-extract-shoreline-area.md
       - How to Extract Shorelines: extract_shorelines.md
       - How to Download the Tide Model: How-to-Download-Tide-Model.md
+      - How to Tidally Correct: How-to-Tidally-Correct.md
       - How to Create Reference Shorelines & Transects: How-to-Create-Reference-Shorelines-&-Transects.md
       - How to Filter Bad Imagery: How-to-Filter-Out-Bad-Imagery.md
       - Settings Guide: settings.md
