### 2024-04-15 fd209f9
Add apply tide correction script to scripts section
