+++ docs/Apply-Tidal_Correction_Script.md
@@ -1,0 +1,147 @@
+# ⚠️ Before you begin ⚠️
+
+Make sure you have downloaded the FES2014 tide model following the guide 'How to Download the Tide Model'
+
+## Apply Tide Correction Script
+
+### Example 1: Apply Tide Correction to a session and provide tide model location
+
+1.**Choose a session**
+
+- In this example I will be using session `dnz3_extract_shorelines_10_yr` and using ROI `ID_dnz3_datetime12-22-23__09_10_44`
+- Example: `C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44`
+
+  2.**Get the location of the config_gdf.geojson file**
+
+- This parameter is used after the `-C `
+- Example: `-C "C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44\config_gdf.geojson"`
+
+  3.**Get the location of the transect_time_series.csv file**
+
+- This parameter is used after the `-T `
+- Example: `-T "C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44\transect_time_series.csv"`
+
+  4.**Get the location of the tide_model**
+
+- This parameter is used after the `-M`
+- Example: `-M "C:\development\doodleverse\coastseg\CoastSeg\tide_model"`
+
+  5.Select a Beach Elevation
+
+- This is the elevation of the beach in meters relative to MSL(Mean Sea Level) for the ROI you have selected
+- Example: `-E 3.2`
+
+  6.Select a Beach Slope
+
+- This is the slope of the beach in meters for the ROI you have selected
+- Example: `-S 2.1`
+
+  7.**Assemble the command**
+
+- Assemble all commands into a single line. I recommended using notepad to combine all the commands into a single line
+
+```
+python apply_tidal_correction.py -E 3.2 -S 2.1 -C
+"C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44\config_gdf.geojson" -T
+"C:\development\doodleverse\coastseg\CoastSeg\sessions\dnz3_extract_shorelines_10_yr\ID_dnz3_datetime12-22-23__09_10_44\transect_time_series.csv"
+-M "C:\development\doodleverse\coastseg\CoastSeg\tide_model"
+```
+
+8.**Run the command**
+
+- It takes a few minutes to predict the tides
+
+![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/13582ba7-63c6-46d4-93c1-aa68db23626c)
+
+### Parameters:
+
+1.**-C (or -c) [CONFIG_FILE_PATH]**
+
+- Description: Path to the configuration file.
+- Example: `-C "path_to_config_file"`
+- `CONFIG_FILE_PATH`: Path to the configuration file. This is the `config_gdf.geojson` in the session directory generated when shorelines were extracted
+
+  2.**-T (or -t) [RAW_TIMESERIES_FILE_PATH]**
+
+- Description: Path to the raw timeseries file.
+- Example: `-T "path_to_timeseries_file"`
+- `RAW_TIMESERIES_FILE_PATH`: Path to a csv file containing the time series created by extracting shorelines with coastseg. This timeseries
+  represents the intersection of each transect with the shoreline extracted for a particular date and time. It is not tidally corrected.
+
+  3.**-E (or -e) [REFERENCE_ELEVATION]**
+
+- Description: Set the reference elevation value. This is a float number.
+- Example: `-E 3`
+
[…]
