+++ mkdocs.yml
@@ -71,6 +71,7 @@ nav:
   - Zoo: models.md
   - Scripts:
       - How to Use Scripts: How-to-Use-Scripts.md
+      - How to Use Apply Tidal Correction Script: Apply-Tidal_Correction_Script.md
       # - CoastSeg: scripts-coastseg.md
       # - CoastSat: scripts-coastsat.md
       # - Zoo: model.md
