### 2024-04-15 b863b8c
remove merge sessions from how to use scripts
