+++ docs/How-to-Use-Scripts.md
@@ -8,19 +8,28 @@ Running Python scripts using the Anaconda Prompt can be an efficient way to exec
 
 ## List of Available Scripts:
 
-1. transects_swap_points.py
+1.transects_swap_points.py
+
    - A script that reads a `config_gdf.geojson` and swaps the origin & end point for each transect.
    - The new transects (default name : `reversed_transects.geojson` ) are saved to the scripts directory
-2. shorten_transects.py
+
+2.shorten_transects.py
+
    - A script that reads a geojson file containing transects and can shorten or length each transect depending on the parameters used
    - The new transects (default name : `shortened_transects.geojson` ) are saved to the scripts directory
-3. get_transects_points.py
+
+3.get_transects_points.py
+
    - A script that reads a `config_gdf.geojson` and creates two geojson files `origin_points.geojson` containing the origin point & `end_points.geojson` for each transect.
    - Both of these geojson files are saved to the scripts directory
-4. georeference_jpgs.py
+
+4.georeference_jpgs.py
+
 - A script that takes a list of ROI directories and creates georeferenced jpegs for the "RGB", "NIR", "SWIR", "MNDWI", "NDWI" subdirectories
 - [See a guide on how to use it](#guide-for-script-georeference_jpgspy)
-5. preview_images_for_dates.py
+
+5.preview_images_for_dates.py
+
 - This script prints the available satellite imagery for ROIs given a list of date ranges. It also saves the results to a file named `results.json`
 - [See a guide on how to use it](#guide-for-script-preview_images_for_datespy)
 
@@ -85,126 +94,6 @@ Most well-maintained scripts include a header comment or a docstring at the begi
 ```
 python preview_images_for_dates.py "C:\development\doodleverse\coastseg\CoastSeg\rois.geojson" "2002-01-01,2023-01-15" "1984-02-01,1995-02-15" "2023-03-01,2023-03-15"
 ```
-# Guide for Script merge_sessions.py (under construction 🚧 )
-Use the `merge_sessions.py` script to combine multiple session folder into a single merged session folder. The script reads the information from the session folders you provide, then it merges the geojson files, json files, and csv files respectively together into a single file that is saved to the new merged session folder. 
-
-
- This script combines the following data together:
-1. Extracted shorelines in geojson files
-2. time series csv files
-3. config_gdf.geojson files
-4. shoreline_dict.json
-5. cross_distance.json
-
-⚠️ **Warning** ⚠️
-
-Merged Sessions should NEVER be loaded into CoastSeg.
-
-## Uses Cases
-<table>
-<tr>
-<td>
-
-### Use Case 1: Merge multiple ROIs at different locations together
-- Example: You have multiple ROIs that cover an entire coastline that you want to merge into a single session. When the extracted shorelines are merged together the resulting session contains long continuous shorelines stored in a single file.
-
-<img src="https://github.com/Doodleverse/CoastSeg/assets/61564689/c816db24-86e2-4a30-994e-23e0cab670b8.png" width="385" height="400"/>
-
-</td>
-<td>
-
-### Use Case 2: Merge multiple sessions for a single ROI together
-- Merging multiple date ranges for a single ROI into a single session.
-   - For example: if you have 2 sessions with extracted shorelines and session 1 contains shorelines for dates  2022-2023 and session 2 contains shorelines for dates 2021 - 2022. You can merge them to have a single session with shorelines for dates 2021 - 2023.
-
-<img src="https://github.com/Doodleverse/CoastSeg/assets/61564689/ed04ddf8-ee3a-4d24-a0e5-fe010c51cf54.png" width="385" height="400"/>
-
-</td>
-</tr>
-</table>
-
-### Warning
[…]
