+++ docs/models.md
@@ -1,10 +1,11 @@
-# CoastSeg Models
+# The CoastSeg Zoo Workflow
+
+This workflow can be found in the 'SDS_zoo_classifier.ipynb' notebook. It is still under active development so use it at your own risk.
+
+The models available in this workflow come from [segmentation zoo](https://github.com/Doodleverse/segmentation_zoo) and were trained using [segmentation gym](https://github.com/Doodleverse/segmentation_gym).
+
+# 🚧 Under development 🚧
+
+- todo add explanation of each model
 
 The following image segmentation models are available in CoastSeg to use on downloaded satellite imagery.
-
-## 5 Band Models
-
-1. sat-5band-2-class_7448390
-   Read the full model description on the official Zenodo release
-   Zenodo Release: https://zenodo.org/record/7448390#.Y75TSnbMJg8
-
