### 2024-04-17 6e0b041
new controls to actually filter no data and cloudy images with new coastsat-package 0.1.65
