### 2024-04-18 eda553e
use coastsat-package 0.1.65 use cloud filter to filter downloaded imagery
