### 2024-04-18 #6584 @2320sharon
Hi @kthyng thanks for the update on the review process. I was beginning to think I did something wrong with the submission.

Quick question, the software  in the paper, CoastSeg, has gotten a few minor patches since we submitted to the paper can the reviewers use the newer versions or do they have to use the version that was submitted?
