### 2024-04-18 #6584 @kthyng
They should use the newest version for sure. They would install it however you say to in your docs to make sure they work so just make sure that is consistent with what you want them to do. I can update the version list in this pre-review issue too if you want, but ultimately what matters is what is listed when the submission is published.
