### 2024-04-18 3391239
update preview_images_for_dates.py with better docstrings
