### 2024-04-18 5a3bc40
update the extract shoreline guide
