### 2024-04-18 7d9b56e
simplify the setting
