+++ docs/settings.md
@@ -1,48 +1,122 @@
-# Settings
+# Settings Explained
 
-Use the settings drop down to modify the settings used by CoastSeg to download ROIs, extract shorelines and compute transect and shoreline intersections.
+## Basic Settings
 
-![Settings Demo](https://github.com/SatelliteShorelines/CoastSeg/blob/main/docs/gifs/settings_demo.gif)
+**1.Reference Shoreline Buffer: `max_dist_ref`**
 
-## Settings Explained
+The `max_dist_ref` parameter defines a buffer (in meters) around the reference shoreline, limiting the extraction of shorelines to this region. Shorelines detected outside this buffer are automatically excluded.
+
+##### Examples of Different `max_dist_ref` Settings
+
+|                                                  ❌ Bad `max_dist_ref`                                                  |                                               ✅ Good `max_dist_ref`                                               |
+| :---------------------------------------------------------------------------------------------------------------------: | :----------------------------------------------------------------------------------------------------------------: |
+|                                                  `max_dist_ref`: 100m                                                   |                                                `max_dist_ref`:309m                                                 |
+|                                           Recommend increasing buffer to 150m                                           |                                 Buffer captures entire shoreline no change needed                                  |
+| ![Increased max_dist_ref](https://github.com/Doodleverse/CoastSeg/assets/61564689/62fefaac-c0c9-40c3-89b8-e433e0555f52) | ![Good max_dist_ref](https://github.com/Doodleverse/CoastSeg/assets/61564689/3ccaf660-cb22-4530-bde0-58d6f6932fa3) |
+
+**2.Minimum Shoreline Length (min_length_sl)**
+Defines the shortest length (in meters) of detected shoreline segments. Adjust this setting to filter out noise or small, irrelevant shoreline fragments.
+
+**3.Distance from Clouds (dist_clouds)**
+Specifies the minimum distance (in meters) from cloud-covered areas to avoid false shoreline detections due to cloud interference.
+
+![dist_clouds example](https://github.com/Doodleverse/CoastSeg/assets/61564689/9fe35a16-72b1-4414-bf1a-82898d103fba)
+
+**4.Minimum Beach Area (min_beach_area)**
+Sets the smallest area (in meters²) that can be classified as shoreline, helping to distinguish between true shoreline and scattered noise.
+
+<details>
+<summary>Click to expand: How this works</summary>
+
+During the image classification, some features (for example, building roofs) may be incorrectly labelled as sand. To correct this, all the objects classified as sand containing less than a certain number of connected pixels are removed from the sand class. The default value is 4500 m^2, which corresponds to 20 connected pixels of 15 m^2. If you are looking at a very small beach (<20 connected pixels on the images), try decreasing the value of this parameter.
+
+</details>
+
+![min_beach_area_coastsat_setting](https://github.com/Doodleverse/CoastSeg/assets/61564689/88f6bc9d-9ee9-42ce-98e3-e4693e8046d4)
+
+**5.Cloud Percentage Threshold (cloud_thresh)**
+This threshold sets the maximum allowed percentage of cloud cover in imagery considered for analysis. Any images that have cloud coverage above this percentage will NOT be considered for shoreline extraction
+
+**6.Sand Color `sand_color`**
+The 'sand_color' setting is used to select the best model for classifying sand. You can choose one of the following options `default` to use the default model, `dark` for grey/black sand beaches or 'bright' for white sand beaches. This setting is only used by the coastsat shoreline extraction model not the one's used by coastseg in the `unet` notebook.
+
+- Only change the `sand_color` parameter if you are seeing that with the default the sand pixels are not being classified as sand (in orange).
+- If your beach has dark sand (grey/black sand beaches), you can set this parameter to dark and the classifier will be able to pick up the dark sand.
[…]
