### 2024-04-18 d9ca772
update the settings guide to be specific to extract shorelines
