### 2024-04-18 1e35166
Update the settings guide
