+++ mkdocs.yml
@@ -60,11 +60,11 @@ nav:
       - How to Use Sessions: sessions.md
       - How to Use Extract Shoreline Area: how-to-extract-shoreline-area.md
       - How to Extract Shorelines: extract_shorelines.md
+      - Extract Shorelines Settings Guide: extract_shoreline_settings.md
       - How to Download the Tide Model: How-to-Download-Tide-Model.md
       - How to Tidally Correct: How-to-Tidally-Correct.md
       - How to Create Reference Shorelines & Transects: How-to-Create-Reference-Shorelines-&-Transects.md
       - How to Filter Bad Imagery: How-to-Filter-Out-Bad-Imagery.md
-      - Settings Guide: settings.md
       - How to Use the Config Files: config.md
   - API Reference:
       - How to use CoastSeg API: CoastSeg-API-Guide.md
