### 2024-04-19 95719a0
v1.2.3 coastsat-package>=0.1.65 update download process to filter clouds and bad pixels from imagery
