### 2024-04-19 be8049b
#219 update tide correction guide by mentioning files will be overwritten, what files are generated, and tell users to load a session
