### 2024-04-19 dfed3f6
#219 mention ROI should NOT be renamed and update getting started guide formatting
