+++ docs/getting-started.md
@@ -28,43 +28,70 @@ git clone coastseg --depth 1 https://github.com/Doodleverse/CoastSeg.git
 
 ## Extract Shorelines
 
-1. Launch Jupyter Lab
+**1.Launch Jupyter Lab**
 
 - Run this command in the coastseg directory to launch the notebook `SDS_coastsat_classifier`
-  ```bash
-  conda activate coastseg
-  jupyter lab SDS_coastsat_classifier.ipynb
-  ```
 
-2. Authenticate with Google Earth Engine
-   ![auth_cell_cropped](https://github.com/Doodleverse/CoastSeg/assets/61564689/642c8353-bfab-4458-a248-a8efce01f1ee)
-3. Draw an Bounding Box
-4. Generate ROI (Region of Interest)
-5. Load Transects
-   ![load_rois_then_transects_on_map_demo](https://github.com/Doodleverse/CoastSeg/assets/61564689/d53154b0-7a63-470f-91ec-dabdf7d4a100)
+```bash
+conda activate coastseg
+jupyter lab SDS_coastsat_classifier.ipynb
+```
 
-6. Modify the Settings
-   - Change the satellites to L8 and L9
-   - Change the dates to 12/01/2023 - 03/01/2024
-   - Change the size of the reference shoreline buffer
-   - Click `Save Settings`
-7. Name the Session
-8. Extract Shorelines
-   ![save_settings_download_extract](https://github.com/Doodleverse/CoastSeg/assets/61564689/3548a9ce-a190-4c95-b495-0ff75484fdb2)
+**2.Authenticate with Google Earth Engine**
+
+![auth_cell_cropped](https://github.com/Doodleverse/CoastSeg/assets/61564689/642c8353-bfab-4458-a248-a8efce01f1ee)
+**3.Draw an Bounding Box**
+**4.Generate ROI (Region of Interest)**
+**5.Load Transects**
+![load_rois_then_transects_on_map_demo](https://github.com/Doodleverse/CoastSeg/assets/61564689/d53154b0-7a63-470f-91ec-dabdf7d4a100)
+
+**6.Modify the Settings**
+
+- Change the satellites to L8 and L9
+
+- Change the dates to 12/01/2023 - 03/01/2024
+
+- Change the size of the reference shoreline buffer
+
+- Click `Save Settings`
+
+**7.Name the Session**
+
+**8.Download the ROIs**
+
+- Click the ROIs you want to download on the map ( they will turn blue when selected)
+
+- Make sure there are transects inside the ROI you have selected otherwise you won't be able to extract shorelines
+
+- If no transects or reference shorelines are available for the region you have uploaded follow the guide here
+
+- NEVER rename your ROIs this is because CoastSeg keeps track of the filename in the 'config.json' and you won't be able to load it into CoastSeg again if you do. The ROI's ID is the filename so that information will be lost if you rename it.
+
+**9.Extract Shorelines**
+![save_settings_download_extract](https://github.com/Doodleverse/CoastSeg/assets/61564689/3548a9ce-a190-4c95-b495-0ff75484fdb2)
 
 ## Apply Tidal Correction to Extracted Shorelines (Optional)
 
-1. Download the tide model
-   - Before tidal correction can be applied the tide model must be downloaded
-   - Follow the tutorial: [How to Download Tide Model](https://github.com/Doodleverse/CoastSeg/wiki/09.-How-to-Download-Tide-Model)
-2. Load the Session with Extracted Shorelines
-   - Re-open the jupyter notebook
-   - Under the Kernal menu Click 'restart and clear outputs of all cells'
-   - Click Load Session and load the same we made before ''
-3. Click Correct Tides
[…]
