### 2024-04-19 fee8089
add a guide on how to upload features
