+++ docs/how-to-upload-features.md
@@ -1,0 +1,17 @@
+## How to Load Features from Geojson Files
+
+**1.Click the dropdown from under 'Load Feature from File'**
+
+- Select the feature you want on the map.
+
+**2.Click the load button from under 'Load Feature from File'**
+
+- This opens a file dialog window where you can select the geojson file you want to load on the map.
+
+![load feature from file](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/3e314de6-a742-46e9-b390-36d66e7aaba9)
+
+**3.Select a geoJSON file containing the feature**
+
+- If the geojson file is too large or has too many large features the map may become very slow. It is highly recommended to only load small features onto the map.
+
+![load_geojson_files_demo](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/4fedfb32-b36b-4203-a9af-228f927ca0a8)
