### 2024-04-19 d4c0c7c
update the getting started and map guide
