### 2024-04-19 aa88368
update gif in config.md & roi.md, add more screenshots to how to create shoreline and transects
