+++ README.md
@@ -143,67 +143,130 @@ conda clean --all
 ```
 
 # Getting Started
-Check out the rest of the [wiki](https://github.com/Doodleverse/CoastSeg/wiki) for more tutorials
 
 ## Prerequisites
-1. Sign up to use Google Earth Engine Python API
 
--Request access to Google Earth Engine at https://signup.earthengine.google.com/
+**1.Sign up to use Google Earth Engine Python API**
 
--It takes about 1 day for Google to approve requests.
+- Request access to Google Earth Engine at https://signup.earthengine.google.com/
+
+- It takes about 1 day for Google to approve requests.
 
 ## Installation & SetUp
-1. Activate the coastseg conda environment
-   ```bash
-   conda activate coastseg
-   ```
+
+**1.Activate the coastseg conda environment**
+
+```bash
+conda activate coastseg
+```
+
 - If you have successfully activated coastseg you should see that your terminal's command line prompt should now start with `(coastseg)`.
 
 <img src="https://user-images.githubusercontent.com/61564689/184215725-3688aedb-e804-481d-bbb6-8c33b30c4607.png" 
      alt="coastseg activated in anaconda prompt" width="350" height="150">
 
-2. Download CoastSeg from GitHub
+**2.Download CoastSeg from GitHub**
+
 ```
 git clone coastseg --depth 1 https://github.com/Doodleverse/CoastSeg.git
 ```
+
 ## Extract Shorelines
-1. Launch Jupyter Lab
+
+**1.Launch Jupyter Lab**
 
 - Run this command in the coastseg directory to launch the notebook `SDS_coastsat_classifier`
-  ```bash
-  conda activate coastseg
-  jupyter lab SDS_coastsat_classifier.ipynb
-  ```
-2. Authenticate with Google Earth Engine
+
+```bash
+conda activate coastseg
+jupyter lab SDS_coastsat_classifier.ipynb
+```
+
+**2.Authenticate with Google Earth Engine**
+
 ![auth_cell_cropped](https://github.com/Doodleverse/CoastSeg/assets/61564689/642c8353-bfab-4458-a248-a8efce01f1ee)
-4. Draw an Bounding Box
-6. Generate ROI (Region of Interest)
-7. Load Transects
+
+**3.Draw an Bounding Box**
+
+**4.Generate ROI (Region of Interest)**
+
+**5.Load Transects**
+
 ![load_rois_then_transects_on_map_demo](https://github.com/Doodleverse/CoastSeg/assets/61564689/d53154b0-7a63-470f-91ec-dabdf7d4a100)
 
-8. Modify the Settings
-   - Change the satellites to L8 and L9
-   - Change the dates to 12/01/2023 - 03/01/2024
-   - Change the size of the reference shoreline buffer
-   - Click `Save Settings`
-9. Name the Session
-10. Extract Shorelines
+- Make sure there are transects inside the ROI you have selected otherwise you won't be able to extract shorelines
+
+- If there isn't a reference shoreline or any transects available for your site check out the guide on how to upload your own [here](https://satelliteshorelines.github.io/CoastSeg/how-to-upload-features/)
+
+**6.Modify the Settings**
+
+- Change the satellites to L8 and L9
+
+- Change the dates to 12/01/2023 - 03/01/2024
+
+- Change the size of the reference shoreline buffer
+
+- Click `Save Settings`
+
+**7.Name the Session**
+
+- Let's call this 'sample_session'
+
+- This is the name of the folder that will be saved in `CoastSeg/sessions`
+
+- It will contain a subdirectory for each ROI that shorelines will be extracted for
+
+**8.Download the ROIs**
+
+- Click the ROIs you want to download on the map ( they will turn blue when selected)
+
[…]
