+++ docs/how-to-upload-features.md
@@ -12,6 +12,8 @@
 
 **3.Select a geoJSON file containing the feature**
 
+- The feature you upload MUST be in CRS 'epsg:4326'
+
 - If the geojson file is too large or has too many large features the map may become very slow. It is highly recommended to only load small features onto the map.
 
 ![load_geojson_files_demo](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/4fedfb32-b36b-4203-a9af-228f927ca0a8)
