### 2024-04-19 ca149ef
#212 add explanation of download setting and explain resumable downloads
