+++ mkdocs.yml
@@ -57,6 +57,7 @@ nav:
       - How to Use the Map: map_guide.md
       - How to use the ROI: roi.md
       - How to Download Imagery: how-to-download.md
+      - Download Settings Guide: download_settings.md
       - How to Upload Features: how-to-upload-features.md
       - How to Use Sessions: sessions.md
       - How to Use Extract Shoreline Area: how-to-extract-shoreline-area.md
