### 2024-04-19 dd211ae
#212 how to load an extracted shoreline session from a user who didn't download your data
