+++ mkdocs.yml
@@ -60,6 +60,7 @@ nav:
       - Download Settings Guide: download_settings.md
       - How to Upload Features: how-to-upload-features.md
       - How to Use Sessions: sessions.md
+      - How to Use Sessions from Another User: how-to-use-a-session-another-user.md
       - How to Use Extract Shoreline Area: how-to-extract-shoreline-area.md
       - How to Extract Shorelines: extract_shorelines.md
       - Extract Shorelines Settings Guide: extract_shoreline_settings.md
