### 2024-04-19 ba6a261
#212 add tide model screenshot, how to create custom ROIs, and how to load a session by another user
