+++ docs/How-to-Download-Tide-Model.md
@@ -3,6 +3,7 @@
 This guide will walk you through signing up for the necessary account and then choosing between two methods to download and clip the tide model for use in CoastSeg.
 
 # Part 1: Signing Up for the FES2014 Model
+
 Step 1: Register for an AVISO+ Account
 
 ## Step 1: Register for an AVISO+ Account
@@ -14,20 +15,24 @@ Step 1: Register for an AVISO+ Account
 
 ![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/5f188ced-0f7e-4732-b065-8eb8f2d54986)
 <br>
+
 ## Step 2: Verify Your AVISO+ Account
+
 1. Check your email for a confirmation from AVISO+.
 2. Click the verification link in the email to activate your account.
 
 ## Step 3: Access Your Subscriptions
+
 1. Log into your [AVISO account](https://www.aviso.altimetry.fr/en/my-aviso-plus.html).
 2. Navigate to My products in the left side menu.
 3. Confirm that FES2014 / FES2012 (Oceanic Tides Heights) is listed under Your current subscriptions.
-<br>
-<img src="https://github.com/Doodleverse/CoastSeg/assets/61564689/bf5382f0-6bc0-4867-893d-c8f84a3d3760" alt="drawing" width="900" height="500"/>
+   <br>
+   <img src="https://github.com/Doodleverse/CoastSeg/assets/61564689/bf5382f0-6bc0-4867-893d-c8f84a3d3760" alt="drawing" width="900" height="500"/>
 
 ![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/d5758b7c-7a4a-4f48-8b0b-244a0f8d7d7a)
 
 # Part 2: Downloading and Clipping the Tide Model
+
 You have two options to download and clip the tide model: using a Jupyter Notebook (Recommended) or using a Script (Alternative). Choose one based on your comfort level and requirements.
 
 ### Space and Time Requirements
@@ -36,50 +41,74 @@ You have two options to download and clip the tide model: using a Jupyter Notebo
 - **Time Commitment:** The download and setup process for the fes2014 model takes approximately 1-2 hours. This is a one-time requirement.
 
 ## Option 1: Using the Download_tide_model.ipynb Notebook (Recommended)
+
 The easiest way to download and clip the tide model is to use this notebook. Run all the code in the notebook and you're done!
+
 1. **Prepare:** Ensure you have you are in an activated coastseg environment and know your AVISO email and password.
 2. **Open Notebook:** Launch `Download_tide_model.ipynb` in Jupyter Notebook.
+
 ```
 cd CoastSeg
 jupyter lab Download_tide_model.ipynb
 ```
+
 3.**Run Code:** Execute all cells in the notebook. The code will guide you through the download and clipping process automatically.
-4. **Check Results:** Confirm that the model files are correctly downloaded and clipped in the specified directory.
--  **Example Output from Downloading Tide Model Step in Notebook**
+
+4.**Check Results:** Confirm that the model files are correctly downloaded and clipped in the specified directory.
+
+- **Example Output from Downloading Tide Model Step in Notebook**
+
 ![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/6f4b8998-2d9f-4284-b0bb-075634efd1e8)
--  **Example Output from Clipping Tide Model Step in Notebook**
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/7251f6ca-acd4-4dda-9429-28bb83daad04)
+
+- **Example Output from Clipping Tide Model Step in Notebook**
+  ![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/7251f6ca-acd4-4dda-9429-28bb83daad04)
+
+  5.**Downloaded Tide Model Contents**
+
[…]
