### 2024-04-19 503159e
#212 add a gif for preview imagery available
