### 2024-04-19 ca29053
#212 update wiki to explain how to use apply cloud mask with examples, dates, cloud threshold
