### 2024-04-19 c84f525
close #212 by adding a specific guide to flip transects
