### 2024-04-22 209a051
update the setting started guide with more gifs and screenshots
