### 2024-04-22 911cbd8
update the resources section of index/,d to remove old wiki link
