### 2024-04-22 4891f25
remove outdated_config files guide
