### 2024-04-22 a94d3a4
update the coastseg API guide
