+++ docs/CoastSeg-API-Guide.md
@@ -17,12 +17,15 @@ This guide provides instructions on how to use the CoastSeg API to download imag
 
 ### Script 1: Download Imagery and Initial Shoreline Extraction
 
-1. **Activate the CoastSeg Conda Environment:**
-   Open your terminal and activate the CoastSeg conda environment by running:
-   ```bash
-     conda activate coastseg
-   ```
-2. **Launch the Script**
+1.**Activate the CoastSeg Conda Environment:**
+
+Open your terminal and activate the CoastSeg conda environment by running:
+
+```bash
+  conda activate coastseg
+```
+
+2.**Launch the Script**
 
 - Navigate to the directory containing your script and run:
 
@@ -37,7 +40,7 @@ The script performs the following actions:
 - Extracts shorelines from the downloaded imagery.
 - Optionally applies tide correction (uncomment the tide correction section if needed).
 
-3. **Understanding the Script Output:**
+  3.**Understanding the Script Output:**
 
 - The script will download imagery to the specified data folder.
 - Extracted shorelines will be saved in the session's directory.
@@ -48,24 +51,32 @@ The script performs the following actions:
 After completing the imagery download and initial extraction, you can proceed with the second script for further shoreline extraction.
 Note: This script should be run after the first one because it loads the 'sample_session1' created by the first script
 
-1. **Activate the CoastSeg Conda Environment:**
-   Open your terminal and activate the CoastSeg conda environment by running:
-   ```bash
-     conda activate coastseg
-   ```
-2. **Apply Tide Correction (Optional)**
+1.**Activate the CoastSeg Conda Environment:**
 
-- Ensure you have downloaded the tide model as per the prerequisites.
-- Uncomment the tide correction section in the script:
-  - make sure to enter the `beach slope` and `reference_elevation` for your site.
+- Open your terminal and activate the CoastSeg conda environment by running:
+
+```bash
+  conda activate coastseg
+```
+
+2.**Uncomment Code to Apply Tide Correction (Optional)**
+
+- The tide model MUST be downloaded as per the prerequisites. Follow a guide here on how to download it [How to Download Tide Model](https://satelliteshorelines.github.io/CoastSeg/How-to-Download-Tide-Model/)
+
+- Uncomment the tide correction section in the script (shown below):
+  - make sure to enter the `beach slope` and `reference_elevation`(relative to MSL) for your site.
 
 ```python
 # coastsegmap.compute_tidal_corrections(roi_ids, beach_slope, reference_elevation)
 ```
 
-3. **Launch the Script**
+3.**Launch the Script**
 
 - Navigate to the directory containing your script and run:
+
+- This script will use the ROI data downloaded by `1_download_imagery.py` and extract shorelines from the imagery.
+
+- If you did not run `1_download_imagery.py` to download data then this script will not work
 
 ```bash
      python 2_extract_shorelines.py
@@ -73,22 +84,6 @@ Note: This script should be run after the first one because it loads the 'sample
 
 It performs the following:
 
-- Loads the session created by the first script.
-- Applies new settings for shoreline extraction.
+- Loads the download data from a session in 'CoastSeg/data' session created by the first script.
+- Applies the new settings to control shoreline extraction from the downloaded imagery
 - Optionally, applies tide correction (uncomment and configure if needed).
-
-## Applying Tide Correction (Optional)
-
-To apply tide correction:
-
-1. **Ensure you have downloaded the tide model as per the prerequisites.**
-
-2. **Uncomment the tide correction section in each script:**
-
-```python
-# coastsegmap.compute_tidal_corrections(roi_ids, beach_slope, reference_elevation)
-```
-
[…]
