### 2024-04-22 7fa1bd2
add the work in progress for case study 1
