+++ docs/case_study_1.md
@@ -1,0 +1,190 @@
+For this case study we will select an easy site located in Santa Cruz, California.
+This site has a simple shoreline, not much cloud cover, and lots of imagery available making it a great test site
+
+## Prerequisites
+
+1. Before you begin make sure you have a GEE account if not follow the guide in [How to Authenticate with Google Earth Engine(GEE)](https://satelliteshorelines.github.io/CoastSeg/how-to-auth/)
+
+2. Have installed CoastSeg by following the [Installation Guide](https://satelliteshorelines.github.io/CoastSeg/basic-install-guide/)
+
+## SetUp
+
+**1.Activate the coastseg conda environment**
+
+```bash
+conda activate coastseg
+```
+
+- If you have successfully activated coastseg you should see that your terminal's command line prompt should now start with `(coastseg)`.
+
+<img src="https://user-images.githubusercontent.com/61564689/184215725-3688aedb-e804-481d-bbb6-8c33b30c4607.png" 
+     alt="coastseg activated in anaconda prompt" width="350" height="150">
+
+**2.Download CoastSeg from GitHub**
+
+- Only do this step if you haven't already cloned CoastSeg to your computer
+
+```
+git clone coastseg --depth 1 https://github.com/Doodleverse/CoastSeg.git
+```
+
+**3.Change to the CoastSeg Directory**
+
+- Open a command prompt like anaconda prompt and go the location you installed CoastSeg
+
+```bash
+cd coastseg
+```
+
+## Launch the Notebook
+
+### 1.Launch Jupyter Lab
+
+- Run this command in the coastseg directory to launch the notebook `SDS_coastsat_classifier`
+
+```bash
+conda activate coastseg
+jupyter lab SDS_coastsat_classifier.ipynb
+```
+
+### 2.Authenticate with Google Earth Engine
+
+- Run the cell located under 'Authenticate with Google Earth Engine (GEE)'
+
+![auth_cell_cropped](https://github.com/Doodleverse/CoastSeg/assets/61564689/642c8353-bfab-4458-a248-a8efce01f1ee)
+
+### 3.Draw an Bounding Box
+
+- Draw a bounding box along the coast in this box is where ROIs will be created
+
+- ROIs can only be generated along a shoreline
+
+- If no shoreline is found then an error message will appear telling you no ROIs can be created. If this happens create your own reference shoreline following the guide here [How to Create Reference Shoreline](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/)
+
+![case_study_1_bbox](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/a3e482fa-8498-4ea9-82b6-c8f1edab45ea)
+
+### 4.Load Available Shorelines
+
+- Click load shoreline to load the one of CoastSeg's default reference shorelines within the bounding box region
+
+![case_study_1_shorelines](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/7723b864-3e47-4a99-ae66-db1a7be9cfa6)
+
+### 5.Load Transects
+
+- Make sure there are transects inside the ROI you have selected otherwise you won't be able to extract shorelines
+
+- If there isn't a reference shoreline or any transects available for your site check out the guide on how to upload your own [here](https://satelliteshorelines.github.io/CoastSeg/how-to-upload-features/)
+
+![case_study_1_transects](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/76a658cc-d5f9-4fcb-a4a4-9b9a474f4d4c)
+
+### 6.Modify the Settings
+
[…]
