### 2024-04-22 19be441
finish the case study 1
