+++ docs/case_study_1.md
@@ -109,7 +109,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![case study 1 session name circle](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/bd85acd1-0cdc-4eba-b3b6-75b0402bbb76)
 
-### 7.Preview the available Imagery
+### 8.Preview the available Imagery
 
 - Preview the amount of available imagery for the selected ROI between the dates
 
@@ -117,7 +117,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![case study 1 preview imagery](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/db42fee9-682b-4e15-8470-b97a166e42a8)
 
-### 8.Download the ROI
+### 9.Download the ROI
 
 - Click the ROI you want to download on the map ( they will turn blue when selected)
 
@@ -127,7 +127,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![case_study_1_download_roi](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/1a30f9c7-fc4d-4e34-a57b-055624ff8464)
 
-### 9.Sort the Downloaded Imagery
+### 10.Sort the Downloaded Imagery
 
 - Open Coastseg/data and open the folder containing the ROI ID, in my case thats 'tto3', so I opened 'CoastSeg\data\ID_tto3_datetime04-22-24\_\_03_47_52'
 
@@ -139,7 +139,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![case_study_1_bad_sort](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/aaba59a9-9523-4419-a80c-9fa860491984)
 
-**10.Extract Shorelines**
+**11.Extract Shorelines**
 
 - Extracting shorelines works by finding the land water interface in the image and drawing a line along it
 
@@ -147,23 +147,57 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![shoreline_transect_intersection](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/e87b8d34-d9a4-4b1e-b3de-8e0be1c16ecd)
 
-### 11. Examine Detection Images for Extracted Shorelines
+### 12. Examine Detection Images for Extracted Shorelines
 
 - The detection images for the extracted shorelines is at 'CoastSeg\sessions\case_study_1\ID_cwm3_datetime04-22-24\_\_02_57_16\jpg_files\detection'
 
 - In these images you can see how well the shoreline were extracted depending on cloud cover, the size of the reference shoreline buffer and the rest of the extract shoreline settings
 
+- There are a few images with some bad shorelines. Lets remove those in step 13
+
 ![case_study_1_detection_images](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/5bd10163-77bc-4fb1-8669-394ddb8a5bf5)
+
+### 13. Remove Outlier/Bad Extracted Shorelines
+
+- Use the Load Extracted Shoreline feature to view all the extracted shorelines on the map
+
+- Find any bad shorelines and click the trash icon to put that shoreline in the trash
+
+- Once you've put all the bad shorelines in the trash click the empty trash button and this will delete all those shorelines from all the files in the session directory.
+
+![case_study_1_remove_outlier_shorelines](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/9e0c1c60-de27-4f5c-a0ab-66eeb20e64ae)
+
+### 14. Open the Extracted Shoreline Session Outputs in QGIS
+
+**Config_gdf.geojson**
+
[…]
