### 2024-04-23 86dcaba
remove gif from save settings
