### 2024-04-23 48e4100
fix typos in case study 1 and getting started
