+++ docs/case_study_2.md
@@ -1,0 +1,252 @@
+# Hey this Guide is Under Construction 🚧 👷‍♀️ 🚧
+
+## Check back later when we're finished!! 🏗️
+
+# 🚧 🚧 🚧 🚧 🚧 🚧 🚧 🚧 🚧
+
+For this case study we will select an difficult site located outside of Unalakleet Alaska.
+This site has a simple shoreline,lots of cloud cover, and not imagery available due to ice and clouds covering the region most of the year making it a difficult test site.
+
+![unakleet_site](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/157636ba-c1bc-4628-8d9a-04bef1ca1246)
+
+## Prerequisites
+
+1. Before you begin make sure you have a GEE account if not follow the guide in [How to Authenticate with Google Earth Engine(GEE)](https://satelliteshorelines.github.io/CoastSeg/how-to-auth/)
+
+2. Have installed CoastSeg by following the [Installation Guide](https://satelliteshorelines.github.io/CoastSeg/basic-install-guide/)
+
+## SetUp
+
+**1.Activate the coastseg conda environment**
+
+```bash
+conda activate coastseg
+```
+
+- If you have successfully activated coastseg you should see that your terminal's command line prompt should now start with `(coastseg)`.
+
+<img src="https://user-images.githubusercontent.com/61564689/184215725-3688aedb-e804-481d-bbb6-8c33b30c4607.png" 
+     alt="coastseg activated in anaconda prompt" width="350" height="150">
+
+**2.Download CoastSeg from GitHub**
+
+- Only do this step if you haven't already cloned CoastSeg to your computer
+
+```
+git clone coastseg --depth 1 https://github.com/Doodleverse/CoastSeg.git
+```
+
+**3.Change to the CoastSeg Directory**
+
+- Open a command prompt like anaconda prompt and go the location you installed CoastSeg
+
+```bash
+cd coastseg
+```
+
+## Launch the Notebook
+
+### 1.Launch Jupyter Lab
+
+- Run this command in the coastseg directory to launch the notebook `SDS_coastsat_classifier`
+
+```bash
+conda activate coastseg
+jupyter lab SDS_coastsat_classifier.ipynb
+```
+
+### 2.Authenticate with Google Earth Engine
+
+- Run the cell located under 'Authenticate with Google Earth Engine (GEE)'
+
+![auth_cell_cropped](https://github.com/Doodleverse/CoastSeg/assets/61564689/642c8353-bfab-4458-a248-a8efce01f1ee)
+
+### 3.Draw an Bounding Box
+
+- Draw a bounding box along the coast in this box is where ROIs will be created
+
+- ROIs can only be generated along a shoreline
+
+- If no shoreline is found then an error message will appear telling you no ROIs can be created. If this happens create your own reference shoreline following the guide here [How to Create Reference Shoreline](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/)
+
+![case_study_2_bbox_for_ak](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/aa4ca94f-3049-4801-99bc-98fe34b7c671)
+
+### 4.Create and Load a Reference Shoreline
+
+- This site does not have a shoreline available so we will have to make our own. Follow the tutorial at [How to Create Reference Shoreline](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/) for a guide on how to create your own reference shoreline
+
+- Here is a guide on how to upload your own [here](https://satelliteshorelines.github.io/CoastSeg/how-to-upload-features/) transects
+
+- I used geojson.io to create my reference shoreline
+
[…]
