+++ mkdocs.yml
@@ -47,7 +47,8 @@ nav:
       - Secure Installation Guide: secure-install-guide.md
   - Getting Started: getting-started.md
   - Case Studies:
-      - Case Study 1 Easy Location in Santa Cruz: case_study_1.md
+      - Case Study 1 An Easy Location in Santa Cruz: case_study_1.md
+      - Case Study 2 A Difficult Location in Alaska: case_study_2.md
   - Contributing:
       - Contribution Guide: Contribution-Guide.md
       - How to Run Test Scripts: contributing-run-tests.md
