### 2024-04-23 cd34a2e
add gifs to case study 2
