### 2024-04-23 ef20246
update case study 2 with extract shorelines examples
