+++ docs/case_study_2.md
@@ -87,7 +87,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - I used geojson.io to create my transects
 
-![case_study_2_gen_and_download_roi](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/e8755f13-e2ef-45e4-a22f-42a2d46b150a)
+![case_study_2_shoreline_transects_upload](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/cd780e5a-39f7-42ba-90ac-d632390c18dc)
 
 ### 6.Modify the Settings
 
@@ -111,6 +111,14 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![case_study_1_extracted_shoreline_rel_sl_buffer](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/c16dc3a2-a211-4e9f-85c0-fc144e9d4f83)
 
+- Turn off the Apply Cloud Mask
+
+  -- The reason for this is that Alaska is very cloudy and often light clouds will be masked out close to the shoreline causing gaps in the shoreline due to the cloud distance threshold. If we don't mask out the clouds then we get a few more usable images
+
+  -- In this example you can see the clouds where masked out in this image but these clouds were onshore so they would not have interfered with the cloud masking. However, the cloud masking was turned on and the 'cloud distance' parameter which removes any shorelines within 'cloud distance' to a cloud causes there to be gaps in the shoreline. You could lower cloud distance to something like 10-20m to fix this or turn off cloud masking
+
+![case_study_2_why_cloud_mask_off](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/3d42653c-54fb-4568-a59c-d5ed75e3d844)
+
 - Click `Save Settings`
 
 ![case_study_1_save_settings](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/aef76b30-3677-417b-8d19-5313649cb877)
@@ -131,7 +139,9 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 ![case study 1 preview imagery](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/db42fee9-682b-4e15-8470-b97a166e42a8)
 
-### 9.Download the ROI
+### 9.Add a Shoreline Extraction Area
+
+### 10.Download the ROI
 
 - Click the ROI you want to download on the map ( they will turn blue when selected)
 
@@ -139,21 +149,45 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - When the download finishes CoastSeg will print the location where the downloads were saved in this case its 'CoastSeg\data\ID_tto3_datetime04-22-24\_\_03_47_52'
 
-![case_study_1_download_roi](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/1a30f9c7-fc4d-4e34-a57b-055624ff8464)
+![case_study_2_gen_and_download_roi](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/e8755f13-e2ef-45e4-a22f-42a2d46b150a)
 
-### 10.Sort the Downloaded Imagery
+### 11.Sort the Downloaded Imagery
 
 - Open Coastseg/data and open the folder containing the ROI ID, in my case thats 'tto3', so I opened 'CoastSeg\data\ID_tto3_datetime04-22-24\_\_03_47_52'
 
 - You can see the ROI ID in the hover menu located to the top right of the map
 
-![roi_id_hover](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/f7b3e931-b511-4cd7-acbf-5badf0f7d382)
[…]
