### 2024-04-23 8bde542
update case study 2 to have shoreline extraction area
