### 2024-04-23 1738617
Organize shoreline extraction area
