### 2024-04-23 3d52a60
remove construction banner and add final gifs to case study 2
