+++ docs/case_study_2.md
@@ -1,9 +1,3 @@
-# Hey this Guide is Under Construction 🚧 👷‍♀️ 🚧
-
-## Check back later when we're finished!! 🏗️
-
-# 🚧 🚧 🚧 🚧 🚧 🚧 🚧 🚧 🚧
-
 For this case study we will select an difficult site located outside of Unalakleet Alaska.
 This site has a simple shoreline,lots of cloud cover, and not imagery available due to ice and clouds covering the region most of the year making it a difficult test site.
 
@@ -75,7 +69,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - This site does not have a shoreline available so we will have to make our own. Follow the tutorial at [How to Create Reference Shoreline](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/) for a guide on how to create your own reference shoreline
 
-- Here is a guide on how to upload your own [here](https://satelliteshorelines.github.io/CoastSeg/how-to-upload-features/) transects
+- Here is a guide on how to upload your own [here](https://satelliteshorelines.github.io/CoastSeg/how-to-upload-features/)
 
 - I used geojson.io to create my reference shoreline
 
@@ -85,7 +79,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - Here is a guide on how to upload your own [here](https://satelliteshorelines.github.io/CoastSeg/how-to-upload-features/) transects
 
-- I used geojson.io to create my transects
+- I used geojson.io to create my transects.This technique is in the tutorial [How to Create Reference Shoreline](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/)
 
 ![case_study_2_shoreline_transects_upload](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/cd780e5a-39f7-42ba-90ac-d632390c18dc)
 
@@ -123,7 +117,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - Click `Save Settings`
 
-![case_study_1_save_settings](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/aef76b30-3677-417b-8d19-5313649cb877)
+![case_study_2_save_settings](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/b502e062-e518-4d26-8fa7-fe99de489693)
 
 ### 7.Name the Session
 
@@ -131,7 +125,7 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - This is the name of the folder that will be saved in `CoastSeg/sessions`
 
-![case study 1 session name circle](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/bd85acd1-0cdc-4eba-b3b6-75b0402bbb76)
+![case_study_2_session](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/be91a824-66ac-4232-8a68-99160eca6ba6)
 
 ### 8.Preview the available Imagery
 
@@ -145,11 +139,13 @@ jupyter lab SDS_coastsat_classifier.ipynb
 
 - Our study site has a small pond located along the shoreline that will get picked up in the reference shoreline buffer and the pond's water land interface will get misidentified as a shoreline
 
-- To get around this issue we have 2 options:
+To get around this issue we have 2 options:
 
-  1. Shrink the reference shoreline buffer
-     -- This would probably be the best solution for this specific, but if you have a dynamic coastline this may not be the best solution for you
[…]
