### 2024-04-23 ec9147f
add search plugin
