### 2024-04-25 9dddb84
Update README.md
Credit to writing this new ReadMe goes to @dbuscombe-usgs 
@2320sharon validated all the links work and merged this new readme in
