+++ README.md
@@ -19,402 +19,128 @@
 
 <img src="https://user-images.githubusercontent.com/61564689/212394936-263ec9fc-fb82-45b8-bc79-bc57dafdae73.gif" width="350" height="350">
 
-### Table of Contents
 
-- [Installation Instructions](#installation-instructions)
-- [Getting Started](#getting-started)
-- [CoastSeg Update Guide](#coastseg-update-guide)
+# Website 
 
-### Useful Links
+Please read our [website](https://satelliteshorelines.github.io/CoastSeg/), which includes installation and usage guides, contribution guide, API description, case studies, and more. 
 
-- [Getting Started Guide](https://satelliteshorelines.github.io/CoastSeg/getting-started/)
-- [Wiki](https://satelliteshorelines.github.io/CoastSeg/)
-- [Discussion](https://github.com/SatelliteShorelines/CoastSeg/discussions)
-- [Contribution Guide](https://satelliteshorelines.github.io/CoastSeg/Contribution-Guide/)
+# What is CoastSeg?
+CoastSeg is an interactive browser-based program that aims to broaden the adoption of satellite-derived shoreline (SDS) detection and coastal landcover mapping workflows among coastal scientists and coastal resource management practitioners. SDS is a sub-field of coastal sciences that aims to detect and post-process a time-series of shoreline locations from publicly available satellite imagery.
 
+CoastSeg is a Python package installed via pip into a conda environment that serves as an toolkit for building custom SDS workflows. CoastSeg also provides full SDS workflow implementations via Jupyter notebooks and Python scripts that call functions and classes in the core CoastSeg toolkit for specific workflows. CoastSeg provides two fully functioning SDS workflows and its design allows for collaborators in the SDS software community to contribute additional workflows. All the codes, notebooks, scripts, and documentation are hosted on the [CoastSeg GitHub repository](https://github.com/SatelliteShorelines/CoastSeg).
 
-## What is CoastSeg?
+So-called `instantaneous' SDS workflows, such as the implementation of [CoastSat](https://github.com/kvos/CoastSat), where shorelines are extracted from each individual satellite image rather than temporal composites, follow a basic recipe, namely:
 
-Coastseg stands for Coastal Segmentation, it is an interactive jupyter notebook for downloading satellite imagery with [CoastSat](https://github.com/kvos/CoastSat) and applying Doodleverse/[Zoo](https://github.com/Doodleverse/segmentation_zoo) image segmentation models to satellite imagery. CoastSeg provides an interactive interface for drawing Regions of Interest (ROIs) on a map, downloading satellite imagery, loading geojson files, extracting shorelines from satellite imagery, and much more. In a nutshell, it is ...
+1) waterline estimation, where the 2D (x,y) location of the land-sea interface is determined
+and 
+2) water-level correction, where the waterline location is mapped onto a shore-perpendicular transect, converted to a linear distance along that transect, then corrected for water level, and referenced to a particular elevation contour on the beach. 
 
-- ... an easy way to download time-series of cloud-masked and post-processed satellite imagery anywhere in the world
-- ... a mapping extension for [CoastSat](https://github.com/kvos/CoastSat) using [Segmentation Zoo](https://github.com/Doodleverse/segmentation_zoo) models.
-- ... an interactive interface to download satellite imagery using CoastSat from Google Earth Engine
-- ... an interactive interface for extracting shorelines from satellite imagery
-- ... an interactive interface to apply segmentation models to satellite imagery
+# Project Aims
[…]
