### 2024-04-25 2cfaeeb
Add the update guide, update index.md to match readme
