+++ docs/index.md
@@ -1,5 +1,3 @@
-# CoastSeg
-
 [![image](https://img.shields.io/pypi/v/coastseg.svg?color=%23ec3dc8)](https://pypi.python.org/pypi/coastseg)
 </br>
 [![image](https://github.com/SatelliteShorelines/CoastSeg/actions/workflows/pip_install_e.yml/badge.svg)](https://github.com/SatelliteShorelines/CoastSeg/actions)
@@ -16,252 +14,107 @@
 
 <img src="https://user-images.githubusercontent.com/61564689/212394936-263ec9fc-fb82-45b8-bc79-bc57dafdae73.gif" width="350" height="350">
 
-### Resources
+## Installation Instructions
 
-- [Getting Started Guide](https://satelliteshorelines.github.io/CoastSeg/getting-started/)
-- [Wiki](https://satelliteshorelines.github.io/CoastSeg/)
-- [Discussion](https://github.com/SatelliteShorelines/CoastSeg/discussions)
-- [Contribution Guide](https://satelliteshorelines.github.io/CoastSeg/Contribution-Guide/)
+CoastSeg is a Jupyter and Python based program that runs in a conda environment. Please see [the installation guide](https://satelliteshorelines.github.io/CoastSeg/basic-install-guide/)
+
+## Getting Started
+
+Please see [Getting Started guide](https://satelliteshorelines.github.io/CoastSeg/getting-started/), which includes instructions for
+
+- Pre-requisites (accounts, data access)
+- Installation and setup
+- Extracting shorelines
+- Applying tidal corrections to extracted shorelines
 
 ## What is CoastSeg?
 
-Coastseg stands for Coastal Segmentation, it is an interactive jupyter notebook for downloading satellite imagery with [CoastSat](https://github.com/kvos/CoastSat) and applying Doodleverse/[Zoo](https://github.com/Doodleverse/segmentation_zoo) image segmentation models to satellite imagery. CoastSeg provides an interactive interface for drawing Regions of Interest (ROIs) on a map, downloading satellite imagery, loading geojson files, extracting shorelines from satellite imagery, and much more. In a nutshell, it is ...
+CoastSeg is an interactive browser-based program that aims to broaden the adoption of satellite-derived shoreline (SDS) detection and coastal landcover mapping workflows among coastal scientists and coastal resource management practitioners. SDS is a sub-field of coastal sciences that aims to detect and post-process a time-series of shoreline locations from publicly available satellite imagery.
 
-- ... an easy way to download time-series of cloud-masked and post-processed satellite imagery anywhere in the world
-- ... a mapping extension for [CoastSat](https://github.com/kvos/CoastSat) using [Segmentation Zoo](https://github.com/Doodleverse/segmentation_zoo) models.
-- ... an interactive interface to download satellite imagery using CoastSat from Google Earth Engine
-- ... an interactive interface for extracting shorelines from satellite imagery
-- ... an interactive interface to apply segmentation models to satellite imagery
+CoastSeg is a Python package installed via pip into a conda environment that serves as an toolkit for building custom SDS workflows. CoastSeg also provides full SDS workflow implementations via Jupyter notebooks and Python scripts that call functions and classes in the core CoastSeg toolkit for specific workflows. CoastSeg provides two fully functioning SDS workflows and its design allows for collaborators in the SDS software community to contribute additional workflows. All the codes, notebooks, scripts, and documentation are hosted on the [CoastSeg GitHub repository](https://github.com/SatelliteShorelines/CoastSeg).
 
-In more detail, CoastSeg represents the following advances in satellite-derived shoreline workflows:
[…]
