### 2024-04-25 #6683 @cheginit
@FlorisCalkoen & @fmemuir, thanks again for accepting to review this submission. This is the review thread for the submission. All of our communications will happen here from now on.

First, please follow the instructions in the first comment of this thread to create your review checklists.

These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied by editing your checklist comment that our EditorBot will create for you.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#6683` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them, instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2–4 weeks. Please reach out to me if you need more time. You can also use `@EditorialBot` to set automatic reminders if you know you'll be away for a known period of time.

Please don't hesitate to ping me (`@cheginit`) with any questions or concerns.
