### 2024-04-26 e269b7b
add a new script for predict tides for coastsat_package users
