+++ scripts/predict_tides.py
@@ -1,0 +1,929 @@
+# Standard library imports
+import os
+import pathlib
+from pathlib import Path
+import time
+import argparse
+from typing import Dict, List, Tuple, Union
+
+# Third-party imports
+import geopandas as gpd
+import numpy as np
+import pandas as pd
+import pyproj
+import pyTMD.io
+import pyTMD.io.model
+import pyTMD.predict
+import pyTMD.spatial
+import pyTMD.time
+import pyTMD.utilities
+from shapely.geometry import Point
+import shapely
+from shapely.geometry import LineString
+
+# Instructions for running the script
+# -----------------------------------
+# Example 1: 
+# Run the script using a single point in the GeoJSON file parranporth.geojson,
+# the time series data in the CSV file transect_time_series.csv,
+# and the FES 2014 model region is located in the directory tide_model.
+# python predict_tides.py -C "C:\Users\sf230\Downloads\parranporth.geojson" -T "C:\development\doodleverse\coastsat_package\coastsat_package\data\NARRA\transect_time_series.csv"  -M "C:\development\doodleverse\coastseg\CoastSeg\tide_model"
+
+# the output will be saved in the current working directory as tidal_predictions.csv
+# the output file will have the columns dates,x,y,tide, where x and y are the coordinates of the point in the GeoJSON file
+# -----------------------------------
+
+def convert_date_gdf(gdf):
+    """
+    Convert date columns in a GeoDataFrame to datetime format.
+
+    Args:
+        gdf (GeoDataFrame): The input GeoDataFrame.
+
+    Returns:
+        GeoDataFrame: The converted GeoDataFrame with date columns in datetime format.
+    """
+    gdf = gdf.copy()
+    if 'dates' in gdf.columns:
+        gdf['dates'] = pd.to_datetime(gdf['dates']).dt.tz_convert(None)
+    if 'date' in gdf.columns:
+        gdf['date'] = pd.to_datetime(gdf['date']).dt.tz_convert(None)
+    gdf = stringify_datetime_columns(gdf)
+    return gdf
+
+def create_complete_line_string(points):
+    """
+    Create a complete LineString from a list of points.
+
+    Args:
+        points (numpy.ndarray): An array of points representing the coordinates.
+
+    Returns:
+        LineString: A LineString object representing the complete line.
+
+    Raises:
+        None.
+
+    """
+    # Ensure all points are unique to avoid redundant looping
+    unique_points = np.unique(points, axis=0)
+    
+    # Start with the first point in the list
+    if len(unique_points) == 0:
+        return None  # Return None if there are no points
+    
+    starting_point = unique_points[0]
+    current_point = starting_point
+    sorted_points = [starting_point]
+    visited_points = {tuple(starting_point)}
+
+    # Repeat until all points are visited
+    while len(visited_points) < len(unique_points):
+        nearest_distance = np.inf
+        nearest_point = None
+        for point in unique_points:
+            if tuple(point) in visited_points:
+                continue  # Skip already visited points
+            # Calculate the distance to the current point
+            distance = np.linalg.norm(point - current_point)
+            if distance < nearest_distance:
+                nearest_distance = distance
+                nearest_point = point
+
+        # Break if no unvisited nearest point was found (should not happen if all points are unique)
+        if nearest_point is None:
+            break
+
+        sorted_points.append(nearest_point)
+        visited_points.add(tuple(nearest_point))
+        current_point = nearest_point
+
+    # Convert the sorted list of points to a LineString
+    return LineString(sorted_points)
+
[…]
