### 2024-04-26 e44e32b
add install instructions for pip instructions
