### 2024-04-26 abe662e
add a guide for the predict tides script
