+++ mkdocs.yml
@@ -82,6 +82,7 @@ nav:
       - How to Use Scripts: How-to-Use-Scripts.md
       - How to Use Apply Tidal Correction Script: Apply-Tidal_Correction_Script.md
       - How to Flip Transects: How-to-Use-Flip-Transects.md
+      - How to Predict Tides: how_to_predicts_tides_script.md
       # - CoastSeg: scripts-coastseg.md
       # - CoastSat: scripts-coastsat.md
       # - Zoo: model.md
