### 2024-04-26 0fa074e
fix bug where the tide model won't load if it wasn't in the default location
