### 2024-05-02 b097614
v1.2.4 update the transects to face the correct direction and remove bad transects #173 #149 and add script to update transects
