+++ docs/contribute-transects.md
@@ -1,0 +1,30 @@
+# Contributing Transects
+
+## Prerequisites
+
+1.Make your transects into a geojson file
+
+2.Make sure your transects are facing the right direction
+
+- Each transect should have its origin on land and its end point on the ocean
+
+  3.Make sure each transect has a unique ID in its 'id' field
+
+- Each transect should have its own unique ID in the 'id' field
+- If you don't assign an id to each transect then run the cell in `add_new_transects_notebook.ipynb` to create a unique ID for each transect
+
+## Adding the New Transects
+
+1.Open `CoastSeg/notebooks/add_new_transects_notebook.ipynb`
+
+2.Follow the instructions in the notebook to add the geojson file
+
+3.Replace the csv in `CoastSeg/src/coastseg/bounding_boxes` with the new `transects_bounding_boxes.csv`
+
+- `transects_bounding_boxes.csv` is generated by `CoastSeg/notebooks/add_new_transects_notebook.ipynb`
+
+  4.Open a pull request
+
+- State which region of the world the transect are for
+
+- Provide a screenshot of the transects you are adding
