### 2024-05-06 SatelliteShorelines/CoastSeg#243 @FlorisCalkoen
Version attribute is missing 
ping #openjournals/joss-reviews#6683

Typically a version is specified in the package at `src/<package>/.__init__.py`. See [this](https://github.com/geopandas/geopandas/blob/main/geopandas/__init__.py) more advanced GeoPandas example, that probably uses versioneer. But simpler approaches exist as well. See for example: https://py-pkgs.org/07-releasing-versioning

Currently this doesn't work:
```python
>>> import coastseg
>>> coastseg.__version__
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: module 'coastseg' has no attribute '__version__'
>>>
```
