### 2024-05-06 SatelliteShorelines/CoastSeg#244 @FlorisCalkoen
Multiple and different installation instructions
ping #openjournals/joss-reviews#6683

The installation instructions [here](https://github.com/SatelliteShorelines/CoastSeg/wiki/01.-How-to-Install-CoastSeg#mac-users) are different than [these](https://satelliteshorelines.github.io/CoastSeg/basic-install-guide/). I would prefer a conda install, but I think it's important to maintain your instructions at one place and make sure that they are consistent.
