### 2024-05-06 SatelliteShorelines/CoastSeg#245 @FlorisCalkoen
Maybe indicate how to install transformers dependency? 
ping #openjournals/joss-reviews#6683

I'm particularly interested in CoastSeg' coastal DL model, so I'm currently installing everything on my PC. So far I have managed to install Tensorflow, but I see in your [docs](https://satelliteshorelines.github.io/CoastSeg/basic-install-guide/) that I also need transformers. Since the installation requirements for all ML libs you use so far where more specific than the regular python libraries, I checked out the transformers docs. However, there I see that there are a whole lot of different ways of installing transformers, also depending on used TF version. Can you maybe provide some guidelines on which one to pick?

I'm a MacOS user, with this CoastSeg and Tensorflow version:

```bash
...
coastseg                  1.2.4              pyhd8ed1ab_0    conda-forge
tensorflow                2.16.1                   pypi_0    pypi
tensorflow-io-gcs-filesystem 0.37.0                   pypi_0    pypi
```

https://huggingface.co/docs/transformers/en/installation
