### 2024-05-06 SatelliteShorelines/CoastSeg#246 @FlorisCalkoen
Examples folder contains data 
ping #openjournals/joss-reviews#6683

I expected to find notebooks in the [examples](https://github.com/SatelliteShorelines/CoastSeg/tree/main/examples) directory, but there is data. I would suggest rename this to data.

Moreover, I would try to avoid contaminating the repository with any data. I think it's better to keep storage and code seperated. Adding data to GH repo's can make them tedious and frustrating to work with. I might consider opening a separate issue on this.
