### 2024-05-06 SatelliteShorelines/CoastSeg#244 @2320sharon
Hi Floris,
Thank you for raising this issue. I need to decommission the old coastseg wiki since we have moved all the install instructions to the website. I'll take care of that today.
