### 2024-05-06 SatelliteShorelines/CoastSeg#247 @FlorisCalkoen
CRS mismatch error when trying to map shorelines with manually-uploaded transects
ping #openjournals/joss-reviews#6683

I have manually uploaded transects for my area of interest in 4326 - please see output below, but I get a CRS mismatch error, when computing the shoreline distances. Can you reproduce this? I'm looking at Terschelling in The Netherlands (see screenshot below).

```ipython
In [5]: transects = gpd.read_file("https://coclico.blob.core.windows.net/public/gcts-terschelling.geojson")

In [6]: transects.crs
Out[6]:
<Geographic 2D CRS: EPSG:4326>
Name: WGS 84
Axis Info [ellipsoidal]:
- Lat[north]: Geodetic latitude (degree)
- Lon[east]: Geodetic longitude (degree)
Area of Use:
- name: World.
- bounds: (-180.0, -90.0, 180.0, 90.0)
Datum: World Geodetic System 1984 ensemble
- Ellipsoid: WGS 84
- Prime Meridian: Greenwich
```

```
CRS mismatch between CRS of the passed geometries and 'crs'. Use 'GeoDataFrame.set_crs(crs, allow_override=True)' to overwrite CRS or 'GeoDataFrame.to_crs(crs)' to reproject geometries.

Additional Information
Traceback (most recent call last): File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/map_UI.py", line 905, in extract_shorelines_button_clicked self.coastseg_map.extract_all_shorelines() File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/coastseg_map.py", line 1858, in extract_all_shorelines self.compute_transects(self.transects.gdf, self.get_settings(), selected_roi_ids) File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/coastseg_map.py", line 1992, in compute_transects self.save_transect_timeseries(session_path,self.rois.get_extracted_shoreline(roi_id),roi_id) File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/coastseg_map.py", line 2098, in save_transect_timeseries common.save_transects( File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/common.py", line 1912, in save_transects merged_timeseries_df,timeseries_df = add_lat_lon_to_timeseries(merged_timeseries_df, transects_gdf.to_crs('epsg:4326'),cross_distance_df, File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/common.py", line 1787, in add_lat_lon_to_timeseries new_gdf_shorelines_wgs84=convert_points_to_linestrings(cross_shore_pts, group_col='date', output_crs='epsg:4326') File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/common.py", line 1998, in convert_points_to_linestrings linestrings_gdf = gpd.GeoDataFrame(linestrings, columns=['geometry'], crs=output_crs) File "/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/geopandas/geodataframe.py", line 152, in __init__ raise ValueError(crs_mismatch_error) ValueError: CRS mismatch between CRS of the passed geometries and 'crs'. Use 'GeoDataFrame.set_crs(crs, allow_override=True)' to overwrite CRS or 'GeoDataFrame.to_crs(crs)' to reproject geometries.
```

![image](https://github.com/SatelliteShorelines/CoastSeg/assets/44444001/2873d24e-73ba-4179-830f-43880dba592f)
