### 2024-05-06 SatelliteShorelines/CoastSeg#248 @FlorisCalkoen
Why do you need transects to extract the shorelines? 
ping #openjournals/joss-reviews#6683

While trying to map some shorelines for Terschelling in the Netherlands I had to upload transects. Why not seperate shoreline mapping from computing series of shoreline positions over a set of transects? IMHO you only need transects if time series of shoreline change, but strictly speaking for the SDS you don't need transects.
