### 2024-05-06 8dd4e24
Fix: Version attribute is missing  #243
