+++ src/coastseg/__init__.py
@@ -1,0 +1,2 @@
+from importlib.metadata import version
+__version__ = version("coastseg")
