### 2024-05-06 SatelliteShorelines/CoastSeg#243 @2320sharon
Hi Floris,

Thanks for pointing this out. I added  the code below to `src>coastseg>__init__.py` and now the code will print the coastseg version. I'll update the next patch release to include this code so any release past 1.2.3 will have this available

![image](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/ddf43681-581b-4bb7-95e1-9c2e735451cd)

```
from importlib.metadata import version
__version__ = version("coastseg")

```
