### 2024-05-06 SatelliteShorelines/CoastSeg#248 @2320sharon
Hi Floris,

Great question, in the early versions of CoastSeg we did separate the steps of extracting the 2D shorelines from the imagery and computing the shoreline change along the transects. However, we decided to move away from this design for two reason. The first being that our targeted user base needs transects to explain shoreline change and while 2D shorelines are useful they are generally not accurate for accessing change at a single point. The second reason is that most of our beta users wanted to find the shoreline change along transects so having the button separated was confusing for them and made the process take longer.
