### 2024-05-06 SatelliteShorelines/CoastSeg#246 @2320sharon
Hi Floris,

I understand the concern about having data in the repository. The only data we have included currently is the default transects data and the example geojson files you pointed out that are used in the `1_download_imagery.py` script, which we have a short guide on at [CoastSeg website](https://satelliteshorelines.github.io/CoastSeg/CoastSeg-API-Guide/#script-1-download-imagery-and-initial-shoreline-extraction). To make things a bit easier for a new user we could add a txt file to the `examples` folder to explain this data is used for the demo scripts, but plenty of other tools include a small amount of demo data for their examples as well.

As for the default transect data we provide having it provided with the CoastSeg project makes it so that users don't need to wait to download transect data and that we don't have to worry about breaking any links if we decide to modify the default transects available to CoastSeg.
