### 2024-05-06 SatelliteShorelines/CoastSeg#242 @2320sharon
Hi @FlorisCalkoen

Thank you for raising this issue. We constantly wrestle with the Tensorflow compatibility issues on Windows. An overwhelming majority of our users use Windows, therefore we have decided to only (for now) support Windows (and Linux, where Tensorflow issues are minimal). We plan on providing Mac support when we can, and in the meantime we are reliant on the community of experimental Mac users. None of the developers have access to a Mac computer.

For now, I'll raise an issue for future support for Mac if the wider community wants to help us build it out and support it.
