+++ .github/workflows/add_to_project.yml
@@ -13,7 +13,7 @@ jobs:
     name: Add issue to project
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/add-to-project@v0.5.0
+      - uses: actions/add-to-project@v1.0.1
         with:
-          project-url: https://github.com/orgs/Doodleverse/projects/5
+          project-url: https://github.com/orgs/SatelliteShorelines/projects/1
           github-token: ${{ secrets.ADD_TO_PROJECT_PAT }}
