### 2024-05-06 ff627b6
Add a mac installation guide to address issue #242
