### 2024-05-06 fe2e9b5
update the basic install guide to link to the new mac install guide #244 #242
