### 2024-05-06 6f2739f
update the secure install guide to mention new mac guide
