### 2024-05-06 ff11ce3
update the secure install guide with new links
