### 2024-05-06 f7628cb
update gif how to auth
