### 2024-05-06 20d6add
update how to load shorelines on the map
