+++ docs/edit_reference_shorelines.md
@@ -14,10 +14,9 @@ You need to have only one set of reference shorelines, otherwise when coastseg a
 1. Click all the shorelines you want to delete
 2. Under **Remove Feature from Map** select 'selected shorelines' from the drop-down
 3. Click the 'Remove Selected' button
-   ![how_to_delete_parts_shoreline](https://github.com/Doodleverse/CoastSeg/assets/61564689/6097ff6c-69e2-4f9f-8637-bec5d21bc988)
-
-## How to Load Features from Geojson Files on Map
-
-To load shorelines, transects or bounding boxes onto the map from a geojson file use the `Load <feature name> file` button. It will open a file dialog window where you can select the geojson file you want to load on the map. If the geojson file is too large or has too many large features the map may become very slow. It is highly recommended to only load small features onto the map.
 
 ![how_to_edit_reference_shoreline_map](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/23871592-5ea6-4572-bf03-35a4bcae2cfa)
+
+## How to Load Shorelines from Geojson Files on the Map
+
+See this [guide](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/) on how to create and load your own reference shorelines onto the map :
