### 2024-05-06 dc3bf4f
migrate gifs for wiki how to filter bad imagery and create ref shorelines #244
