### 2024-05-06 80393e4
fix typo in basic install guide
