### 2024-05-06 ac4bbef
fix typo with uninstall transformers
