### 2024-05-07 SatelliteShorelines/CoastSeg#245 @2320sharon
Hi @FlorisCalkoen,

Thank you for your interest in the CoastSeg's coastal DL model and for detailing your setup. Regarding the installation of the transformers library, the general recommendation for most users is to use:

```
pip install transformers

```

This method has proved effective in most cases. Our team has explored a few alternative methods such as pip install transformers[tf-cpu], but this specific installation option has not been successful in our tests and might not be compatible with the TensorFlow version you are using.

We understand the importance of supporting various operating systems and installation methods. However, due to the complexity involved in maintaining compatibility between transformers and TensorFlow across different platforms, our resources are currently limited. Specifically, our team does not have access to Mac computers, which restricts our ability to test and provide tailored support for MacOS.
