### 2024-05-07 SatelliteShorelines/CoastSeg#245 @2320sharon
I want to provide some additional context as to why CoastSeg isn't using other versions of tensorflow or transformers.

 Coastseg deliberately utilizes only the CPU versions of machine learning libraries like Tensorflow because CoastSeg doesn't rely on the GPU for model inference.  We have attempted on multiple occasions to integrate GPU-accelerated workflows into CoastSeg. However, the complexity of setting up CUDA drivers and the inconsistent support for TensorFlow across various platforms have presented significant challenges. After careful consideration, we decided to prioritize a CPU-only approach to avoid these complexities and ensure that CoastSeg remains as compatible and user-friendly as possible across all supported platforms.
