### 2024-05-07 a5eb3b4
#244 update imgs in download tide model script and add a new contribution guide for testing zoo workflow
