+++ docs/How-to-Download-Tide-Model.md
@@ -13,7 +13,8 @@ Step 1: Register for an AVISO+ Account
 3. Make sure to select FES2014 / FES2012 (Oceanic Tides Heights) in the model selection section.
 4. Submit the form and wait for a confirmation email.
 
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/5f188ced-0f7e-4732-b065-8eb8f2d54986)
+![aviso_select_product](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/d1aa062e-d4f6-40a1-8dea-efa23c6ac386)
+
 <br>
 
 ## Step 2: Verify Your AVISO+ Account
@@ -27,9 +28,9 @@ Step 1: Register for an AVISO+ Account
 2. Navigate to My products in the left side menu.
 3. Confirm that FES2014 / FES2012 (Oceanic Tides Heights) is listed under Your current subscriptions.
    <br>
-   <img src="https://github.com/Doodleverse/CoastSeg/assets/61564689/bf5382f0-6bc0-4867-893d-c8f84a3d3760" alt="drawing" width="900" height="500"/>
+   <img src="https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/26118c1c-87a0-4754-85f0-ee9d141f3fcf" alt="drawing" width="900" height="500"/>
 
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/d5758b7c-7a4a-4f48-8b0b-244a0f8d7d7a)
+![download_tide_model_select_product](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/a475fe2c-38b7-4337-9e10-90bf7cd115bb)
 
 # Part 2: Downloading and Clipping the Tide Model
 
@@ -57,11 +58,10 @@ jupyter lab Download_tide_model.ipynb
 4.**Check Results:** Confirm that the model files are correctly downloaded and clipped in the specified directory.
 
 - **Example Output from Downloading Tide Model Step in Notebook**
-
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/6f4b8998-2d9f-4284-b0bb-075634efd1e8)
+  ![download_tide_model_notebook](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/eecae8d2-cb5e-4e50-a587-3260ff9469b7)
 
 - **Example Output from Clipping Tide Model Step in Notebook**
-  ![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/7251f6ca-acd4-4dda-9429-28bb83daad04)
+  ![clip_tide_model](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/abba33ae-1c63-4c48-be14-4e51e8224870)
 
   5.**Downloaded Tide Model Contents**
 
@@ -80,22 +80,23 @@ mkdir tide_model
 cd scripts
 ```
 
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/abe482c6-e1d8-4d0b-adba-45ce6a0fd952)
+![activate_coastseg_tide-Model_download](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/9706d481-5538-456d-8c79-b80d0c27b055)
 
 ## Download & Clip
 
-1. **Run Download Script:** Execute the `aviso_fes_tides.py` script with the necessary parameters, including your AVISO credentials and target directory for the tide model.
+1.**Run Download Script:** Execute the `aviso_fes_tides.py` script with the necessary parameters, including your AVISO credentials and target directory for the tide model.
 
[…]
