+++ mkdocs.yml
@@ -60,6 +60,7 @@ nav:
       - How to Add a New Workflow: contribute-a-workflow.md
       - Getting Started to Contribute: Getting-Started_to_contribute.md
       - How to Contribute Transects: contribute-transects.md
+      - How to Test the Zoo Workflow: test-zoo-workflow.md
       # - Technical Documentation: technical-documentation.md
   - Tutorials:
       - How to authenticate with Google Earth Engine: how-to-auth.md
