### 2024-05-07 f00579d
#244 update the how to extract shoreline area guide with new screenshot links
