+++ docs/how-to-extract-shoreline-area.md
@@ -5,26 +5,29 @@ You can draw multiple Shoreline Extraction Area polygons if your shoreline is co
 
 ## Steps
 
-1. Navigate to `Draw Controls` and click `Shoreline Extraction Area`
-   </br>
+1.Navigate to `Draw Controls` and click `Shoreline Extraction Area`
+</br>
 
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/a8aa2126-a706-46fd-9460-ef3b9297db13)
+![select_shoreline_extraction_Area4](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/4fa41590-dbdb-46d7-a38f-4ec8b136da21)
 
-2. Click either the polygon or rectangle drawing tool on the map
+2.Click either the polygon or rectangle drawing tool on the map
 
-![draw_controls_shape](https://github.com/Doodleverse/CoastSeg/assets/61564689/5acba879-a527-45ad-981f-a2a7956fb47e) 3. Draw the Shoreline Extraction Area on the map (it will be shown in pink)
+![draw_shoreline_extraction_area_tools](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/1e796685-ee6f-4653-a38a-95f69a86b696)
+
+3.Draw the Shoreline Extraction Area on the map (it will be shown in pink)
 
 - draw as many of these as you like
 
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/215d4f1b-9ec2-43c9-9c7c-f75dd2757c82)
+![shoreline_extraction_Area](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/c2470bed-ef4c-4f94-9ae9-b4ec8ba6e813)
 
-4. (Optional) If you make a mistake navigate to `Remove Feature from Map` and select `Shoreline Extraction Area` from the dropdown to remove the Shoreline Extraction Area from the map
+4.(Optional) If you make a mistake navigate to `Remove Feature from Map` and select `Shoreline Extraction Area` from the dropdown to remove the Shoreline Extraction Area from the map
 
-![image](https://github.com/Doodleverse/CoastSeg/assets/61564689/cfc00702-cfff-4c30-b129-988681db1cc0)
+![remove shoreline extraction area](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/d011338b-894e-40e7-9b79-a127485c8fdc)
 
-5. (Optional) To switch back to drawing bounding boxes navigate to `Draw Controls` and click `Bounding Box`
-6. Click Extract Shorelines
+5.(Optional) To switch back to drawing bounding boxes navigate to `Draw Controls` and click `Bounding Box`
+
+6.Click Extract Shorelines
 
 - The Shoreline Extraction Area you drew will now be used to extract only the shorelines in this region
 
-![shoreline_extraction_demo_screenshot](https://github.com/Doodleverse/CoastSeg/assets/61564689/7b925a9a-c812-48ce-bbc4-2902ba4845a2)
+![shoreline_extraction_area_with_shorelines](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/5de6ef78-98c3-450e-b407-45fa5118060b)
