+++ mkdocs.yml
@@ -90,3 +90,4 @@ nav:
       # - CoastSeg: scripts-coastseg.md
       # - CoastSat: scripts-coastsat.md
       # - Zoo: model.md
+  - Roadmap: roadmap.md
