+++ docs/extract_shorelines.md
@@ -42,7 +42,7 @@ If its your first time extracting shorelines from imagery we recommend using a s
   3. Delete the reference shoreline
   4. Click Load Shoreline Button to load the reference shoreline in the remaining ROIs
 
-![ROI_removal_and_shoreline_loading](https://github.com/Doodleverse/CoastSeg/assets/61564689/0928145a-e5f4-4d3b-bda4-315e17b14263)
+![ROI_removal_and_shoreline_loading](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/9437c93d-096a-4a01-b0fd-9a97d228e8bd)
 
 **Option 2 : Load Your Own Shorelines**
 
@@ -50,7 +50,7 @@ If its your first time extracting shorelines from imagery we recommend using a s
 2. Load shorelines from your geojson file
    - Check out the guide for how to do this here [How to Load Features from Geojson Files on Map](https://github.com/SatelliteShorelines/CoastSeg/wiki/04.-How-to-Use-the-Map#how-to-load-features-from-geojson-files-on-map)
 
-![how_to_make_ref_shoreline](https://github.com/Doodleverse/CoastSeg/assets/61564689/3cd70302-9bc0-411c-87f1-b25d86ac2280)
+![how_to_load_default_sl](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/c26b14a8-fef9-462c-8ed5-cad93f1f3d25)
 
 ## Step 2: Load Transects
 
@@ -77,11 +77,11 @@ CoastSeg extract shorelines from your imagery by checking if the water line inte
 
 If most of your imagery looks like these examples you may need to change your ROI.
 
-| Example Image                                                                                                               | Description                                                | Solution                                                            |
-| --------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------- | ------------------------------------------------------------------- |
-| ![1985-09-17-15-02-56_RGB_L5](https://github.com/Doodleverse/CoastSeg/assets/61564689/37d470e2-e2fe-4489-966f-a2a2384b928b) | ROI is too small with excessive water coverage.            | Make your ROI larger to include more land.                          |
-| ![2020-03-09-15-33-35_RGB_L8](https://github.com/Doodleverse/CoastSeg/assets/61564689/cb66cedb-9e90-446a-a6d1-3ac20f1ef0df) | Cloud masking failed and mistakenly masked the shoreline.  | Try turning off cloud masking or set 'cloud_mask_issue' to True     |
-| ![2020-07-16-15-51-52_RGB_S2](https://github.com/Doodleverse/CoastSeg/assets/61564689/78071e57-a6dd-4ea1-b9aa-9731b9ad3723) | Image is likely too cloudy to extract accurate shorelines. | Try using imagery with less clouds or try turning off cloud masking |
+| Example Image                                                                                                              | Description                                                | Solution                                                            |
+| -------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------- | ------------------------------------------------------------------- |
+| ![bad_img_too_small](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/949f11d7-64e5-4886-aee9-f6aec296ea39) | ROI is too small with excessive water coverage.            | Make your ROI larger to include more land.                          |
+| ![masked_out_clouds](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/39529ab1-33d6-4ef8-a2b4-5d80ec92cdbf) | Cloud masking failed and mistakenly masked the shoreline.  | Try turning off cloud masking or set 'cloud_mask_issue' to True     |
[…]
