### 2024-05-07 SatelliteShorelines/CoastSeg#244 @2320sharon
All of the installation instructions along with the wiki have all been migrated to the official CoastSeg website.  I also added specific [installation instructions for Mac users](https://satelliteshorelines.github.io/CoastSeg/mac-install-guide/).
