### 2024-05-07 SatelliteShorelines/CoastSeg#248 @FlorisCalkoen
Hi @2320sharon, thank you for your prompt response!

I see that most users are probably interested in satellite-derived shoreline series, but I'm also very interested in the shorelines, and probably there are others like me. I know that [Mao et al 2021](https://www.sciencedirect.com/science/article/abs/pii/S0924271621002598) actually computes shoreline change from shorelines without any transects --- so change analysis can also be performed at that level.

If you (still) decide to provide your users with one method I think it would be better to explictely call the method `extract shoreline series`  or even `extract shoreline series at transects`. IMHO, I think it's a general misconception (in the coastal monitoring field) to refer to SDS time-series as SDS -; to me it should be SDS for the shorelines (linestring geometries) and SDS series for the transect series (point geometries or translated position wrt to a reference point on a transect).

Do you save the shoreline geometries while extracting the series? Sorry I didn't check the data output yet.
