### 2024-05-07 SatelliteShorelines/CoastSeg#247 @2320sharon
Hi @FlorisCalkoen,

Sorry for the delay getting back to you. I was having some difficulty reproducing the error you showed here.

To attempt reproduce the error I did the following but I did not get the error you displayed here:
1. I loaded the sample transects you provided into a geodataframe and saved it to a geojson file

![image](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/3febc211-5c77-4331-b26c-3716dac735b9)

2. I used the load transects from file button to load in the transects

3. I drew a bounding box and loaded the default shoreline available in the region

4. I downloaded a small subset of imagery and extracted shorelines with no error (I did have to turn off cloud masking due to the cloud masking algorithm for LandSat masking out the shoreline)

5. I ran all these steps and did not get any error messages.

I ran all these steps on CoastSeg version 1.2.4 on a windows 11 computer.

Could you run `conda list coastseg` in your coastseg environment so I could see which coastseg version you're using?

I'm also going to attach the transect files I used.

[transects.zip](https://github.com/SatelliteShorelines/CoastSeg/files/15238665/transects.zip)
