### 2024-05-07 SatelliteShorelines/CoastSeg#247 @2320sharon
Hi @FlorisCalkoen,

Sorry for the delay getting back to you. I was having some difficulty reproducing the error you showed here, until I realized I was running CoastSeg 1.2.2 instead of 1.2.4 . I'm currently working on understanding this error and fixing it. Thanks for finding this
