### 2024-05-07 e548fef
fix tide correct so it won't break if none of the transects were in the raw timeseries.csv file
