### 2024-05-07 2c63130
Fix bug found in issue #247
