+++ src/coastseg/common.py
@@ -934,6 +934,8 @@ def filter_images_by_roi(roi_settings: list[dict]):
     it constructs a GeoDataFrame and filters images located in a predefined directory based
     on the constructed ROI. The function logs a warning and skips to the next ROI
     if the specified directory for an ROI does not exist.
+    
+    This function assumes the ROI coordinates are in EPSG:4326.
 
     Args:
         roi_settings (list[dict]): A list of dictionaries, each containing the settings
@@ -943,7 +945,7 @@ def filter_images_by_roi(roi_settings: list[dict]):
                                      'roi_id': <int>,
                                      'sitename': <str>,
                                      'filepath': <str>,  # Base filepath for the ROI
-                                     'polygon': <list>,  # List of coordinates representing the ROI polygon
+                                     'polygon': <list>,  # List of coordinates representing the ROI polygon (in EPSG:4326)
                                    }
 
     Returns:
@@ -1664,11 +1666,19 @@ def order_linestrings_gdf(gdf,dates, output_crs='epsg:4326'):
         GeoDataFrame: The ordered GeoDataFrame with linestrings.
 
     """
+    gdf = gdf.copy()
+    # Convert to the output CRS
+    if gdf.crs is not None:
+        gdf.to_crs(output_crs, inplace=True)
+    else:
+        gdf.set_crs(output_crs, inplace=True)
+        
     all_points = [shapely.get_coordinates(p) for p in gdf.geometry]
     lines = []
     for points in all_points:
         line_string = create_complete_line_string(points)
         lines.append(line_string)
+    
     gdf = gpd.GeoDataFrame({'geometry': lines,'date': dates},crs=output_crs)
     return gdf
 
@@ -1763,12 +1773,14 @@ def add_lat_lon_to_timeseries(merged_timeseries_df, transects_gdf,timeseries_df,
     
     # add the shoreline position as an x and y coordinate to the csv called shore_x and shore_y
     merged_timeseries_df = add_shore_points_to_timeseries(merged_timeseries_df, transects_gdf)
+    
     # convert to geodataframe
     merged_timeseries_gdf = gpd.GeoDataFrame(
         merged_timeseries_df, 
         geometry=[Point(xy) for xy in zip(merged_timeseries_df['shore_x'], merged_timeseries_df['shore_y'])], 
         crs="EPSG:4326"
     )
+    merged_timeseries_gdf.to_crs("EPSG:4326",inplace=True)
     if only_keep_points_on_transects:
         merged_timeseries_gdf,dropped_points_df = filter_points_outside_transects(merged_timeseries_gdf,transects_gdf,save_location,ext)
         if not dropped_points_df.empty:
@@ -1909,6 +1921,7 @@ def save_transects(
 
     # re-order columns
     merged_timeseries_df = merged_timeseries_df[['dates', 'x', 'y', 'transect_id', 'cross_distance']]
+    # add the shore_x and shore_y columns to the merged time series which are the x and y coordinates of the shore points along the transects
     merged_timeseries_df,timeseries_df = add_lat_lon_to_timeseries(merged_timeseries_df, transects_gdf.to_crs('epsg:4326'),cross_distance_df,
                               save_location,
                               drop_intersection_pts,
@@ -1987,15 +2000,24 @@ def convert_points_to_linestrings(gdf, group_col='date', output_crs='epsg:4326')
         gpd.GeoDataFrame: A new GeoDataFrame containing LineStrings created from the points.
     """
     # Group the GeoDataFrame by date
+    gdf = gdf.copy()
+    print(f"gdf.crs: {gdf.crs}")
+    # Convert to the output CRS
+    if gdf.crs is not None:
[…]
