### 2024-05-07 1a1cede
fix tide correction so that even if no intersections occured the rest of the tide corrections will proceed
