### 2024-05-07 7ec7140
fix #247 CRS mismatch with transects uploaded release v1.2.5
