### 2024-05-07 SatelliteShorelines/CoastSeg#246 @2320sharon
@FlorisCalkoen

I understand the confusion regarding the 'examples' directory. Originally I wanted to put our example scripts `1_download_imagery.py` and `2_extract_shorlines.py` in that folder, but doing so caused errors because the underlying coastsat code expects the `/data` to be at the same level as the script running.

Thank you for the suggestion to use [pooch](https://github.com/fatiando/pooch) we might consider using pooch to download the default transects and shorelines at some point so that the code to download the data won't need to be coded into coastseg.
