### 2024-05-07 SatelliteShorelines/CoastSeg#248 @2320sharon
@FlorisCalkoen

Thanks for bringing up the [Mao et al 2021](https://www.sciencedirect.com/science/article/abs/pii/S0924271621002598) research paper.  I can make an issue to allow CoastSeg to run without transects meaning it will output the 2D shorelines but none of the files related to shoreline change along the transects. This leaves this feature open to development.

We do save the shoreline geometries while extracting the series as both LineStrings [`extracted_shorelines_lines.geojson`], MultiPoints [`extracted_shorelines_points.geojson`] and as a dictionary [`extracted_shorelines_dict.json`] in the same format CoastSat does. One of the reasons we save them in these different formats is because we found from months of testing with our users that the multipoint format made it easier to see outliers (cloud detected as shorelines, ponds etc.). We also store the shorelines as LineStrings as well because it makes it easy to see the shoreline vector, but if clouds are detected as shoreline off shore this can results in strangely connected shorelines as seen CoastSat. The dictionary [`extracted_shorelines_dict.json`] contains all the data for the extracted shorelines in the same format CoastSat stored it except its no longer in a pickle format making it much more portable.

![image](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/8dd434e6-7122-4068-89d3-2f14e901e9d5)
