### 2024-05-07 39d6c45
fix typo in case_study_1 section 11 was not a header
