### 2024-05-07 4dfecd6
fix typo in the basic install guide
