### 2024-05-08 5eb76ad
update the extract shorelines guide with a figure and improve the intro
