### 2024-05-08 SatelliteShorelines/CoastSeg#248 @dbuscombe-usgs
In addition to the files you request already being available (i.e. the feature you request is already available), I will note that the Mao algorithm retrieves annual shoreline at high tide with image compositing methods, and is therefore not really a good analog for coastseg, which, as stated in numerous places, implements the CoastSat-style, transect-based, instantaneous shorelines mapped onto transects. The key advantage in the latter is data density, and high-frequency data are easily compared using a transect system. Transects can be as close together as you wish.

During extensive beta testing, our many users requested the 2d shorelines, so they are also outputted as standard, according to the graphic that Sharon shared above, as a convenience. But CoastSeg does not generally provide post-processing workflows - those tasks will be offloaded to another compatible toolbox to avoid software and dependency bloat.

We use the term SDS to mean Satellite Derived Shoreline, regardless of how it is computed. We also define this in the paper. This is the standard terminology adopted in the literature. We use waterline for uncorrected measurements, which is also standard - see this recent paper https://doi.org/10.1016/j.coastaleng.2024.104536

As for your terminology, namely SDS for the shorelines (linestring geometries) and SDS-series for the transect series, feel free to propose such a change in a research paper. This is just a software package adopting current standard terminology.
