### 2024-05-08 SatelliteShorelines/CoastSeg#246 @dbuscombe-usgs
We're talking about 32.3 kB of data, in just 4 files, that are named unambiguously.

Pooch is a fine idea for larger data files, but I personally don't think it is worth the extra dependency. What happens when pooch is no longer maintained? Also, why is python 3.11 hard-coded into their [requirements.yml](https://github.com/fatiando/pooch/blob/main/environment.yml)?
