### 2024-05-08 SatelliteShorelines/CoastSeg#245 @dbuscombe-usgs
Good points being discussed here, thanks.

@2320sharon I'm curious as to why `transformers` isn't listed in https://github.com/SatelliteShorelines/CoastSeg/blob/main/requirements.txt?

Floris, as you will have seen, our website Zoo [page](https://satelliteshorelines.github.io/CoastSeg/models/) is still under development. I recommend the following environment if you planned to contribute new DL models to this project. Thanks for the offer! That would be great. We don't have the bandwidth to train new models at the moment.

Assuming you are on Linux (recommended for any DL model training workflow - dont mess around with trying to get things to work on Mac or Windows, you'll be there all week)

set up miniconda:
```
sudo apt-get update
sudo apt-get install wget
wget https://repo.anaconda.com/miniconda/Miniconda3-py39_4.12.0-Linux-x86_64.sh
bash Miniconda3-py39_4.12.0-Linux-x86_64.sh
bash
```

then set up mamba:

```
conda update -n base conda
conda clean --all -y
python -m pip install --upgrade pip

conda install -n base conda-libmamba-solver
conda config --set solver libmamba
```

then configure your GPU dependencies

```
conda create -n gym_gpu python=3.10 -y
conda activate gym_gpu
conda install -c conda-forge cudatoolkit=11.8.0 -y
conda install -c nvidia cuda-nvcc -y

python3 -m pip install nvidia-cudnn-cu11 tensorflow[and-cuda]
```

then finally the rest of the dependencies
```
conda install -c conda-forge scikit-image ipython tqdm pandas natsort matplotlib -y
python -m pip install doodleverse_utils chardet

python -m pip install transformers
```

If you are contributing a new model, you'll be using Segmentation Gym, and these install instructions and more are detailed on the  [readme](https://github.com/Doodleverse/segmentation_gym) there
