### 2024-05-08 SatelliteShorelines/CoastSeg#247 @2320sharon
@FlorisCalkoen

Good news! I was able to replicate your bug and its now fixed in CoastSeg 1.2.5 .  Once you've updated your coastseg environment this issue should be fixed
