### 2024-05-09 a6147bb
add a guide on how to edit the website
