### 2024-05-09 230780c
#171 add code to support a good bad classifier for imagery it runs when shorelines are extracted
