### 2024-05-09 ff6a208
try to add best.h5 good/bad classifier
