+++ src/coastseg/classifier_model/coastal_image_classification.py
@@ -1,0 +1,331 @@
+"""
+Mark Lundine
+Good/bad coastal image classification
+adapted from https://keras.io/examples/vision/image_classification_from_scratch/
+"""
+import os
+import numpy as np
+import glob
+import tensorflow as tf
+from tensorflow import keras
+from tensorflow.keras import layers
+import pandas as pd
+import shutil
+import matplotlib.pyplot as plt
+    
+def load_dataset(training_data_directory,
+                 image_size,
+                 batch_size):
+    """
+    loads in training data in training and validation sets
+    inputs:
+    training_data_directory (str): path to the training data
+    returns:
+    train_ds, val_ds
+    """
+    train_ds, val_ds = tf.keras.preprocessing.image_dataset_from_directory(training_data_directory,
+                                                                        validation_split=0.2,
+                                                                        color_mode="grayscale",
+                                                                        subset="both",
+                                                                        seed=1337,
+                                                                        image_size=image_size,
+                                                                        batch_size=batch_size
+                                                                        )
+    return train_ds, val_ds
+
+def data_augmentation(images):
+    """
+    applies data augmentation to images
+    """
+    data_augmentation_layers = [
+        layers.RandomFlip("horizontal"),
+        layers.RandomFlip("vertical"),
+        layers.RandomRotation(0.25),
+        layers.RandomZoom(0.5, 0.5),
+        
+    ]
+    return images
+
+def define_model(input_shape, num_classes=2):
+    """
+    Defines the classification model
+    inputs:
+    input_shape (tuple (xdim, ydim)): shape of images for model
+    num_classes (int, optional): number of classes for the model
+    """
+    inputs = keras.Input(shape=input_shape)
+
+    # Entry block
+    x = data_augmentation(inputs)
+    # Entry block
+    x = layers.Rescaling(1.0 / 255)(inputs)
+    x = layers.Conv2D(128, 3, strides=2, padding="same")(x)
+    x = layers.BatchNormalization()(x)
+    x = layers.Activation("relu")(x)
+
+    previous_block_activation = x  # Set aside residual
+
+    for size in [256, 512, 728]:
+        x = layers.Activation("relu")(x)
+        x = layers.SeparableConv2D(size, 3, padding="same")(x)
+        x = layers.BatchNormalization()(x)
+
+        x = layers.Activation("relu")(x)
+        x = layers.SeparableConv2D(size, 3, padding="same")(x)
+        x = layers.BatchNormalization()(x)
+
+        x = layers.MaxPooling2D(3, strides=2, padding="same")(x)
+
+        # Project residual
+        residual = layers.Conv2D(size, 1, strides=2, padding="same")(
+            previous_block_activation
+        )
+        x = layers.add([x, residual])  # Add back residual
+        previous_block_activation = x  # Set aside next residual
+
+    x = layers.SeparableConv2D(1024, 3, padding="same")(x)
+    x = layers.BatchNormalization()(x)
+    x = layers.Activation("relu")(x)
+
+    x = layers.GlobalAveragePooling2D()(x)
+    if num_classes == 2:
+        units = 1
+    else:
+        units = num_classes
+
+    x = layers.Dropout(0.25)(x)
+    outputs = layers.Dense(units, activation=None)(x)
+
+    return keras.Model(inputs, outputs)
+
+def train_model(model,
+                image_size,
+                train_ds,
+                val_ds,
+                model_folder,
+                epochs=100):
+    """
+    Trains the good/bad classification model
+    inputs:
+    model (keras model): this is defined in define_model()
[…]
