### 2024-05-10 e1c0c37
add more pages to zoo model page
