### 2024-05-16 SatelliteShorelines/CoastSeg#250 @fmemuir
Small edit to git clone command in documentation 
For: https://github.com/openjournals/joss-reviews/issues/6683

Very minor issue, but trying to clone the repo using `git clone coastseg --depth 1 https://github.com/Doodleverse/CoastSeg.git` gives the error:
`fatal: repository 'coastseg' does not exist`
if the user has not made a folder called 'coastseg' first. Probably better to just have the command as:
`git clone --depth 1 https://github.com/Doodleverse/CoastSeg.git`
as a folder called 'CoastSeg' then gets created anyway.
