### 2024-05-16 SatelliteShorelines/CoastSeg#251 @fmemuir
"No rois found on the map" when redrawing a second/new ROI polygon
For: https://github.com/openjournals/joss-reviews/issues/6683

When an ROI is generated that doesn't have any transects available, generating subsequent ROIs causes a pop-up to appear saying "No rois found on the map".

### To recreate error
1. Draw an ROI around a place with no transects
2. Select bounding boxes of choice within ROI
3. Click 'Load Transects'
4. Close error pop-up and locate to a different place with known transects (e.g. Santa Cruz)
5. Draw new ROI
6. Click 'Generate ROI'

To solve this, the cell must be re-run to create a new dashboard. Is there a button or command to clear the cache of failed attempts to find an ROI with transects, rather than having to reload a new dashboard?
