### 2024-05-16 SatelliteShorelines/CoastSeg#250 @2320sharon
Hi @fmemuir

Thank you for reviewing CoastSeg. I'll update the documentation with this change.
