### 2024-05-17 SatelliteShorelines/CoastSeg#251 @2320sharon
Hi @fmemuir

Thank you for reviewing CoastSeg and for laying out the steps to replicate your issue. I was able to replicate the issue you encountered successfully.

## The Solutions

There are 2 ways to fix this issue:

1.  You can select 'shoreline' from the dropdown under 'Remove Feature From Map' and remove the shoreline at the old location from the map. You'll also need to remove the rois  as well, use the same dropdown, click 'ROIs' and remove the ROIs.
![image](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/f05cffc6-5c78-4bea-b96c-a747abddb488)

2. You can use the remove all button to remove all the features on the map then re-draw the bounding box & generate the ROIs again.

After removing the old features the new ROIs should generate just fine.

## Why it happened
To explain why this issue happened let me first explain what happens when the 'generate ROI' button is clicked. First, CoastSeg loads a shoreline within the bounding box drawn by the user and attempts to generate ROIs along the shoreline. If no shoreline is found in the bounding box or within any existing ROIs on the map, then the error message appears saying "No rois found on the map" because ROIs have to generated along a shoreline. What's happening in this case is the shoreline still exists at the first location where no transects were available. Since this shoreline is not within the bounding box where ROIs are supposed to be formed  the error message appears.

I'll work on patching out this issue so no else encounters this problem in the future. In the meantime the two solutions I outlined above will work.
