### 2024-05-17 f690ef2
fix #250 update the clone command on the website
