### 2024-05-17 SatelliteShorelines/CoastSeg#250 @2320sharon
Thank you for this suggestion its much more convenient to have a single CoastSeg folder for most users. The website is now updated with the new git clone command you suggested here. I appreciate it!
