### 2024-05-17 SatelliteShorelines/CoastSeg#251 @fmemuir
Thanks for the response and solution for this, I'd missed the 'Remove Feature From Map' buttons!
