### 2024-05-17 SatelliteShorelines/CoastSeg#248 @fmemuir
Just jumping on this thread (let me know if I should open a separate issue), to ask if available transects exist as a feature somewhere? It might be helpful to have locations with available transects (or even a super simple shoreline boundary) available to toggle on/off within the Leaflet map, so that beginner users can see where transects currently exist (without having to scroll around creating ROIs).
