### 2024-05-17 SatelliteShorelines/CoastSeg#248 @2320sharon
Hi @fmemuir

Good question, if a beginner use wants to see all the transects that are available in CoastSeg they either download the transects from the official [zenodo  release ](https://zenodo.org/records/8187949) we have linked in the readme or they can load from the geojson files located in `CoastSeg/src/coastseg/transects` into a free open source mapping software such as [QGIS](https://www.qgis.org/en/site/).

The main reason we don't render the transects on the map is because rendering thousands of vectors on ipyleaflet is very resource intensive. We've explored rendering the regions where transects are available with basic shoreline vectors, but given the large spread of transects we have this option isn't much better because it still would consume a lot of memory. We explored the idea of showing all the transects we presently have available as a static web map in our documentation, but since we are planning to add more transects to CoastSeg in the near future and continuously we decided against it because it would require frequent updating. The transects that CoastSeg provides by default are not required for the program to work. We designed CoastSeg so that users can upload their own transects, shorelines, rois, and bounding boxes if they find that CoastSeg either doesn't have or generate the features they want. The default shorelines and transects available are a 'fallback' for the user.

We provide a guide for users for how they can generate and upload their own transects and shorelines as well available on the [CoastSeg website](https://satelliteshorelines.github.io/CoastSeg/How-to-Create-Reference-Shorelines-%26-Transects%26ROIs/)

The current workflow to view the transects available only a takes a few seconds:
1. Draw a Bounding Box
2. Click  'load transects' Button
3. If none were available: Click 'Remove All' then repeat step 1 at a new location

This approach saves on a lot of memory since only a limited number of transects can be loaded at a given time. Unfortunately, rendering all the available transects on the map is non-trivial issue and would add significant overhead to the CoastSeg's performance.  Additionally this simplistic design frees the team behind CoastSeg to work on improvements to the shoreline mapping functionality and more.
