### 2024-05-21 SatelliteShorelines/CoastSeg#252 @FlorisCalkoen
Would it be possible to run the UI in the browser outside Jupyter? 
ping #openjournals/joss-reviews#6683

So far I have been mapping shorelines using the example notebooks in a JupyterLab IDE. This works, but sometimes searching for right settings becomes a bit tedious because the UI is relatively small in a Jupyter NB. You may consider running the UI in a web server instead of interacting with it through Jupyter. You could put the underlying code in a script that you launch through a CLI, e.g., `mamba activate coastseg; coastseg start` or `mamba activate coastseg; python -m coastseg main.py`. You may find some inspiration [here](https://panel.holoviz.org/gallery/)  - not necessasrily saying that you should port your code to holoviews, but it might give you a good idea on what's currently possible!
