### 2024-05-21 SatelliteShorelines/CoastSeg#247 @FlorisCalkoen
@2320sharon, many thanks for the fix, as I can confirm that with CoastSeg 1.2.5 I can map SDS with my own transects.
