### 2024-05-21 SatelliteShorelines/CoastSeg#253 @FlorisCalkoen
In SDS_zoo_classifier.ipynb the instructions and default settings are inconsistent
ping #openjournals/joss-reviews#6683

Just a small enhancement. In the SDS_zoo_classifier.ipynb you recommend in the introduction to use ensemble as the model, but this setting is actually hidden under the advanced settings, where `BEST` is the default setting. I recommend either changing the recommendation to `BEST` or change the default value to `ENSEMBLE`
