### 2024-05-21 SatelliteShorelines/CoastSeg#254 @FlorisCalkoen
SDS Zoo classifier: NameError: name 'TFSegformerForSemanticSegmentation' is not defined
ping #openjournals/joss-reviews#6683

I'm trying to run the CoastSeg BEST model to segment some RGB imagery. My workflow has been like this:

1. The imagery was downloaded using the
SDS_coastsat_classifier (although I don't think I'm using this imagery yet).
2. I try to select the directory with data from another session, but there I cannot find the RGB directory, so I'm guided towards `/Users/calkoen/dev/CoastSeg/data/ID_bza12_datetime05-06-24__07_09_58/jpg_files/preprocessed/RGB/`. So far all good.
3. When I click extract shorelines some things are happening, at least the first time I ran. Probably it downloads the model etc. However, eventuelly it fails with the posted error below. @2320sharon, do you maybe know what's causing this?

```bash
# packages in environment at /Users/calkoen/mambaforge/envs/coastseg:
...
coastseg                  1.2.5              pyhd8ed1ab_0    conda-forge
tensorflow                2.16.1                   pypi_0    pypi
tensorflow-io-gcs-filesystem 0.37.0                   pypi_0    pypi
...
```

```txt
Running the model. Please wait.

Running segformer_RGB_4class_8190958 model and extracting shorelines:   0%
 0/2 [00:00<?, ?it/s]

Running model segformer_RGB_4class_8190958

Downloading Downloading best_model.txt: 100%
 44.0/44.0 [00:00<00:00, 3.01kB/s]
Downloading sat4class_rgb_512_v5_segformer_fullmodel.h5: 100%
 14.4M/14.4M [00:07<00:00, 2.09MB/s]
...
```

But then I run into this error:
```python-traceback
File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py:1254](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py#line=1253), in Zoo_Model.get_model(self, weights_list)
   1253 try:
-> 1254     model = tf.keras.models.load_model(weights)
   1255     #  nclasses=NCLASSES, may have to replace nclasses with NCLASSES

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/keras/src/saving/saving_api.py:183](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/keras/src/saving/saving_api.py#line=182), in load_model(filepath, custom_objects, compile, safe_mode)
    182 if str(filepath).endswith((".h5", ".hdf5")):
--> 183     return legacy_h5_format.load_model_from_hdf5(
    184         filepath, custom_objects=custom_objects, compile=compile
    185     )
    186 elif str(filepath).endswith(".keras"):

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/keras/src/legacy/saving/legacy_h5_format.py:125](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/keras/src/legacy/saving/legacy_h5_format.py#line=124), in load_model_from_hdf5(filepath, custom_objects, compile)
    124 if model_config is None:
--> 125     raise ValueError(
    126         f"No model config found in the file at {filepath}."
    127     )
    128 if hasattr(model_config, "decode"):

ValueError: No model config found in the file at [/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/downloaded_models/segformer_RGB_4class_8190958/sat4class_rgb_512_v5_segformer_fullmodel.h5.](http://localhost:8888/Users/calkoen/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/downloaded_models/segformer_RGB_4class_8190958/sat4class_rgb_512_v5_segformer_fullmodel.h5.)

NameError                                 Traceback (most recent call last)
File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/ipywidgets/widgets/widget_output.py:103](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/ipywidgets/widgets/widget_output.py#line=102), in Output.capture.<locals>.capture_decorator.<locals>.inner(*args, **kwargs)
    101     self.clear_output(*clear_args, **clear_kwargs)
    102 with self:
--> 103     return func(*args, **kwargs)

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/models_UI.py:582](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/models_UI.py#line=581), in UI_Models.run_model_button_clicked(self, button)
    580 shoreline_path = self.fileuploader.files_dict.get("shorelines", "")
    581 shoreline_extraction_area_path = self.fileuploader.files_dict.get("shoreline extraction area", "")
--> 582 zoo_model_instance.run_model_and_extract_shorelines(
    583     self.model_dict["sample_direc"],
    584     session_name=session_name,
    585     shoreline_path=shoreline_path,
    586     transects_path=transects_path,
    587     shoreline_extraction_area_path = shoreline_extraction_area_path
    588 )

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py:733](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py#line=732), in Zoo_Model.run_model_and_extract_shorelines(self, input_directory, session_name, shoreline_path, transects_path, shoreline_extraction_area_path, use_tta, use_otsu, percent_no_data, use_GPU)
    728 prog_bar = tqdm.auto.tqdm(range(2),
    729     desc=f"Running {model_name} model and extracting shorelines",
    730     leave=True,
    731 )
    732 # run the model
--> 733 self.run_model(
    734     img_type,
    735     model_implementation,
    736     session_name,
    737     input_directory,
    738     model_name=model_name,
    739     use_GPU=use_GPU,
    740     use_otsu=use_otsu,
    741     use_tta=use_tta,
    742     percent_no_data=percent_no_data,
    743 )
    744 prog_bar.update(1)
    745 prog_bar.set_description_str(
    746                         desc=f"Ran model now extracting shorelines", refresh=True
    747 )

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py:1075](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py#line=1074), in Zoo_Model.run_model(self, img_type, model_implementation, session_name, src_directory, model_name, use_GPU, use_otsu, use_tta, percent_no_data)
   1072 logger.info(f"use_tta: {use_tta}")
   1074 print(f"Running model {model_name}")
-> 1075 self.prepare_model(model_implementation, model_name)
   1077 # create a session
   1078 session = sessions.Session()

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py:1011](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py#line=1010), in Zoo_Model.prepare_model(self, model_implementation, model_id)
   1008 weights_list = self.get_weights_list(model_implementation)
   1010 # Load the model from the config files
-> 1011 model, model_list, config_files, model_types = self.get_model(weights_list)
   1012 logger.info(f"self.TARGET_SIZE: {self.TARGET_SIZE}")
   1013 logger.info(f"self.N_DATA_BANDS: {self.N_DATA_BANDS}")

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py:1341](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/coastseg/zoo_model.py#line=1340), in Zoo_Model.get_model(self, weights_list)
   1339     for k in range(self.NCLASSES):
   1340         id2label[k] = str(k)
-> 1341     model = segformer(id2label, num_classes=self.NCLASSES)
   1342     model.compile(optimizer="adam")
   1343 # 242,812
   1344 else:

File [~/mambaforge/envs/coastseg/lib/python3.10/site-packages/doodleverse_utils/model_imports.py:63](http://localhost:8888/lab/tree/~/mambaforge/envs/coastseg/lib/python3.10/site-packages/doodleverse_utils/model_imports.py#line=62), in segformer(id2label, num_classes)
     60 label2id = {label: id for id, label in id2label.items()}
     61 model_checkpoint = "nvidia[/mit-b0](http://localhost:8888/mit-b0)"
---> 63 model = TFSegformerForSemanticSegmentation.from_pretrained(
     64     model_checkpoint,
     65     num_labels=num_classes,
     66     id2label=id2label,
     67     label2id=label2id,
     68     ignore_mismatched_sizes=True,
     69 )
     70 return model

NameError: name 'TFSegformerForSemanticSegmentation' is not defined
```
