### 2024-05-21 #6683 @cheginit
@FlorisCalkoen That's great! Thanks.

For comments on the paper, I'd just open one issue in their repo for all paper comments and add them there. You can either make them a check-list so you can checkmark them when they are addressed, or if you prefer, you can upload an annotated PDF of your comments in that issue.
