### 2024-05-21 SatelliteShorelines/CoastSeg#247 @2320sharon
@FlorisCalkoen that's great to hear! I'm going to close this issue now that it has been fixed.
