### 2024-05-21 SatelliteShorelines/CoastSeg#246 @2320sharon
@FlorisCalkoen Thank you for the suggestion to use a `configure_instance()` function that would solve a few of the problems we've been faced before. If you could open a PR with that code that would be awesome.
