### 2024-05-21 SatelliteShorelines/CoastSeg#252 @2320sharon
Hey @FlorisCalkoen we ran into a similar issue where some of our super users like @mlundine,  who are using CoastSeg on a few hundred sites, were finding that adjusting the settings manually was time-consuming. To make things faster we developed CoastSeg API to enable your own scripting, i.e., 'headless mode', which we have a tutorial on [here](https://satelliteshorelines.github.io/CoastSeg/CoastSeg-API-Guide/). Basically this lets you run CoastSeg as a script which makes testing settings much faster.

Thanks for your question as to why CoastSeg is not a dedicated web server application instead of a Jupyter instance. Early on we explored turning CoastSeg into a Holoviews application and one of our team members at the time worked on developing an [application](https://github.com/venuswku/data-visualizer) using Holoviews to visualize geospatial data. She ran into significant  performance issues rendering items on the map efficiently, difficulty creating custom components, and each of these issues took weeks to months to fix. The difficulty of developing with this software made us decide to stick with Jupyter as it is easy to customize, add new features to, and almost everyone already knows how to use. Moreover, and most importantly, as we state repeatedly in the paper, the community are encouraged to contribute notebooks to CoastSeg and add new features, hence we want to keep the Jupyter interface that makes this easy to do. We strongly feel Jupyter is widely accessible and familiar, so we plan to stick with it for the foreseeable.
