### 2024-05-21 SatelliteShorelines/CoastSeg#252 @FlorisCalkoen
Hi @2320sharon, many for your extensive answer and those scripts should indeed help a lot when doing a more extensive search towards finding the right settings. Please don't take all my issues as "these have to be addressed"; I'm following JOSS guidelines in opening issues for every thing I come accross as they have found that to be the most constructive way of working towards better software. I fully understand you don't want to port the whole thing to Holoviews! Do you know what performance issues you encountered before with Holoviews? Because I typically use Holoviews for big-data visualization (although the datashader makes working with geospatial features more difficult at an item level).
