### 2024-05-21 a366bb0
add some docstrings, improve error messages and organize loading from config
