+++ src/coastseg/coastseg_map.py
@@ -6,7 +6,7 @@ import logging
 import glob
 from datetime import datetime
 from collections import defaultdict
-from typing import Collection, Dict, List, Optional, Tuple, Union
+from typing import Any, Collection, Dict, List, Optional, Tuple, Union
 import traceback
 
 # Third-party imports
@@ -797,27 +797,30 @@ class CoastSeg_Map:
             "shoreline_extraction_area": ["geometry"],
         }
 
+        # attempt to load each feature type onto the map from the config_gdf.geojson file
         for feature_name, columns in feature_types.items():
             feature_gdf = self._extract_feature_gdf(gdf, feature_name, columns)
-            if feature_name == "roi":
-                exception_handler.check_if_gdf_empty(
-                    feature_gdf, "ROIs", "Cannot load empty ROIs onto map"
-                )
-                if self.rois is None:
-                    self.rois = ROI(rois_gdf=feature_gdf)
-                    self.load_feature_on_map(
-                        feature_name, gdf=feature_gdf, zoom_to_bounds=True
-                    )
-                elif self.rois is not None:
-                    # add the new roi to the existing rois
-                    self.rois = self.rois.add_geodataframe(feature_gdf)
-                    # load the new rois onto the map
-                    self.add_feature_on_map(self.rois, feature_name)
-            else:
-                # load shorelines, transects, or bbox features onto the map
+            
+            # if the feature is not an ROI use the load_feature_on_map method to load it. (eg. shorelines, transects, bbox)
+            if feature_name != "roi":
+                self.load_feature_on_map(feature_name, gdf=feature_gdf, zoom_to_bounds=True)
+                continue
+            
+            # at this point assume the feature is an ROI
+            exception_handler.check_if_gdf_empty(feature_gdf, "ROIs", "Cannot load empty ROIs onto map")
+
+            if self.rois is None:
+                # if no ROIs exist on the map, create a new ROI object and load the ROIs onto the map
+                self.rois = ROI(rois_gdf=feature_gdf)
                 self.load_feature_on_map(
-                    feature_name, gdf=feature_gdf, zoom_to_bounds=True
+                        feature_name, gdf=feature_gdf, zoom_to_bounds=True
                 )
+            else:
+                # add the new roi to the existing rois
+                self.rois = self.rois.add_geodataframe(feature_gdf)
+                # load the new rois onto the map
+                self.add_feature_on_map(self.rois, feature_name)  
+            
         del gdf
 
     def _extract_feature_gdf(
@@ -1080,7 +1083,7 @@ class CoastSeg_Map:
 
         """
         logger.info(f"Loading json config from filepath: {filepath}")
-        exception_handler.check_if_None(self.rois)
+        exception_handler.check_if_None(self.rois,"rois")
 
         json_data = file_utilities.read_json_file(filepath, raise_error=True)
         json_data = json_data or {}
@@ -1639,8 +1642,23 @@ class CoastSeg_Map:
         return self.get_settings()
         
 
-    def validate_transect_inputs(self, settings:dict, roi_ids:list=None):
-        # ROIs,settings, roi-settings cannot be None or empty
+    def validate_transect_inputs(self, settings: dict, roi_ids: list = None):
+        """
+        Validates the inputs for generating transects.
+
+        Args:
+            settings (dict): A dictionary containing the settings for generating transects.
+            roi_ids (list, optional): A list of ROI IDs. Defaults to None.
+
+        Raises:
+            ValueError: If any of the following conditions are met:
+                - The session name is an empty string.
+                - The ROIs, transects, or extracted shorelines are None.
+                - The extracted shorelines dictionary is empty.
+                - The 'along_dist' key is missing from the settings dictionary.
[…]
