### 2024-05-21 441a06a
helper function to infer data dir
