### 2024-05-21 fdcc737
helper function to infer data dir
