### 2024-05-21 SatelliteShorelines/CoastSeg#254 @2320sharon
Hi @FlorisCalkoen
I recognize this issue as its one our team spent weeks dealing with. The issue here is that the segformer library was not configured correctly in the CoastSeg environment and that's causing the `NameError: name 'TFSegformerForSemanticSegmentation' is not defined` because python can't recognize that the segformer library has been installed. If you run this command in the activated CoastSeg environment you'll get the same error `python -c "from transformers import TFSegformerForSemanticSegmentation"`, which means it wasn't configured correctly.

No one on the CoastSeg development team has a Mac which makes finding the environment that will successfully run the segformer model almost impossible.  You can see in this issue #136 it took us quite some time to discover the right order that the CoastSeg dependencies had to be installed in to get the segformer model to run for Windows and Mac.   If want to try to run the model on Mac you would need create a new CoastSeg environment and try installing the transformers library in a different way.

We mention in the [Mac Installation Guide](https://satelliteshorelines.github.io/CoastSeg/mac-install-guide/) that we can't support the zoo workflow on Mac due to the numerous compatibility issues with tensorflow and Mac that our team can't troubleshoot.
