### 2024-05-21 0b68d41
force roi to turn all incoming ids into type string
