### 2024-05-21 7a397c9
fix preview available imagery to accept ids
