### 2024-05-22 SatelliteShorelines/CoastSeg#256 @fmemuir
CoastSeg paper comments and edits
For: https://github.com/openjournals/joss-reviews/issues/6683

I'm attaching here an [annotated PDF](https://github.com/SatelliteShorelines/CoastSeg/files/15401873/10.21105.joss.06683_FMannotated.pdf) of the paper for review. The suggestions are minor and mainly grammatical, so I will hold off checking the last few tasks in the "Software paper" part of my checklist until these are addressed. One of my comments was on the referencing style, @cheginit is it customary for JOSS referencing to have papers with four authors appear in-text as "Lastname1, Lastname2 et al., 2020"? If so, please ignore those comments!
