### 2024-05-22 SatelliteShorelines/CoastSeg#256 @cheginit
> is it customary for JOSS referencing to have papers with four authors appear in-text as "Lastname1, Lastname2 et al., 2020"?

Yes, since there are two papers by the same first author and the same year, the second one is added to distinguish the two references.
