### 2024-05-22 SatelliteShorelines/CoastSeg#254 @2320sharon
@FlorisCalkoen
I'm not sure if this will work or not but try using tensorflow 2.15 `pip install tensorflow==2.15` then see if it works by running
```
python
from transformers import TFSegformerForSemanticSegmentation
```
