### 2024-05-27 SatelliteShorelines/CoastSeg#257 @FlorisCalkoen
CoastSeg manuscript review
ping #https://github.com/openjournals/joss-reviews/issues/6683

### Summary

Fitzpatrick et al. (2024) have made significant effort in improving community-driven satellite-derived shoreline mapping. The authors have refactored the well-established CoastSat software package, resulting in more robust and maintainable code. Additionally, they have developed Jupyter widgets that simplify shoreline mapping for less-technical users by providing them with a user interface. Furthermore, they provide access to a deep learning model trained for image segmentation.

Their efforts to streamline desktop shoreline mapping, a practice now standard in coastal science, are new and likely to make this package widely used. The authors' active engagement with the open-source coastal community further also underscores commitment to maintain this software in the future. I have reviewed the software, and most of it functions effectively. I would like to thank Fitzpatrick and Buscombe for their prompt responses on GitHub, ensuring the software also worked on my machine.

However, I do have a few comments regarding the manuscript and recommend addressing these before publishing.

### Software capabilities

While the manuscript presents the software as capable of landcover mapping, it primarily focuses on satellite-derived shorelines (SDS). The landcover mapping functionality appears to be less developed and undocumented in comparison to SDS capabilities. To manage expectations, I would recommend to clearly convey this to the reader, so by not specifically highlighting the landcover mapping capabilities (L11-12), but refrain to SDS.

Additionally, the manuscript mentions a new deep learning (DL) model, but this feature is not extensively covered. If the DL model is intended to be a key feature, it should be properly benchmarked using statistical performance indicators, so that users know when to choose a specific model. If the inclusion of this DL model is a separate effort, it may be better suited for a dedicated paper or feature release.  At this stage, I would expect that the DL model, that currently does not take the infrared bands into account, is performing less in wave-dominated environments, because it will probably map the surf zone instead of the shoreline, but I cannot see that - some performance metrics would be very useful!  Do you have some benchmark statistics on the Zoo model available? Then I would recommend including these into the paper!

Moreover, what about enhancing the hub with a standardized workflow to benchmark new models? Or maybe add that to the roadmap? Ideally, if you decide to include performance statistics on the DL model, I would recommend to develop a generic workflow (NB based), so that users can benchmark their SDS advancements. Would it, for example, be possible to make a simple connection point between the hub and the  iconic coastal sites that were used in Vos et al 2023?

### Literature

The current manuscript provides an overview of several satellite-derived shorelines (SDS) methods and applications, but it would be better to give a more comprehensive overview of historical work as several pioneering contributions have been overlooked. To engage with a broader audience and enhance the manuscript's depth, I suggest incorporating additional key references.

- [x] Firstly, the manuscript could benefit from including the earliest known SDS references. Please verify and consider adding the earliest references on lines 13-14.

	https://www.sciencedirect.com/science/article/pii/S0034425712001174?casa_token=cjDozJ4mn_MAAAAA:j1SB_wiVLoXGyT63NunDeZ2hPIaTL1bMpudTpPW_Ni_q9R3GUAKd68sOvss9iaWcASnFMxNw6RY

	https://www.sciencedirect.com/science/article/pii/S002532271530086

- [ ] Additionally, it is important to recognize a wider array of shoreline mapping efforts. The composite method for shoreline mapping is extensively discussed in Luijendijk 2018 and Hagenaars et al 2018. If the reference to Bishop-Taylor (2021) is there because that software is openly available, I would say that that is an additional feature that is worth mentioning separately.
- [x] Finally, if the reference shoreline that you provide globally is from Sayre et al 2018, I guess you should add a reference (and maybe check the license).

I guess that by incorporating these references, the manuscript will not only provide a more thorough historical overview but also engage with a wider audience.

### What is the difference between coastsat-package and CoastSat?

The coastsat-package on [PyPi](https://pypi.org/project/coastsat-package/) points to the CoastSat repository, but the description explains that its a slightly modified version. What are the differences? Where is this slightly-different version hosted? Who is going to maintain the coastsat-package? How are CoastSat community efforts cascaded into coastsat-package? What do we reference if we map shorelines using CoastSeg that wraps CoastSat underneath?

Description from the PyPi repository:

""""
This is coastsat-package the pip and conda package extension of [CoastSat](https://github.com/kvos/CoastSat). CoastSat-package is a slightly modified version of coastsat to make it compatible with [CoastSeg](https://github.com/Doodleverse/CoastSeg).
"""

### Minor

- L21-22: add a URL to GitHub repository?
- L50: IMHO, I don't consider CoastSeg a platform. I would say it's a (very extensive) collection of interactive widgets that run Jupyter notebook, that together are an application. Maybe you could use (Desktop) Jupyter UI/App?
- L177: doesn't is colloquial, maybe better "does not"
- 233-34: should be Normalized Difference Water Index
