### 2024-05-28 3dc0318
#256 #257 address comments and edit for the coastseg manuscript
