+++ paper/paper.bib
@@ -19,6 +19,13 @@
   publisher={Elsevier}
 }
 
+@misc{CoastSeg,
+  title={{CoastSat v2.4}},
+  author={Fitzpatrick, S. and Buscombe,D. and Lundine, M. and Warrick,J. and Vos, K.},
+  year={2024},
+  publisher={GitHub},
+  howpublished = {\url{https://github.com/kvos/CoastSat/releases/tag/v1.2.5}}
+  }
 
 @article{vos2023benchmarking,
   title={Benchmarking satellite-derived shoreline mapping algorithms},
