### 2024-05-29 SatelliteShorelines/CoastSeg#254 @dbuscombe-usgs
This is a known error, caused by not setting up the conda environment properly, or using a Mac. Since we've had now quite a few people verify this workflow on Windows and Linux, I vote for closing this issue
