### 2024-05-29 SatelliteShorelines/CoastSeg#253 @dbuscombe-usgs
@2320sharon I suggest removing the recommendation to use ENSEMBLE model, by removing "(Ensemble is recommended because it uses multiple models to create more accurate predictions)." from the preamble. We don't want to be overly prescriptive until the benchmark paper is out
