### 2024-05-29 SatelliteShorelines/CoastSeg#248 @dbuscombe-usgs
I believe this issue has been addressed and can be closed. Is that ok, @fmemuir?
