### 2024-05-29 SatelliteShorelines/CoastSeg#252 @dbuscombe-usgs
Seems like this can be closed.
