### 2024-05-29 SatelliteShorelines/CoastSeg#242 @dbuscombe-usgs
Mac was never supported, and JOSS even noted that ahead of review
