### 2024-05-29 SatelliteShorelines/CoastSeg#253 @2320sharon
@dbuscombe-usgs I'm updating it now. It should be alright to recommend Best for now because it runs a bit faster than ensemble right?
