### 2024-05-29 SatelliteShorelines/CoastSeg#253 @dbuscombe-usgs
Sure, or just leave it blank and allow users to discover what works for them. It's easy enough to do, just make a new session ...
