### 2024-05-29 SatelliteShorelines/CoastSeg#253 @2320sharon
I decided to give the zoo notebook an update while I changed the recommended model from ENSEMBLE to BEST. Below is the new instructions for the zoo notebook

Thanks for catching this @FlorisCalkoen

## How to Run Image Segmentation Models on the Downloaded Imagery
---

1. **Select Imagery Type**: Choose the type of imagery you have: **RGB**, **NDWI**, or **MNDWI**.
   - If you select **NDWI** or **MNDWI**, the **RGB** directory will automatically be converted to the selected type and saved to a new folder named after the selected type. For example, if you choose **NDWI**, a new directory called **NDWI** will appear next to the **RGB** directory.

2. **Choose a Model**: Select the image segmentation model you want to use.

3. **Select Performance Mode**: Choose between **Ensemble** or **Best** (Best is recommended for faster performance).
   - **Best**: Uses only the best model available on the images. This option is faster since it involves a single model making the prediction.
   - **Ensemble**: Utilizes all available models to create more accurate predictions. Although it takes longer to make a single prediction, the resulting segmentation is generally more accurate.

4. **Select Imagery Directory**: Choose the directory of imagery to run the model on.
   - Select an ROI (Region of Interest) directory under the `data` directory.
   - Always select a directory named 'RGB'.

## Need More Help?

Check out the CoastSeg website: [CoastSeg Models](https://satelliteshorelines.github.io/CoastSeg/models/)
