### 2024-05-30 17762d6
close issue In SDS_zoo_classifier.ipynb the instructions and default settings are inconsistent #253
