### 2024-05-30 7f0e302
fix the test_coastseg_map and test_import test so that it won't attempt to import models_UI if running on MAc OS
