### 2024-05-30 afa7a23
fix test_imports.py to not import zoo_model if Mac OS is used
