### 2024-05-30 3bd9944
Update the model description page for the zoo workflow page. Mention benchmarking, contributing models, and model stats
