### 2024-05-30 43860b5
update the models page and describe the spectral indices currently compatible
