+++ docs/models.md
@@ -26,6 +26,27 @@ pip install tensorflow
 pip install transformers
 ```
 
+# Model-Compatible Spectral Indices
+
+The models currently available with CoastSeg can use the following spectral indices. We welcome contribution for models that use different spectral indices, or any combination of bands. However, our team prefers our models to apply equally to all Landsat and Sentinel imagery, so ideally models are trained on those data, not some other data source.
+
+1.**Red Green Blue (RGB)**
+
+- These files are generated by default.
+
+  2.**Normalized Difference Water Index (NDWI)**
+
+- `NDWI = (Green - NIR) / (Green + NIR)`
+- Uses the green and NIR bands to enhance the appearance of open water features.
+- [Read more here](https://eos.com/make-an-analysis/ndwi/)
+
+  3.**Modified Normalized Difference Water Index (MNDWI)**
+
+- `MNDWI = (Green - SWIR) / (Green + SWIR)`
+- Uses green and SWIR bands to enhance open water features and diminish built-up area features that often correlate with open water in other indices.
+- Reduces vegetation effects better than NDWI and is more suitable for enhancing and extracting water information in regions with a background dominated by built-up land areas.
+- [Read more here](https://www.tandfonline.com/doi/full/10.1080/01431160600589179)
+
 ## Available Models
 
 ---
