### 2024-05-30 SatelliteShorelines/CoastSeg#253 @2320sharon
The notebook has been updated with the newest instructions posted here. Closing the issue
