+++ paper/paper.bib
@@ -9,6 +9,17 @@
   publisher={Elsevier}
 }
 
+@article{sayre2019global,
+  author = {Roger Sayre, Suzanne Noble, Sharon Hamann, Rebecca Smith, Dawn Wright, Sean Breyer, Kevin Butler, Keith Van Graafeiland, Charlie Frye, Deniz Karagulle, Dabney Hopkins, Drew Stephens, Kevin Kelly, Zeenatul Basher, Devon Burton, Jill Cress, Karina Atkins, D. Paco Van Sistine, Beverly Friesen, Rebecca Allee, Tom Allen, Peter Aniello, Irawan Asaad, Mark John Costello, Kathy Goodin, Peter Harris, Maria Kavanaugh, Helen Lillis, Eleonora Manca, Frank Muller-Karger, Bjorn Nyberg, Rost Parsons, Justin Saarinen, Jac Steiner and Adam Reed},
+  title = {A new 30 meter resolution global shoreline vector and associated global islands database for the development of standardized ecological coastal units},
+  journal = {Journal of Operational Oceanography},
+  volume = {12},
+  number = {sup2},
+  pages = {S47--S56},
+  year = {2019},
+  publisher = {Taylor \& Francis},
+  doi = {10.1080/1755876X.2018.1529714}
+
 @article{mclean202350,
   title={50 Years of Beach--Foredune Change on the Southeastern Coast of Australia: Bengello Beach, Moruya, NSW, 1972--2022},
   author={McLean, R. and Thom, B. and Shen, J. and Oliver, T.},
