+++ paper/paper.md
@@ -96,7 +96,7 @@ SDS workflows require manipulating various settings in order to extract optimal 
 
 ### Accessibility
 
-`CoastSeg` facilitates entirely browser-based workflows with an interactive webmap and `ipywidget` controls. It interfaces with the Zenodo API to download reference shorelines for any location in the world, organized into 5x5 degree chunks in GeoJSON format [@buscombe_2023_7786276]. CoastSeg also provides transects for specific locations, offering beachface slope metadata [@buscombe_2023_8187949] that is available when users hover over each transect with their cursor. We have improved the reliability of `CoastSeg` through rigorous error handling, which includes developer log files for in-depth diagnostics, user report files for transparency, and detailed error messages that provide guidance for troubleshooting and problem resolution. We have also provided a set of utility scripts for common data input/output tasks, often the result of specific requests from our software testers (see Acknowledgments). In addition to a project wiki and improved documentation, we have researched minimum, maximum, and recommended values for all settings, set suggested default values, and have provided visual project management aids.
+`CoastSeg` facilitates entirely browser-based workflows with an interactive webmap and `ipywidget` controls. It interfaces with the Zenodo API to download reference shorelines [@sayre2019global] for any location in the world, organized into 5x5 degree chunks in GeoJSON format [@buscombe_2023_7786276]. CoastSeg also provides transects for specific locations, offering beachface slope metadata [@buscombe_2023_8187949] that is available when users hover over each transect with their cursor. We have improved the reliability of `CoastSeg` through rigorous error handling, which includes developer log files for in-depth diagnostics, user report files for transparency, and detailed error messages that provide guidance for troubleshooting and problem resolution. We have also provided a set of utility scripts for common data input/output tasks, often the result of specific requests from our software testers (see Acknowledgments). In addition to a project wiki and improved documentation, we have researched minimum, maximum, and recommended values for all settings, set suggested default values, and have provided visual project management aids.
 
 ### Performance
 
