### 2024-05-30 cfbae62
update paper.bib with new references
