+++ paper/paper.bib
@@ -19,6 +19,7 @@
   year = {2019},
   publisher = {Taylor \& Francis},
   doi = {10.1080/1755876X.2018.1529714}
+  }
 
 @article{mclean202350,
   title={50 Years of Beach--Foredune Change on the Southeastern Coast of Australia: Bengello Beach, Moruya, NSW, 1972--2022},
