### 2024-05-30 9501efc
fix minor typos in the paper.md
