### 2024-05-30 SatelliteShorelines/CoastSeg#256 @2320sharon
Hi @fmemuir

Thank you for reviewing the CoastSeg manuscript. I've updated the CoastSeg manuscript with all of your suggested edits as well as the some of the edits suggested Floris Calkoen.
