### 2024-05-30 SatelliteShorelines/CoastSeg#257 @2320sharon
Hi @FlorisCalkoen

First of all thank you for reviewing CoastSeg. We appreciate the suggested improvements for CoastSeg and they have helped make CoastSeg a more robust piece of software. Thank you for working with us throughout this review process and for taking the time to understand the different design decisions behind CoastSeg.

> While the manuscript presents the software as capable of landcover mapping, it primarily focuses on satellite-derived shorelines (SDS). The landcover mapping functionality appears to be less developed and undocumented in comparison to SDS capabilities. To manage expectations, I would recommend to clearly convey this to the reader, so by not specifically highlighting the landcover mapping capabilities (L11-12), but refrain to SDS.

In the zoo workflow, CoastSeg first performs landcover mapping followed by a pixel-wise segmentation of the geospatial image to extract the shoreline. Applying a landcover mapping model to identify additional features in the image, such as sand, vegetation, whitewater, and water, to then find the shoreline is a common practice in SDS. In the manuscript, we do not claim that CoastSeg does anything beyond SDS. We anticipate future contributions to continue to focus solely on SDS, however, a future optimal solution to SDS may involve landcover mapping as an intermediate step, like the zoo workflow within CoastSeg presently does. We believe this is clear, and leave the reference to landcover mapping as-is. It is technically correct, and encourages the community to continue to explore landcover mapping (i.e. dense pixelwise predictions on maps) as a necessary intermediate step in SDS.

> Additionally, the manuscript mentions a new deep learning (DL) model, but this feature is not extensively covered. If the DL model is intended to be a key feature, it should be properly benchmarked using statistical performance indicators, so that users know when to choose a specific model. If the inclusion of this DL model is a separate effort, it may be better suited for a dedicated paper or feature release. At this stage, I would expect that the DL model, that currently does not take the infrared bands into account, is performing less in wave-dominated environments, because it will probably map the surf zone instead of the shoreline, but I cannot see that - some performance metrics would be very useful! Do you have some benchmark statistics on the Zoo model available? Then I would recommend including these into the paper!

The new deep learning model, used in the Zoo workflow mentioned in the manuscript, does not currently have benchmark statistics for its performance on coastal environments. We are working on a separate paper that will benchmark the models’ performance on a variety of coastal landscapes, including those used in the SDS benchmark. We plan to take a similar approach by co-author Kilian Vos, and create a separate repository and paper for benchmarking the deep learning models in CoastSeg. We have already provided preliminary validation statistics of the model with the published models (see the links we have made to the various models in the documentation). However, these only speak to the quality of image segmentation on a  limited validation dataset, and not the quality of extracted SDS, hence the need for a separate paper. This is a software paper, that describes the CoastSeg project. JOSS papers are short and we don't have the space to include all the details you demand, but also these details are out of scope. Finally, our deep learning models do use NIR and SWIR. As documented, we have developed several models; one that uses RGB only, one that uses MNDWI, and one that uses NDWI. We look forward to sharing this work in peer-reviewed paper format shortly. In the meantime we have updated our [website](https://satelliteshorelines.github.io/CoastSeg/models/) to state where to find the statistical performance indicators for each model, a description of the current spectral indices our models support, and an update on the status of our model benchmarking exercise.

> Moreover, what about enhancing the hub with a standardized workflow to benchmark new models? Or maybe add that to the roadmap? Ideally, if you decide to include performance statistics on the DL model, I would recommend to develop a generic workflow (NB based), so that users can benchmark their SDS advancements. Would it, for example, be possible to make a simple connection point between the hub and the iconic coastal sites that were used in Vos et al 2023?

CoastSeg is primarily designed for SDS workflows, and we plan to keep any developments for benchmarking models within a separate repository, similar to how CoastSat has [SDS Benchmark](https://github.com/SatelliteShorelines/SDS_Benchmark) to track its benchmarking efforts. We have [a guide for contributing models](https://satelliteshorelines.github.io/CoastSeg/contribute-a-workflow/), and users will be able to benchmark their models in the separate repository. That being said, benchmarking models is outside the scope of CoastSeg’s intended functionality.

> Firstly, the manuscript could benefit from including the earliest known SDS references. Please verify and consider adding the earliest references on lines 13-14.

Thank you for suggesting some additional references to include in the manuscript. Since this is a software paper, not a peer-reviewed research article, we believe our referencing is sufficient, including recent review papers that provide an overview of the state of the field. Since ours is an extension of open source SDS software, naturally we reference only those open source contributions.

> Finally, if the reference shoreline that you provide globally is from Sayre et al 2018, I guess you should add a reference (and maybe check the license).

We added the reference for the Sayre et al. 2018 shorelines to the paper. The license linked [here](https://www.tandfonline.com/action/showCopyRight?scroll=top&doi=10.1080%2F1755876X.2018.1529714) states that it is permissible for non-commercial use, which aligns with our intended usage.

> What is the difference between coastsat-package and CoastSat? The coastsat-package on PyPi points to the CoastSat repository, but the description explains that its a slightly modified version. What are the differences? Where is this slightly-different version hosted? Who is going to maintain the coastsat-package? How are CoastSat community efforts cascaded into coastsat-package? What do we reference if we map shorelines using CoastSeg that wraps CoastSat underneath?

In the paper, we listed a number of improvements we have made to the original CoastSat workflow. This forms a whole section of the paper, in fact, so we're a little unsure of the confusion. Simply put, CoastSeg relies on the coastsat-package because CoastSat does not currently have an officially supported package available. The coastsat-package is a slightly modified version of CoastSat, with changes documented under the [What’s Changed](https://github.com/SatelliteShorelines/coastsat_package) section on the package's ReadMe page (as well as listed in the paper at a higher level). These modifications include minor organizational adjustments, additional parameters for functions to give end users more control, and changes to the create jpg function to default to creating a jpg. The design of the coastsat-package ensures it remains closely aligned with CoastSat's workflow while providing added utility for CoastSeg.

The CoastSeg team presently maintains the coastsat-package, but since it is open-source, contributions from the broader community are welcome and encouraged. The package's description and homepage have been updated to link to the actual coastsat-package repository instead of directly to CoastSat. When changes are made to CoastSat, the coastsat-package maintainers updates the relevant code to maintain synchronization between the two. Consequently, any references to CoastSat using the coastsat-package should be considered references to CoastSat itself.
