### 2024-06-03 SatelliteShorelines/CoastSeg#257 @FlorisCalkoen
@2320sharon,

Thank you very much for your detailed answer! Although I'll leave it up to you and the editor @cheginit to decide, I would like say that I have some remaining concerns on a few points.

- I believe the manuscript would benefit from referencing key SDS studies from a broader scientific community, even though it is a software paper.
- It is not ideal to release scientific software without benchmark statistics. While it is promising to hear that you are preparing a manuscript on this topic, relying on future publications is risky and it eventually might impact downstream science. I appreciate that you have linked to a [Zenodo repository](https://zenodo.org/records/6950472) for the ML model. However, the performance metrics provided as numpy arrays may not be user-friendly for non-programmers or even experts, as these arrays do not have any labeling. It would be more accessible to present these metrics directly on the CoastSeg website or as text in the Zenodo repository.
- Finally, the license that is used by Sayre et al 2018 states that:
"""
This is an Open Access article distributed under the terms of the Creative Commons Attribution-NonCommercial-NoDerivatives License (http://creativecommons.org/licenses/by-nc-nd/4.0/), which permits non-commercial re-use, distribution, and reproduction in any medium, provided the original work is properly cited, and is not altered, transformed, or built upon in any way.
"""

1. I'm not entirely sure how you use the coastline exactly, but probably you have to check that you do not break their clause.  I read somwhere the NoDerivates implies that "If you remix, transform, or build upon the material, you may not distribute the modified material."
2. You are using a GNU License, which is more open and conflicts with with the license used by Sayre et al 2018.

Again, I'm not entirely sure how you use your coastline, but this is probably an action point to check. Maybe you can reconsider your use of the coastline? Or discuss this with Sayre et al 2018?  Or adopt seperate licensing?
