### 2024-06-03 SatelliteShorelines/CoastSeg#257 @2320sharon
@FlorisCalkoen

Thank you for your detailed review of CoastSeg. I hope to address some of the points you raised here:

- Regarding your suggestion about citing the additional paper, we have opted not to do so because the dataset used in that paper is not open source. There are many SDS papers going back more than 10 years and we do not have sufficient space for a comprehensive literature review, however we link to recent review and benchmark papers that better review the current state of the science.

 - We will update the CoastSeg website with the validation statistics and metrics from the zoo. Additionally, we plan to provide even more detail as we continue our model validation exercises.

-  Thank you for raising the question about the license used by Sayre et al. (2018). In CoastSeg, we do not transform or alter the data from the coastline; instead, we simply segment the larger coastline into smaller chunks. Since we credit the original authors of the dataset and do not modify the coastline data, we believe we are adhering to the terms stated in the license. @cheginit what do you think about this?

Thank you again for reviewing CoastSeg and for all your help.
