### 2024-06-04 5e5da50
#257 #246 refactor coastseg to infer the CoastSeg directory
