### 2024-06-04 111d8ae
Fix the issue where ROI IDs were not converted to strings if passed as ints
