### 2024-06-04 d14fce4
#246 update the download and extract shorelines script to use new infer project dir codes
