### 2024-06-04 7ffdf9d
#259 fix the Kmeans cluster filter
