### 2024-06-04 e7d8e89
#259 #197 review kmeans filtering, update no data filtering, move predicition segmentation png to good/bad & npz to good/bad
