### 2024-06-06 24d6ba6
replace COPERNICUS/S2 with COPERNICUS/S2_HARMONIZED
