### 2024-06-06 a631569
Squashed commit of the following:
commit 4a270e1ba2b8c554d00e2e9d7b1d2e17a754aade
Merge: 64bd0b0 d14fce4
Author: Sharon Fitzpatrick <SF2309@Gmail.Com>
Date:   Thu Jun 6 07:04:33 2024 -0700

    Merge branch 'pr-255' into new_release/1.2.6

commit 64bd0b0d2f6ab286815daf78d19a4ec63ba4a1b4
Merge: 2f84fa2 e7d8e89
Author: Sharon Fitzpatrick <SF2309@Gmail.Com>
Date:   Thu Jun 6 07:01:40 2024 -0700

    Merge branch 'issue/259/get_kmeans_cluster' into new_release/1.2.6

commit 2f84fa2b120d415fc1dd737ee85174da58b2c315
Merge: c0b2d01 606488f
Author: Sharon Fitzpatrick <SF2309@Gmail.Com>
Date:   Thu Jun
[…]
