### 2024-06-06 dc98c77
move the PR code to get_base_dir & is_interactive to core_utilities.py
