### 2024-06-08 SatelliteShorelines/CoastSeg#257 @cheginit
@FlorisCalkoen Thanks for providing a thorough review of the paper with constructive comments.

@2320sharon Thanks for working on addressing the issues raised by @FlorisCalkoen.

@2320sharon, in response to:

> To manage expectations, I would recommend to clearly convey this to the reader, so by not specifically highlighting the landcover mapping capabilities (L11-12), but refrain to SDS.

you stated:

> We believe this is clear, and leave the reference to landcover mapping as-is. It is technically correct, and encourages the community to continue to explore landcover mapping

But, after reading the paper, I agree with @FlorisCalkoen, this point is not clear to me as well. Please make the necessary changes to address this issue.

Regarding this point that you made:

> We are working on a separate paper that will benchmark the models’ performance on a variety of coastal landscapes, including those used in the SDS benchmark. We plan to take a similar approach by co-author Kilian Vos, and create a separate repository and paper for benchmarking the deep learning models in CoastSeg.

Please add a few sentences to the paper (and documentation), explaining this point, so user are aware of the state of this part of the software before using it.

About the benchmarking issue, I agree with these points:

> CoastSeg is primarily designed for SDS workflows, and we plan to keep any developments for benchmarking models within a separate repository, similar to how CoastSat has [SDS Benchmark](https://github.com/SatelliteShorelines/SDS_Benchmark) to track its benchmarking efforts. We have [a guide for contributing models](https://satelliteshorelines.github.io/CoastSeg/contribute-a-workflow/), and users will be able to benchmark their models in the separate repository. That being said, benchmarking models is outside the scope of CoastSeg’s intended functionality.

> We will update the CoastSeg website with the validation statistics and metrics from the zoo. Additionally, we plan to provide even more detail as we continue our model validation exercises.

Please reflect this properly in the paper and documentation, so users know about the ongoing efforts.

On the issue of adding new citation and references, note that we have soft limit on the number of pages for the paper. So, please add the requested references, you don't have to provide a comprehensive literature review, but notable, important, and relevant previous work (regardless of their publication date) should be cited.

About this:

> In the paper, we listed a number of improvements we have made to the original CoastSat workflow. This forms a whole section of the paper, in fact, so we're a little unsure of the confusion.

Please make the necessary changes to the paper and the documentation to clarify the differences and improvements, then ask @FlorisCalkoen if the changes address the confusion.

Regarding the license issue, I will get back to you. In the meanwhile, please also consult with other license experts, just to make sure we are seeking opinions of several people on this.

As a general note, whenever you make any changes to address a specific comment, whether in the paper or the software, it's important to clearly reference the changes that you made in the repo and ask @FlorisCalkoen if the change addresses the issues.
