### 2024-06-09 SatelliteShorelines/CoastSeg#260 @fmemuir
Doc tree specifically for 'data' and 'sessions' 
For [review 6683](https://github.com/openjournals/joss-reviews/issues/6683)

Small suggestion for the wiki, to make it clearer on the filing structure of where imagery and outputs are stored. The 'data' folder is mentioned in the Tutorials and Case Studies when bad images are filtered out, but it is only clear from the screenshots where the RGB folder of imagery sits and that this is the one to filter images out of. Additionally, the difference between 'data' and the output data stored in 'sessions' could be clearer; it's not immediately obvious that the extracted data is here. It's recommended that the wiki is given a short explanation (or even just a document tree fig) of where different files are being stored when the shorelines are extracted.
