### 2024-06-09 SatelliteShorelines/CoastSeg#261 @fmemuir
Extract shorelines as multipart lines instead of points (for better representation of broken lines)
For [review 6683](https://github.com/openjournals/joss-reviews/issues/6683)

When shorelines are extracted as points and then exported as geoJSON lines, shorelines that have breaks in them due to the shoreline shape or cloud objects are not preserved and points are threaded together across breaks. For example, at a harbour entrance, shoreline points are extracted and able to be visualised, but shoreline lines thread straight across the entrance, which means an incorrect map representation and extra work to manually create multipart lines with breaks in them.

**Describe the solution you'd like**
Ideally, shoreline lines should be separate line objects with shared attributes (such as dates, image names and ROI IDs). One way to achieve this would be to convert contours to multiline features at the 'process shorelines' step, using a format like a geopandas GeoDataFrame. The multiline feature would (for the harbour entrance example) be made up of one shoreline with one set of attributes, but encompassing two separate LineString objects. The GeoDataFrame could then be exported to a geoJSON (or alternative file format) where the separate LineString objects all create individual features, but share the same attributes.
