### 2024-06-10 SatelliteShorelines/CoastSeg#257 @2320sharon
Hi @cheginit

Thank you for reviewing this issue and your input on the issue. I appreciate your guidance throughout this process. I will address each of the points raised and update the issue accordingly.

To begin:
> To manage expectations, I would recommend to clearly convey this to the reader, so by not specifically highlighting the landcover mapping capabilities (L11-12), but refrain to SDS.

I will address this issue by removing the reference to landcover mapping capabilities from the paper.

Regarding the benchmarking mentioned:
> We are working on a separate paper that will benchmark the models’ performance on a variety of coastal landscapes, including those used in the SDS benchmark. We plan to take a similar approach by co-author Kilian Vos, and create a separate repository and paper for benchmarking the deep learning models in CoastSeg.

I have already updated our documentation for the zoo workflow to state that:

> These models have not been thoroughly tested yet, but we are currently undergoing the process of benchmarking these models in a variety of coastal environments. We will be documenting the results of this benchmark in a separate repository and we will link it here when its ready.

This can be found on the documentation website [here ](https://satelliteshorelines.github.io/CoastSeg/models/).  I'll also add a sentence to the Project Roadmap section of the paper that explains the models used in the deep learning based SDS workflow are actively being benchmarked at several locations.

Regarding the new citations and references:
> On the issue of adding new citation and references, note that we have soft limit on the number of pages for the paper. So, please add the requested references, you don't have to provide a comprehensive literature review, but notable, important, and relevant previous work (regardless of their publication date) should be cited.

I'll add the requested references to the paper.

Regarding coastsat_package:

> In the paper, we listed a number of improvements we have made to the original CoastSat workflow. This forms a whole section of the paper, in fact, so we're a little unsure of the confusion.

> Please make the necessary changes to the paper and the documentation to clarify the differences and improvements, then ask @FlorisCalkoen if the changes address the confusion.

I'll make a page on the CoastSeg website that explains why we use `coastsat_package` instead of `CoastSat` and link the `coastsat_package`  repository that lists the improvements & changes made to it.  I will check with Floris Calkoen if this resolves the confusion.

As for the license issue:
> Regarding the license issue, I will get back to you. In the meanwhile, please also consult with other license experts, just to make sure we are seeking opinions of several people on this.

Thank you for looking into the license issue. We are confident that we are not violating the terms of the license as we do not modify the data and provide multiple citations of the source. We will consult other experts to confirm this.

I'll be sure to tag @FlorisCalkoen to validate that the changes I make address the issues raised.

Thank you both for your time and for all your help in this review process.
