### 2024-06-10 SatelliteShorelines/CoastSeg#260 @2320sharon
Hi @fmemuir

Thank you for the suggestions. These are excellent ideas. I'll update the wiki pages to explain why the downloaded imagery and sessions are separated and where the outputs go when the shoreline extraction process occurs. This should make the idea of using sessions easier for new users and help them understand where their data is stored.

## Tasks

- [ ] Explain the difference between 'data' and `sessions` in the `Getting Started`, Case Studies and `How to use Sessions` guides in the wiki

- [ ] Add a doctree to explain the file structure of a session to `How to use Sessions` and `How to Extract Shorelines` guides

- [ ] Explain how the downloaded data from 'data' is used in a session in the `Getting Started`, Case Studies and `How to use Sessions` guides in the wiki

- [ ] Add a doctree to the [How to Filter Bad Imagery](https://satelliteshorelines.github.io/CoastSeg/How-to-Filter-Out-Bad-Imagery/) and link to guide in [case_study_2](https://satelliteshorelines.github.io/CoastSeg/case_study_2/)
