### 2024-06-10 455a9d2
#257 add two references, remove landcover references from paper, and add disclaimer that models are undergoing benchmarking
