+++ paper/paper.bib
@@ -9,6 +9,30 @@
   publisher={Elsevier}
 }
 
+@article{luijendijk2018state,
+  title={The State of the World's Beaches},
+  author={Luijendijk, Arjen and Hagenaars, Gerben and Ranasinghe, Roshanka and Baart, Fedor and Donchyts, Gennadii and Aarninkhof, Stefan},
+  journal={Scientific Reports},
+  volume={8},
+  number={1},
+  pages={6641},
+  year={2018},
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41598-018-24630-6},
+  url={https://doi.org/10.1038/s41598-018-24630-6}
+}
+
+@article{pardopascual20121,
+title = {Automatic extraction of shorelines from Landsat TM and ETM+ multi-temporal images with subpixel precision},
+journal = {Remote Sensing of Environment},
+volume = {123},
+pages = {1-11},
+year = {2012},
+doi = {https://doi.org/10.1016/j.rse.2012.02.024},
+url = {https://www.sciencedirect.com/science/article/pii/S0034425712001174},
+author = {Josep E. Pardo-Pascual and Jaime Almonacid-Caballer and Luis A. Ruiz and Jesús Palomar-Vázquez}
+}
+
 @article{sayreEtAl2019,
 author = {Roger Sayre, Suzanne Noble, Sharon Hamann, Rebecca Smith, Dawn Wright, Sean Breyer, Kevin Butler, Keith Van Graafeiland, Charlie Frye, Deniz Karagulle, Dabney Hopkins, Drew Stephens, Kevin Kelly, Zeenatul Basher, Devon Burton, Jill Cress, Karina Atkins, D. Paco Van Sistine, Beverly Friesen, Rebecca Allee, Tom Allen, Peter Aniello, Irawan Asaad, Mark John Costello, Kathy Goodin, Peter Harris, Maria Kavanaugh, Helen Lillis, Eleonora Manca, Frank Muller-Karger, Bjorn Nyberg, Rost Parsons, Justin Saarinen, Jac Steiner and Adam Reed},
 title = {A new 30 meter resolution global shoreline vector and associated global islands database for the development of standardized ecological coastal units},
