### 2024-06-11 59a7c3b
update sessions to explain where the session folder is and what each of the files do
