+++ docs/sessions.md
@@ -3,6 +3,96 @@
 A session is a saved state that CoastSeg was in. A session can be used to load in the ROIs that were downloaded previously or the previously extracted shorelines.
 A session is typically stored as a directory containing a minimum of two files a `config.json` and a `config_gdf.geojson` file that lets you restore the state coastseg was in.
 
+### Where are sessions located?
+
+Sessions are saved to `CoastSeg/sessions`. A session contains the all the files related to extract shorelines for a particular set of ROIs in the downloaded data located in `CoastSeg/data`. Each session follows a similar format depending on what actions were taken by the user. For instance, if tide correction was not applied during the session then none of the files with the tidal correction applied will be present.
+
+Here is a sample session that had 2 ROIs where shorelines were successfully extracted and tide correction was applied to both ROIs.
+
+```
+├── CoastSeg
+|
+|___session
+|    |_ ID_4_datetime11-22-22__11_15_15
+│   |   |_ jpg_files
+│   |   |  |_ detection # this folder contains images of the extracted shorelines overlaid on the original scene
+│   |   |  |  |_ 2023-12-09-18-40-08_L9.jpg   # this is the shoreline extracted from satellite L9 on 2023-12-09 at 18:40:08
+│   |   |  |  |_ <rest of shoreline detection images>
+│   |   |_config.json
+│   |   |_config_gdf.json
+│   |   |_extracted_shorelines_lines.geojson
+│   |   |_extracted_shorelines_points.geojson
+│   |   |_raw_transect_time_series.csv
+│   |   |_raw_transect_time_series_merged.csv
+│   |   |_raw_transect_time_series_points.geojson
+│   |   |_raw_transect_time_series_vectors.geojson
+│   |   |_tidally_corrected_transect_time_series.csv        # this file will only appear if tidal correction was applied
+│   |   |_tidally_corrected_transect_time_series_merged.csv # this file will only appear if tidal correction was applied
+│   |   |_tidally_corrected_transect_time_series_points.geojson # this file will only appear if tidal correction was applied
+│   |   |_tidally_corrected_transect_time_series_vectors.geojson # this file will only appear if tidal correction was applied
+│   |   |_shoreline_settings.json
+│   |   |_transects_cross_distances.json
+│   |   |_transects_settings.json
+│   |   |
+|   |_ ID_3_datetime11-22-22__11_15_15
+│   |   |_ jpg_files
+│   |   |_config.json
+│   |   |_config_gdf.json
+│   |   |_<rest of files>
+```
+
+## Session Files
+
+Each session contains the following files that contain information related to the shorelines extracted from each ROI. Below is an explanation of each file:
+
+## Configuration Files
+
+- **config.json**: Contains the settings that were used to download the ROI, the location of the downloaded ROI, and the settings that were used to extract shorelines.
+- **config_gdf.json**: Contains all the features that were on the map when shorelines were extracted this includes:
+  - Selected ROIs
+  - Reference Shoreline
+  - Transects
+  - Bounding Box (if present)
+  - Shoreline Extraction Area (if present)
+
+## Extracted Shorelines
+
+- **extracted_shorelines_lines.geojson**: A GeoJSON file containing the extracted shorelines represented as lines.
+- **extracted_shorelines_points.geojson**: A GeoJSON file containing the extracted shorelines represented as points.
+
+## Raw Transect Time Series
+
+- **raw_transect_time_series.csv**: Contains raw transect time series data in CSV format in the projected CRS (listed as `output_epsg` in the `config.json` settings)
+  - This file contains the intersection between the shoreline and all the transects for each date. This intersection is the distance along the transect the shoreline was captured at.
+- **raw_transect_time_series_merged.csv**: Merged raw transect time series data in CSV format in the projected CRS (listed as `output_epsg` in the `config.json` settings)
+
[…]
