### 2024-06-11 46bed8b
#260 add doctree for sample session and new page what is in a session
