+++ docs/sessions.md
@@ -1,9 +1,9 @@
-# What is a session?
+# What is a Session?
 
 A session is a saved state that CoastSeg was in. A session can be used to load in the ROIs that were downloaded previously or the previously extracted shorelines.
 A session is typically stored as a directory containing a minimum of two files a `config.json` and a `config_gdf.geojson` file that lets you restore the state coastseg was in.
 
-### Where are sessions located?
+### Where are Sessions Located?
 
 Sessions are saved to `CoastSeg/sessions`. A session contains the all the files related to extract shorelines for a particular set of ROIs in the downloaded data located in `CoastSeg/data`. Each session follows a similar format depending on what actions were taken by the user. For instance, if tide correction was not applied during the session then none of the files with the tidal correction applied will be present.
 
@@ -40,73 +40,6 @@ Here is a sample session that had 2 ROIs where shorelines were successfully extr
 │   |   |_config_gdf.json
 │   |   |_<rest of files>
 ```
-
-## Session Files
-
-Each session contains the following files that contain information related to the shorelines extracted from each ROI. Below is an explanation of each file:
-
-## Configuration Files
-
-- **config.json**: Contains the settings that were used to download the ROI, the location of the downloaded ROI, and the settings that were used to extract shorelines.
-- **config_gdf.json**: Contains all the features that were on the map when shorelines were extracted this includes:
-  - Selected ROIs
-  - Reference Shoreline
-  - Transects
-  - Bounding Box (if present)
-  - Shoreline Extraction Area (if present)
-
-## Extracted Shorelines
-
-- **extracted_shorelines_lines.geojson**: A GeoJSON file containing the extracted shorelines represented as lines.
-- **extracted_shorelines_points.geojson**: A GeoJSON file containing the extracted shorelines represented as points.
-
-## Raw Transect Time Series
-
-- **raw_transect_time_series.csv**: Contains raw transect time series data in CSV format in the projected CRS (listed as `output_epsg` in the `config.json` settings)
-  - This file contains the intersection between the shoreline and all the transects for each date. This intersection is the distance along the transect the shoreline was captured at.
-- **raw_transect_time_series_merged.csv**: Merged raw transect time series data in CSV format in the projected CRS (listed as `output_epsg` in the `config.json` settings)
-
-  - This file contains the intersection between the shoreline and all the transects for each date along with the x and y point in crs epsg 4326 of where the shoreline intersected the transect, and the x and y coordinate in crs epsg 4326 of the end point of the transect.
-
-  ![raw_timeseries_merged.csv](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/39f6f725-481b-48f8-8be4-4829939107d8)
-
-- **raw_transect_time_series_points.geojson**: Raw transect time series data represented as points in GeoJSON format in crs epsg 4326.
-- **raw_transect_time_series_vectors.geojson**: Raw transect time series data represented as vectors in GeoJSON format in crs epsg 4326.
-
-## Tidally Corrected Transect Time Series
-
-These files will only be present if tidal correction was applied during the analysis:
-
-- **tidally_corrected_transect_time_series.csv**: Tidally corrected transect time series data in CSV format in the projected CRS (listed as `output_epsg` in the `config.json` settings)
-  - This file contains the intersection between the shoreline and all the transects for each date. This intersection is the distance along the transect the shoreline was captured at.
-- **tidally_corrected_transect_time_series_merged.csv**: Merged tidally corrected transect time series data in the projected CRS (listed as `output_epsg` in the `config.json` settings)
[…]
