+++ mkdocs.yml
@@ -71,6 +71,7 @@ nav:
       - How to Download Imagery: how-to-download.md
       - Download Settings Guide: download_settings.md
       - How to Upload Features: how-to-upload-features.md
+      - What is in a Session : what-is-in-a-session.md
       - How to Use Sessions: sessions.md
       - How to Use Sessions from Another User: how-to-use-a-session-another-user.md
       - How to Use Extract Shoreline Area: how-to-extract-shoreline-area.md
