+++ tests/test_imports.py
@@ -1,7 +1,7 @@
 import os
 import platform
 from coastseg.file_utilities import load_package_resource
-import importlib
+import pytest
 
 
 def test_import_load_package_resource():
@@ -24,9 +24,10 @@ def test_import_bounding_boxes():
     except ImportError:
         assert False, "Failed to import bounding_boxes"
 
-
-def test_import_coastseg_logs():
+@pytest.mark.parametrize('named_temp_dir', ['CoastSeg'], indirect=True)
+def test_import_coastseg_logs(named_temp_dir):
     try:
+        # The named_temp_dir fixture created a temporary directory named 'CoastSeg'
         from coastseg import coastseg_logs
     except ImportError:
         assert False, "Failed to import coastseg_logs"
@@ -53,8 +54,11 @@ def test_import_downloads():
         assert False, "Failed to import downloads"
 
 
-def test_import_download_tide_model():
+# Use the `pytest.mark.parametrize` to pass the parameter to the fixture
+@pytest.mark.parametrize('named_temp_dir', ['CoastSeg'], indirect=True)
+def test_import_download_tide_model(named_temp_dir):
     try:
+        # The named_temp_dir fixture created a temporary directory named 'CoastSeg'
         from coastseg import download_tide_model
     except ImportError:
         assert False, "Failed to import download_tide_model"
