+++ tests/conftest.py
@@ -17,6 +17,34 @@ from tempfile import TemporaryDirectory
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
 
+# create a custom context manager to create a temporary directory & clean it up after use
+class NamedTemporaryDirectory:
+    def __init__(self, name):
+        self.name = name
+        self.path = os.path.join(tempfile.gettempdir(), name)
+        os.makedirs(self.path, exist_ok=True)
+    
+    def __enter__(self):
+        return self.path
+    
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        for root, dirs, files in os.walk(self.path, topdown=False):
+            for name in files:
+                os.remove(os.path.join(root, name))
+            for name in dirs:
+                os.rmdir(os.path.join(root, name))
+        os.rmdir(self.path)
+        
+@pytest.fixture
+def named_temp_dir(request):
+    # Retrieve the parameter from the request (this would be the name of the temporary directory)
+    dir_name = request.param
+    # Setup phase: create the temporary directory with the provided name
+    with NamedTemporaryDirectory(dir_name) as temp_dir:
+        yield temp_dir  # Provide the directory to the test function
+    # Teardown phase: cleanup is handled by the NamedTemporaryDirectory context manager
+
+
 @pytest.fixture(scope="session")
 def box_no_shorelines_transects():
     geojson = {
