### 2024-06-11 d11b904
fix broken tests use pathlib to create temp coastseg directory along cwd
