+++ tests/test_imports.py
@@ -2,7 +2,7 @@ import os
 import platform
 from coastseg.file_utilities import load_package_resource
 import pytest
-
+import pathlib
 
 def test_import_load_package_resource():
     try:
@@ -24,7 +24,7 @@ def test_import_bounding_boxes():
     except ImportError:
         assert False, "Failed to import bounding_boxes"
 
-@pytest.mark.parametrize('named_temp_dir', ['CoastSeg'], indirect=True)
+@pytest.mark.parametrize('named_temp_dir', [('CoastSeg',str(pathlib.Path(__file__).parent))], indirect=True)
 def test_import_coastseg_logs(named_temp_dir):
     try:
         # The named_temp_dir fixture created a temporary directory named 'CoastSeg'
@@ -55,7 +55,7 @@ def test_import_downloads():
 
 
 # Use the `pytest.mark.parametrize` to pass the parameter to the fixture
-@pytest.mark.parametrize('named_temp_dir', ['CoastSeg'], indirect=True)
+@pytest.mark.parametrize('named_temp_dir', [('CoastSeg',str(pathlib.Path(__file__).parent))], indirect=True)
 def test_import_download_tide_model(named_temp_dir):
     try:
         # The named_temp_dir fixture created a temporary directory named 'CoastSeg'
