### 2024-06-11 0f6fdfd
update core_utilities get_base_dir to work on windows, check for the CoastSeg directory on both the file and cwd paths and create the CoastSeg directory if not
