### 2024-06-13 SatelliteShorelines/CoastSeg#257 @dbuscombe-usgs
Hi @cheginit I have looked into the license issue and believe I have a definite answer.

To recap, the dataset in question is https://pubs.usgs.gov/publication/70202401. This dataset is public domain, since it originates from the U.S. Geological Survey (USGS), a US government agency whose works are not subject to copyright. It technically is a USGS data product because the first author is a USGS employee https://www.sciencebase.gov/catalog/catalogParty/show?partyId=9062. All works by US government employees are public domain.

The reviewer said that the dataset has a "Creative Commons Attribution-NonCommercial-NoDerivatives License" and they link to  https://creativecommons.org/licenses/by-nc-nd/4.0/

However, the reviewer is mistaken. He refers to the license of the paper, not the dataset. The license language reads:

"""
This is an Open Access article distributed under the terms of the Creative Commons Attribution-NonCommercial-NoDerivatives License (http://creativecommons.org/licenses/by-nc-nd/4.0/), which permits non-commercial re-use, distribution, and reproduction in any medium, provided the original work is properly cited, and is not altered, transformed, or built upon in any way.
"""

This comes from the journal paper’s license:
https://www.tandfonline.com/action/showCopyRight?scroll=top&doi=10.1080%2F1755876X.2018.1529714

Note that the license specifically refers to the article that describes the dataset, not the dataset itself.  Note the license solely refers to the ‘Open access article’. I believe this settles the matter. Otherwise, I'd be happy to get the U.S. Geological Survey legal team involved to clarify the situation over licensing of datasets that originate from the US government.
