### 2024-06-13 SatelliteShorelines/CoastSeg#257 @cheginit
@dbuscombe-usgs Thanks for looking into this and resolving the matter by providing valid points, appreciate it.

@2320sharon, @FlorisCalkoen Please consider the license issue resolved.
