### 2024-06-18 0f991cc
update the paper to list the coastsat_package improvements
