### 2024-06-18 c509a4b
#257 update the paper pdf with coastsat_package improvements
