### 2024-06-18 SatelliteShorelines/CoastSeg#257 @2320sharon
@FlorisCalkoen is it alright if I close this issue?
