### 2024-06-18 c17d7a9
#260 update the extract shorelines guide, getting started guide and what is in a session guide to explain where the extracted shoreline figures are generated, how downloaded data and sessions are kept track of, and add new figures to show contents of downloaded data and extracted shorelines folders
