+++ docs/sessions.md
@@ -1,7 +1,7 @@
 # What is a Session?
 
 A session is a saved state that CoastSeg was in. A session can be used to load in the ROIs that were downloaded previously or the previously extracted shorelines.
-A session is typically stored as a directory containing a minimum of two files a `config.json` and a `config_gdf.geojson` file that lets you restore the state coastseg was in.
+A session is typically stored as a directory containing a minimum of two files a `config.json` and a `config_gdf.geojson` file that lets you restore the state coastseg was in when the session was created.
 
 ### Where are Sessions Located?
 
@@ -70,99 +70,3 @@ To load config files for ROIs that have been downloaded before go to the `data` 
 2. Select a directory from the sessions directory
    ![load_extracted_shorelines_session](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/4d3ba936-3d86-48e1-9d2d-ea67a355c71d)
 
-## Config.json
-
-Config.json files can be found in session directories as well as within ROI directories within the data directory.
-
-### Sample Config.json in the data directory
-
-The config files are organized into 2 sections:
-
-1. Settings used to download each ROI organized by each ROI's id
-2. Settings that were saved when the ROI was downloaded under the 'settings' section
-
-a. Settings used to download each ROI organized by each ROI's id
-
-- One setting to pay attention to is the `output epsg` this is the CRS the ROI and all its features are re-projected features into and these re-projected features are the ones used to calculate the shoreline transect intersections.
-
-- The following settings are associated with ROI 0:
-  - ROI 0 has imagery for dates "2018-01-01" - "2019-01-01" and it was saved to the `ID_0_datetime03-22-23__07_29_15` directory at
-    `C:\\1_CoastSeg\\1_official_CoastSeg_repo\\CoastSeg\\data`.
-
-```
-{
-    "0": {
-        "dates": [
-            "2018-01-01",
-            "2019-01-01"
-        ],
-        "sitename": "ID_0_datetime03-22-23__07_29_15",
-        "polygon": [
-            [
-                [
-                    -75.75713505156588,
-                    36.189897585260795
-                ],
-                [
-                    -75.7514757042764,
-                    36.17420897122136
-                ],
-                [
-                    -75.74012332310643,
-                    36.176873810012815
-                ],
-                [
-                    -75.74578267039591,
-                    36.191909415885604
-                ],
-                [
-                    -75.75713505156588,
-                    36.189897585260795
-                ]
-            ]
-        ],
-        "roi_id": "0",
-        "sat_list": [
-            "L5",
-            "L7",
-            "L8"
-        ],
-        "landsat_collection": "C02",
-        "filepath": "C:\\1_CoastSeg\\1_official_CoastSeg_repo\\CoastSeg\\data"
-    },
-    "roi_ids": [
-        "0"
-    ],
-    "settings": {
-        "landsat_collection": "C02",
-        "dates": [
-            "2018-01-01",
-            "2019-03-01"
-        ],
-        "sat_list": [
-            "L5",
-            "L7",
-            "L8"
-        ],
-        "cloud_thresh": 0.68,
-        "dist_clouds": 100,
-        "output_epsg": 32618,
-        "check_detection": false,
-        "adjust_detection": false,
-        "save_figure": true,
-        "min_beach_area": 2100,
-        "min_length_sl": 200,
-        "cloud_mask_issue": false,
-        "sand_color": "default",
-        "pan_off": "False",
-        "max_dist_ref": 396,
-        "along_dist": 25,
-        "min_points": 3,
-        "max_std": 15.0,
-        "max_range": 30.0,
-        "min_chainage": -100.0,
-        "multiple_inter": "auto",
-        "prc_multiple": 0.1
-    }
-}
-```
