### 2024-06-18 5181aeb
#260 add doctree to how to filter out bad imagery and link guide in case study 2. Explain with more detail how downloaded data is loaded to extract shorelines in the getting started guide and extract shorelines guide
