+++ docs/extract_shorelines.md
@@ -85,9 +85,9 @@ If most of your imagery looks like these examples you may need to change your RO
 
 ## Step 4: Examine Shorelines Extracted from 3-5 years of imagery
 
-- When you click extract shorelines a new directory with the session name you provided will be created at `CoastSeg/sessions`. It will contain a subdirectory for each ROI you selected and within each subdirectory will be all the files generated by extracting shorelines which are detailed in [Extracted Shoreline Session Contents](#extracted-shoreline-session-contents). 
+- When you click "Extract Shorelines," a new directory with the session name you provided will be created at `CoastSeg/sessions`. This directory will contain a subdirectory for each ROI you selected. Within each subdirectory, you will find all the files generated by extracting shorelines, which are detailed in [Extracted Shoreline Session Contents](#extracted-shoreline-session-contents).
 
-- You can find images that show the extracted shoreline on each image in `coastseg/sessions/<YOUR SESSION NAME>/ROI_ID_NAME/jpg_files/detection` as shown below:
+- You can find images that show the extracted shorelines on each image in `CoastSeg/sessions/<YOUR SESSION NAME>/ROI_ID_NAME/jpg_files/detection`. An example is shown below:
 
 ![coastseg/sessions/session/roi/jpg_files/detection](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/f5be75fb-1c75-46c2-bbec-b9b1e716fea4)
 
