+++ docs/How-to-Filter-Out-Bad-Imagery.md
@@ -1,37 +1,65 @@
 # Filtering Imagery
 
-Extracting the best shorelines requires that only the best imagery is used which means that bad imagery needs to be sorted out. You should filter out bad imagery in `data/roi_id/jpg_files/preprocessed/RGB` by moving any bad imagery to a designated subdirectory named 'bad'.
 
-It is important to note that this operation does not delete any TIFF files, but rather helps to improve the efficiency of shoreline extraction and enhance the quality of the extracted shorelines. Check out the [wiki page about the shoreline extraction process](https://github.com/Doodleverse/CoastSeg/wiki/6.-How-to-Extract-Shorelines) for more information about how to extract shorelines.
+To extract the best shorelines, it's essential to use only high-quality imagery, which means filtering out any poor-quality images. You should filter out bad imagery in `data/roi_id/jpg_files/preprocessed/RGB` by moving any undesirable images to a designated subdirectory named 'bad'.
 
-## Step by Step Guide
+Filtering the RGB images does not delete any TIFF files but helps to improve the efficiency of shoreline extraction and enhance the quality of the extracted shorelines. For more information on the shoreline extraction process, check out the [wiki page about the shoreline extraction process](https://github.com/Doodleverse/CoastSeg/wiki/6.-How-to-Extract-Shorelines).
 
-### Before you begin
 
-Download data with the `SDS_coastsat_classifier.ipynb` and make sure it downloaded to your `data` directory within coastseg
+## Step-by-Step Guide
 
-1. Open the `data` directory within coastseg
-2. Navigate to the ROI directory
+### Before You Begin
 
-- If you're not sure which ROI directory you used for a particular session go to the `sessions` directory, click the session you are interested in and open the `config.json` file. Look for the variable `sitename` and the sitename will be in a format similar to `ID_yvk1_datetime06-05-23__07_07_42`. This is the name of the ROI directory containing the downloaded data in the `data` directory within coastseg
+Download data with the `SDS_coastsat_classifier.ipynb` and ensure it is saved to your `data` directory within CoastSeg.
 
-### Example config file from data
+### 1. Open the `data` Directory
 
-![config_json_example_screenshot_sitename_from_data](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/cc172608-895d-456b-8789-0725fb338805)
+1. Navigate to the `data` directory within CoastSeg.
+2. Click on the ROI directory you want to sort
 
-### Example config file from sessions
+In this example, you can see that ROI 3 is located within the `CoastSeg/data` folder. The `jpg_files > preprocessed > RGB` directory contains all the RGB imagery for all the satellites.
 
-![config_json_sitename_from_session](https://github.com/SatelliteShorelines/CoastSeg/assets/61564689/45d2f48e-abe9-4dd7-bcd0-8253092c6613)
+```
+├── CoastSeg
+|
+|___data
+|    |_ ID_3_datetime11-22-22__11_15_15
+│   |   |_ L8
+│   |   |_ L9
+│   |   |_ S2
+│   |   |_ jpg_files
+│   |   |  |_ preprocessed
+|   │   |   |  |_ RGB
+|   |   │   |  |  |_ 2018-12-06-19-04-16_RGB_S2.jpg
+|   |   │   |  |  |_ 2018-12-06-19-04-16_RGB_L8.jpg
+|   |   │   |  |  |_ 2018-12-06-19-04-16_RGB_L9.jpg
+|   │   |   |  |_ NIR
+|   |   │   |  |  |_ 2018-12-06-19-04-16_NIR_S2.jpg
[…]
