### 2024-06-18 eb37423
#260 Add link to the what is in a session guide to getting started guide
