### 2024-06-20 SatelliteShorelines/CoastSeg#260 @2320sharon
@fmemuir

I have implemented your feedback and made several enhancements to the documentation as detailed below:

- Screenshot Addition to Getting Started Guide: I've included a screenshot in the [getting started guide](https://satelliteshorelines.github.io/CoastSeg/getting-started/#:~:text=9.-,Download%20the%20ROIs,-Click%20the%20ROIs) to help users locate the RGB imagery for sorting.

- I updated the Getting Started Guide to clarify the difference between the ROIs in `CoastSeg/data` and `CoastSeg/sessions` and added a screenshot to show users where to find the shoreline detection images [here](https://satelliteshorelines.github.io/CoastSeg/getting-started/#:~:text=Extracting%20shorelines%20involves,this%20guide.)

- I updated the [How to Filter Bad Imagery](https://satelliteshorelines.github.io/CoastSeg/How-to-Filter-Out-Bad-Imagery/) and link to this guide in [case_study_2](https://satelliteshorelines.github.io/CoastSeg/case_study_2/)

- I updated the [How to Use Sessions](https://satelliteshorelines.github.io/CoastSeg/sessions/) guide to tell the user where the sessions are saved as well as how each session is related to ROIs in CoastSeg/data

- I also added a new page to the website called [What is in a Session](https://satelliteshorelines.github.io/CoastSeg/what-is-in-a-session/) that breaks down each kind of file in a session and  includes a doctree of the session contents.

Thank you for your feedback and I hope this will make sessions easier to understand. Please let me know if these changes work or if there needs to be any additional clarification.

Thank you again for reviewing CoastSeg and for your helpful feedback.
