### 2024-06-21 3a414af
update the extracted shorelines to extract individual contours
