+++ src/coastseg/extracted_shoreline.py
@@ -2,6 +2,7 @@
 import colorsys
 import copy
 import shutil
+import pytz
 import fnmatch
 import json
 import json
@@ -55,7 +56,7 @@ from coastsat.SDS_tools import (
     remove_duplicates,
     remove_inaccurate_georef,
 )
-from coastsat.SDS_transects import compute_intersection_QC
+# from coastsat.SDS_transects import compute_intersection_QC
 from coastsat import SDS_preprocess, SDS_shoreline, SDS_tools
 from ipyleaflet import GeoJSON
 from matplotlib import gridspec
@@ -70,6 +71,9 @@ from coastseg.validation import get_satellites_in_directory
 from coastseg.filters import filter_model_outputs, apply_land_mask
 from coastseg.common import get_filtered_files_dict, edit_metadata
 
+from scipy.spatial import KDTree
+from shapely.geometry import LineString
+from skimage import transform
 
 # Set pandas option
 pd.set_option("mode.chained_assignment", None)
@@ -78,6 +82,138 @@ pd.set_option("mode.chained_assignment", None)
 logger = logging.getLogger(__name__)
 __all__ = ["Extracted_Shoreline"]
 
+
+def compute_intersection_QC(
+    output,
+    transects,
+    along_dist=25,
+    min_points=3,
+    max_std=15,
+    max_range=30,
+    min_chainage=-100,
+    multiple_inter="auto",
+    prc_multiple=0.1,
+    use_progress_bar: bool = True,
+):
+    """
+    More advanced function to compute the intersection between the 2D mapped shorelines
+    and the transects. Produces more quality-controlled time-series of shoreline change.
+
+    Arguments:
+    -----------
+        output: dict
+            contains the extracted shorelines and corresponding dates.
+        transects: dict
+            contains the X and Y coordinates of the transects (first and last point needed for each
+            transect).
+        along_dist: int (in metres)
+            alongshore distance to calculate the intersection (median of points
+            within this distance).
+        min_points: int
+            minimum number of shoreline points to calculate an intersection.
+        max_std: int (in metres)
+            maximum std for the shoreline points when calculating the median,
+            if above this value then NaN is returned for the intersection.
+        max_range: int (in metres)
+            maximum range for the shoreline points when calculating the median,
+            if above this value then NaN is returned for the intersection.
+        min_chainage: int (in metres)
+            furthest landward of the transect origin that an intersection is
+            accepted, beyond this point a NaN is returned.
+        multiple_inter: mode for removing outliers ('auto', 'nan', 'max').
+        prc_multiple: float, optional
+            percentage to use in 'auto' mode to switch from 'nan' to 'max'.
+        use_progress_bar(bool,optional). Defaults to True. If true uses tqdm to display the progress for iterating through transects.
+            False, means no progress bar is displayed.
+
+    Returns:
+    -----------
+        cross_dist: dict
+            time-series of cross-shore distance along each of the transects. These are not tidally
+            corrected.
+    """
+
+    cross_dist = {}
+
+    shorelines = output["shorelines"]
+    transect_keys = list(transects.keys())
+    if use_progress_bar:
+        transect_keys = tqdm(
+            transect_keys, desc="Computing transect shoreline intersections"
+        )
+
+    for key in transect_keys:
+        std_intersect = np.full(len(shorelines), np.nan)
+        med_intersect = np.full(len(shorelines), np.nan)
[…]
