### 2024-06-24 SatelliteShorelines/CoastSeg#261 @fmemuir
Thanks @2320sharon for your thorough response and efforts to address this suggestion, I believe it will really improve the overall user experience! Happy to close this whenever you are.
