### 2024-06-24 SatelliteShorelines/CoastSeg#260 @fmemuir
@2320sharon Great updates, really appreciate the added clarification!
