### 2024-06-24 b3c5fa4
#263 add smoothening function to zoo extract shorelines workflow
