+++ src/coastseg/extracted_shoreline.py
@@ -1,6 +1,7 @@
 # Standard library imports
 import colorsys
 import copy
+import shapely
 import shutil
 import fnmatch
 import json
@@ -55,7 +56,7 @@ from coastsat.SDS_tools import (
     remove_duplicates,
     remove_inaccurate_georef,
 )
-from coastsat.SDS_transects import compute_intersection_QC
+# from coastsat.SDS_transects import compute_intersection_QC
 from coastsat import SDS_preprocess, SDS_shoreline, SDS_tools
 from ipyleaflet import GeoJSON
 from matplotlib import gridspec
@@ -70,6 +71,10 @@ from coastseg.validation import get_satellites_in_directory
 from coastseg.filters import filter_model_outputs, apply_land_mask
 from coastseg.common import get_filtered_files_dict, edit_metadata
 
+from scipy.spatial import KDTree
+from shapely.geometry import LineString
+from skimage import transform
+import pytz
 
 # Set pandas option
 pd.set_option("mode.chained_assignment", None)
@@ -90,6 +95,273 @@ def time_func(func):
 
     return wrapper
 
+def filter_shoreline_new(
+    shoreline,
+    shoreline_extraction_area,
+    output_epsg,
+):
+    """Filter the shoreline based on the extraction area.
+    Args:
+        shoreline (array): The original shoreline data.
+        shoreline_extraction_area (GeoDataFrame): The area to extract the shoreline from.
+        shoreline_extraction_area (GeoDataFrame): The area to extract the shoreline from.
+    Returns:
+        np.array: The filtered shoreline as a numpy array of shape (n,2).
+    """
+    if shoreline_extraction_area is not None:
+        # Ensure both the shoreline and extraction area are in the same CRS.
+        shoreline_extraction_area_gdf = shoreline_extraction_area.to_crs(
+            f"epsg:{output_epsg}"
+        )
+
+        if isinstance(shoreline, gpd.GeoDataFrame):
+            shoreline_gdf = shoreline.to_crs(f"epsg:{output_epsg}")
+        else:
+            # Convert the shoreline to a GeoDataFrame.
+            shoreline_gdf = SDS_shoreline.create_gdf_from_type(
+                shoreline,
+                "lines",
+                crs=f"epsg:{output_epsg}",
+            )
+            if shoreline_gdf is None:
+                return shoreline
+
+        # Filter shorelines within the extraction area.
+        filtered_shoreline_gdf = common.ref_poly_filter(
+            shoreline_extraction_area_gdf, shoreline_gdf
+        )
+        return filtered_shoreline_gdf
+
+    return shoreline
+
+def compute_intersection_QC(
+    output,
+    transects,
+    along_dist=25,
+    min_points=3,
+    max_std=15,
+    max_range=30,
+    min_chainage=-100,
+    multiple_inter="auto",
+    prc_multiple=0.1,
+    use_progress_bar: bool = True,
+    **kwargs,
+):
+    """
+    More advanced function to compute the intersection between the 2D mapped shorelines
+    and the transects. Produces more quality-controlled time-series of shoreline change.
+    Arguments:
+    -----------
+        output: dict
+            contains the extracted shorelines and corresponding dates.
+        transects: dict
+            contains the X and Y coordinates of the transects (first and last point needed for each
+            transect).
+        along_dist: int (in metres)
+            alongshore distance to calculate the intersection (median of points
+            within this distance).
+        min_points: int
+            minimum number of shoreline points to calculate an intersection.
+        max_std: int (in metres)
+            maximum std for the shoreline points when calculating the median,
[…]
