### 2024-06-24 c7a83f8
update filters.py to save the kmeans classifer scores to model_settings.json
