### 2024-06-25 e150f0c
#261 make independent linestrings work in the zoo workflow
