+++ src/coastseg/extracted_shoreline.py
@@ -95,6 +95,30 @@ def time_func(func):
 
     return wrapper
 
+def stringify_datetime_columns(gdf: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+    """
+    Check if any of the columns in a GeoDataFrame have the type pandas timestamp and convert them to string.
+
+    Args:
+        gdf: A GeoDataFrame.
+
+    Returns:
+        A new GeoDataFrame with the same data as the original, but with any timestamp columns converted to string.
+    """
+    timestamp_cols = [
+        col for col in gdf.columns if pd.api.types.is_datetime64_any_dtype(gdf[col])
+    ]
+
+    if not timestamp_cols:
+        return gdf
+
+    gdf = gdf.copy()
+
+    for col in timestamp_cols:
+        gdf[col] = gdf[col].astype(str)
+
+    return gdf
+
 def filter_shoreline_new(
     shoreline,
     shoreline_extraction_area,
@@ -564,12 +588,6 @@ def combine_satellite_data(satellite_data: dict) -> dict:
         satellite_data[satname].setdefault("cloud_cover", [])
         satellite_data[satname].setdefault("filename", [])
         satellite_data[satname].setdefault("idx", [])
-        satellite_data[satname].setdefault("dates", [])
-        satellite_data[satname].setdefault("geoaccuracy", [])
-        satellite_data[satname].setdefault("shorelines", [])
-        satellite_data[satname].setdefault("cloud_cover", [])
-        satellite_data[satname].setdefault("filename", [])
-        satellite_data[satname].setdefault("idx", [])
         # For each key in the nested dictionary
         for key, value in sat_data.items():
             # Wrap non-list values in a list and concatenate to merged_satellite_data
@@ -655,12 +673,11 @@ def smooth_lines(lines:gpd.GeoDataFrame,refinements=5):
         coords = LineString_to_arr(line.geometry)
         refined = chaikins_corner_cutting(coords, refinements=refinements)
         refined_geom = arr_to_LineString(refined)
-        # new_lines['geometry'][i] = refined_geom
         new_lines.loc[i,'geometry'] = refined_geom
     return new_lines
 
 def process_shoreline_zoo(
-    contours, cloud_mask, im_nodata, georef, image_epsg, settings, date, **kwargs
+    contours, cloud_mask, im_nodata, georef, image_epsg, settings, date,satname:str,**kwargs
 ):
     # convert the contours that are currently pixel coordinates to world coordiantes
     contours_world = SDS_tools.convert_pix2world(contours, georef)
@@ -706,6 +723,7 @@ def process_shoreline_zoo(
     gdf = gpd.GeoDataFrame(
         {
             "date": np.tile(date_obj, len(contours_shapely)), # type: ignore
+            "satname": np.tile(satname, len(contours_shapely)), # type: ignore
             "cloud_cover": np.tile(cloud_cover, len(contours_shapely)),
         },
         geometry=contours_shapely,
@@ -714,13 +732,6 @@ def process_shoreline_zoo(
 
     # smooth the shorelines in the GeoDataFrame
     gdf = smooth_lines(gdf)
-
-    # print(
-    #     os.path.abspath(f"shoreline_{date_obj.strftime('%Y-%m-%d-%H-%M-%S')}.geojson")
-    # )
-    # gdf.to_file(
-    #     f"shoreline_{date_obj.strftime('%Y-%m-%d-%H-%M-%S')}.geojson", driver="GeoJSON"
-    # )
     return gdf
 
 def find_shoreline(
@@ -734,6 +745,7 @@ def find_shoreline(
     im_labels: np.ndarray,
[…]
