### 2024-06-26 #6683 @cheginit
👋🏼 @2320sharon, please fix the issues with the DOIs as you can see in the previous comments and add the missing ones. I also noticed that there are two `CoastSat` references in the references; one is a reference to a tagged release on GitHub and the other to its paper. Please use the paper one for both instances. In general, you should try to avoid using links to GitHub in the refences as a software release since they are not permanent identifiers. So, if you can find a refence on archiving sites like Zenodo, please use those as reference and make sure that DOIs are included. The same goes for references to `Coastsat-package` on PyPi. There's also an issue in the "Pardo-Pascual, J. E ..." reference where the link has an additional "https": `https://doi.org/https:316//doi.org/10.1016/j.rse.2012.02.024`. Please fix all the issues in the bib file and check if all issues are resolved by using using our EditorBot and the command: `@editorialbot check references`
