### 2024-06-26 6ead75f
#261 comment out code in save_extracted_shoreline that saves the extracted shorelines to file
