### 2024-06-26 d8a0fd2
fix DOIs in paper.bib, remove CoastSeg github ref and remove coastsat GitHub reference
