+++ paper/paper.bib
@@ -28,8 +28,8 @@ journal = {Remote Sensing of Environment},
 volume = {123},
 pages = {1-11},
 year = {2012},
-doi = {https://doi.org/10.1016/j.rse.2012.02.024},
-url = {https://www.sciencedirect.com/science/article/pii/S0034425712001174},
+doi = {10.1016/j.rse.2012.02.024},
+url = {https://doi.org/10.1016/j.rse.2012.02.024},
 author = {Josep E. Pardo-Pascual and Jaime Almonacid-Caballer and Luis A. Ruiz and Jesús Palomar-Vázquez}
 }
 
@@ -58,16 +58,10 @@ eprint = {
   volume={439},
   pages={108850},
   year={2023},
+  doi = {10.1016/j.geomorph.2023.108850},
+  url = {https://doi.org/10.1016/j.geomorph.2023.108850,
   publisher={Elsevier}
 }
-
-@misc{CoastSeg,
-  title={{CoastSat v2.4}},
-  author={Fitzpatrick, S. and Buscombe,D. and Lundine, M. and Warrick,J. and Vos, K.},
-  year={2024},
-  publisher={GitHub},
-  howpublished = {\url{https://github.com/kvos/CoastSat/releases/tag/v1.2.5}}
-  }
 
 @article{vos2023benchmarking,
   title={Benchmarking satellite-derived shoreline mapping algorithms},
@@ -77,9 +71,10 @@ eprint = {
   number={1},
   pages={345},
   year={2023},
+  doi = {10.1038/s43247-023-01001-2},
+  url = {https://doi.org/10.1038/s43247-023-01001-2},
   publisher={Nature Publishing Group UK London}
 }
-
 
 @article{turner2021satellite,
   title={Satellite optical imagery in Coastal Engineering},
@@ -88,7 +83,9 @@ eprint = {
   volume={167},
   pages={103919},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.coastaleng.2021.103919},
+  url={https://doi.org/10.1016/j.coastaleng.2021.103919}
 }
 
 @article{vos2023pacific,
@@ -99,9 +96,10 @@ eprint = {
   number={2},
   pages={140--146},
   year={2023},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41561-022-01117-8},
+  url={https://doi.org/10.1038/s41561-022-01117-8}
 }
-
 
 @article{castelle2021satellite,
   title={Satellite-derived shoreline detection at a high-energy meso-macrotidal beach},
@@ -110,8 +108,11 @@ eprint = {
   volume={383},
   pages={107707},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.geomorph.2021.107707},
+  url={https://doi.org/10.1016/j.geomorph.2021.107707}
 }
+
 
 @article{warrick2023large,
   title={A Large Sediment Accretion Wave Along a {N}orthern {C}alifornia Littoral Cell},
@@ -119,8 +120,11 @@ eprint = {
   journal={Journal of Geophysical Research: Earth Surface},
   pages={e2023JF007135},
   year={2023},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2023JF007135},
+  url={https://doi.org/10.1029/2023JF007135}
 }
+
 
 @article{vitousek2023future,
   title={The future of coastal monitoring through satellite remote sensing},
@@ -129,7 +133,9 @@ eprint = {
   volume={1},
   pages={e10},
   year={2023},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
[…]
