+++ paper/paper.bib
@@ -51,6 +51,7 @@ eprint = {
 }
 }
 
+
 @article{mclean202350,
   title={50 Years of Beach--Foredune Change on the Southeastern Coast of Australia: Bengello Beach, Moruya, NSW, 1972--2022},
   author={McLean, R. and Thom, B. and Shen, J. and Oliver, T.},
@@ -59,9 +60,10 @@ eprint = {
   pages={108850},
   year={2023},
   doi = {10.1016/j.geomorph.2023.108850},
-  url = {https://doi.org/10.1016/j.geomorph.2023.108850,
+  url = {https://doi.org/10.1016/j.geomorph.2023.108850},
   publisher={Elsevier}
 }
+
 
 @article{vos2023benchmarking,
   title={Benchmarking satellite-derived shoreline mapping algorithms},
