### 2024-06-26 53d821f
fix remaining references missing DOIs
