+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{vos2019coastsat,
-  title={{CoastSat: A Google Earth Engine-enabled Python toolkit to extract shorelines from publicly available satellite imagery}},
+  title={CoastSat: A Google Earth Engine-enabled Python toolkit to extract shorelines from publicly available satellite imagery},
   author={Vos, K. and Splinter, K. D. and Harley, M. D. and Simmons, J. A. and Turner, I. L.},
   journal={Environmental Modelling \& Software},
   volume={122},
@@ -176,15 +176,14 @@ eprint = {
   url={https://doi.org/10.1016/j.geomorph.2022.108360}
 }
 
-
-@misc{alley2017pytmd,
-  title={py{TMD}: {P}ython based tidal prediction software},
+@misc{tyler_sutterley_2024,
+  author       = {Tyler Sutterley},
+  title        = {tsutterley/pyTMD: v2.1.1},
+  year         = 2024,
   publisher    = {Zenodo},
-  howpublished = {\url{https://doi.org10.5281/zenodo.5555395}},
-  author={Alley, K. and Brunt, K. and Howard, S. and Padman, L. and Siegfried, M. and Sutterly, T.},
-  year={2017},
+  doi          = {10.5281/zenodo.10929240},
+  url          = {https://doi.org/10.5281/zenodo.10929240},
 }
-
 
 @article{lyard2021fes2014,
   title={{FES2014} global ocean tide atlas: {D}esign and performance},
@@ -224,12 +223,13 @@ eprint = {
   url={https://doi.org/10.3390/rs9070676}
 }
 
-@Misc{planet,
+
+@misc{planet,
   author =    {{Planet Labs}},
   organization = {Planet Labs},
   title =     {{Planet Application Program Interface: In Space for Life on Earth}},
   year =      {2018},
-  url = "https://api.planet.com"
+  url = {https://www.planet.com/}
 }
 
 @dataset{buscombe_2023_8187949,
@@ -258,21 +258,26 @@ eprint = {
   url          = {\url{https://doi.org/10.5281/zenodo.7786276}}
 }
 
+
 @misc{voscoastsat,
   title={{C}oastsat-package},
   author={Vos, K. and Fitzpatrick, S.},
   publisher={PyPi},
-  journal = {PyPi},
   year={2023},
-  howpublished = {\url{https://pypi.org/project/coastsat-package/}},
+  doi={10.5281/zenodo.12553179},
+  url={https://doi.org/10.5281/zenodo.12553179}
 }
 
 @misc{krause2021dea,
-  title={{DEA Notebooks contributors 2021: Digital Earth Australia notebooks and tools repository, Geoscience Australia, Canberra}},
-  author={Krause, C. and Dunn, B. and Bishop-Taylor, R. and Adams, C. and Burton, C. and Alger, M. and Chua, S. and Phillips, C. and Newey, V. and Kouzoubov, K. and others},
+  title={{Digital Earth Australia notebooks and tools repository}},
+  author={Krause, C. and Dunn, B. and Bishop-Taylor, R. and Adams, C. and Burton, C. and Alger, M. and Chua, S. and Phillips, C. and Newey, V. and Kouzoubov, K. and Leith, A. and Ayers, D. and Hicks, A.},
   year={2021},
-  howpublished = {\url{https://knowledge.dea.ga.gov.au/notebooks/README/}},
+  publisher={Commonwealth of Australia (Geoscience Australia)},
+  url={https://doi.org/10.26186/145234},
+  howpublished={\url{https://github.com/GeoscienceAustralia/dea-notebooks/}},
+  doi={10.26186/145234}
 }
+
 
 @article{konstantinou2023satellite,
   title={Satellite-based shoreline detection along high-energy macrotidal coasts and influence of beach state},
