### 2024-06-26 3c336db
fix core_utilities so that the "CoastSeg" directory is not required to run CoastSeg. CWD will be used if CoastSeg does not exist
