### 2024-06-26 5921c75
v1.2.8 update core_utilities.py to not require "CoastSeg" in working directory
