### 2024-06-26 cf0d4a5
update transects.py and common.py to fix the broken tests
