### 2024-06-26 0e9eab8
remove references without DOIs and replace coastseg reference with DOI version
