+++ paper/paper.bib
@@ -224,14 +224,6 @@ eprint = {
 }
 
 
-@misc{planet,
-  author =    {{Planet Labs}},
-  organization = {Planet Labs},
-  title =     {{Planet Application Program Interface: In Space for Life on Earth}},
-  year =      {2018},
-  url = {https://www.planet.com/}
-}
-
 @dataset{buscombe_2023_8187949,
   author       = {Buscombe, D. and
                   Fitzpatrick, S.},
@@ -266,6 +258,16 @@ eprint = {
   year={2023},
   doi={10.5281/zenodo.12553179},
   url={https://doi.org/10.5281/zenodo.12553179}
+}
+
+@misc{CoastSeg,
+  author       = {Fitzpatrick, S. and Buscombe,D. and Lundine, M. and Warrick,J. and Vos, K.},
+  title        = {SatelliteShorelines/CoastSeg: v1.2.9},
+  year         = 2024,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.12555413},
+  url          = {https://doi.org/10.5281/zenodo.12555413},
+  note         = {\url{https://github.com/SatelliteShorelines/CoastSeg/tree/v1.2.9}}
 }
 
 @misc{krause2021dea,
@@ -325,21 +327,6 @@ eprint = {
   publisher={Elsevier},
   doi={10.1016/j.rse.2021.112734},
   url={https://doi.org/10.1016/j.rse.2021.112734}
-}
-
-@misc{modelzoo,
-  title={{PyTorch Model Zoo}},
-  author={PyTorch},
-  year={2020},
-  howpublished = {\url{https://pytorch.org/serve/model_zoo.html}},
-}
-
-@misc{modelzoo2,
-  title={{NVIDIA Model Zoo}},
-  author={NVIDIA},
-  year={2023},
-  howpublished = {\url{https://docs.nvidia.com/tao/tao-toolkit/text/model_zoo/overview.html
-}},
 }
 
 @article{garcia2015evaluating,
