### 2024-06-26 767acd1
add last missing DOI for coastsat
