+++ paper/paper.bib
@@ -6,6 +6,8 @@
   volume={122},
   pages={104528},
   year={2019},
+  doi={10.1016/j.envsoft.2019.104528},
+  url={https://doi.org/10.1016/j.envsoft.2019.104528},
   publisher={Elsevier}
 }
 
