### 2024-06-26 #6683 @2320sharon
Hi @cheginit,

Thank you for looking over the references. I have fixed all the references as requested. The duplicate CoastSat entries have been removed, and there is now a single reference to the paper. I also replaced the GitHub links with Zenodo generated DOIs where possible. The issue with the "Pardo-Pascual, J. E ..." reference has also been corrected. I have checked the references using the EditorBot command @editorialbot check references, and all issues are now resolved.

Thank you!
