### 2024-06-27 2a09573
#261 add code in coastsat workflow to make shoreline contours independent
