+++ src/coastseg/extracted_shoreline.py
@@ -12,7 +12,7 @@ import os
 import datetime
 from glob import glob
 from time import perf_counter
-from typing import Optional, Union, List, Dict
+from typing import Any, Optional, Union, List, Dict
 from time import perf_counter
 from typing import Dict, List, Optional, Union
 from itertools import islice
@@ -48,7 +48,7 @@ import skimage.measure as measure
 import skimage.morphology as morphology
 
 from coastsat.SDS_download import get_metadata
-from coastsat.SDS_shoreline import extract_shorelines
+from coastsat import SDS_download
 from coastsat.SDS_tools import (
     get_filenames,
     get_filepath,
@@ -57,7 +57,6 @@ from coastsat.SDS_tools import (
     remove_inaccurate_georef,
 )
 # from coastsat.SDS_transects import compute_intersection_QC
-from coastsat import SDS_preprocess, SDS_shoreline, SDS_tools
 from ipyleaflet import GeoJSON
 from matplotlib import gridspec
 from matplotlib.colors import rgb2hex
@@ -73,7 +72,6 @@ from coastseg.common import get_filtered_files_dict, edit_metadata
 
 from scipy.spatial import KDTree
 from shapely.geometry import LineString
-from skimage import transform
 import pytz
 
 # Set pandas option
@@ -94,6 +92,522 @@ def time_func(func):
         return result
 
     return wrapper
+
+
+def save_shorelines_to_geojson(all_shorelines_gdf: gpd.GeoDataFrame, session_path: str) -> Dict[str, str]:
+    """
+    Processes and saves shoreline data to GeoJSON files.
+
+    Args:
+        all_shorelines_gdf (geopandas.GeoDataFrame): GeoDataFrame containing shoreline data.
+        session_path (str): The directory path where the GeoJSON files will be saved.
+
+    Returns:
+        Dict[str, str]: A dictionary with paths to the saved GeoJSON files.
+    """
+    all_shorelines_gdf = all_shorelines_gdf.reset_index(drop=True)
+    
+    # Convert to EPSG 4326
+    all_shorelines_gdf_4326 = all_shorelines_gdf.to_crs(epsg=4326)
+    # Drop the filename column
+    all_shorelines_gdf_4326.drop(columns=["filename"], inplace=True)
+
+    if all_shorelines_gdf_4326.empty:
+        print("No shorelines were extracted.")
+        logger.warning("No shorelines were extracted.")
+        return {}
+
+    # Save extracted shorelines to GeoJSON files
+    lines_geojson_path = os.path.join(session_path, 'extracted_shorelines_lines.geojson')
+    all_shorelines_gdf_4326.to_file(lines_geojson_path, driver="GeoJSON")
+    print(f"extracted_shorelines_lines.geojson saved to {lines_geojson_path}")
+
+    # Convert linestrings to multipoints
+    points_gdf = convert_linestrings_to_multipoints(all_shorelines_gdf_4326)
+    projected_gdf = stringify_datetime_columns(points_gdf)
+    
+    # Save extracted shorelines as a GeoJSON file
+    points_geojson_path = os.path.join(session_path, 'extracted_shorelines_points.geojson')
+    projected_gdf.to_file(points_geojson_path, driver="GeoJSON")
+    print(f"extracted_shorelines_points.geojson saved to {points_geojson_path}")
+
+    return {
+        "lines_geojson": lines_geojson_path,
+        "points_geojson": points_geojson_path
+    }
+
+def aggregate_shoreline_data(shoreline_geodataframe: gpd.GeoDataFrame) -> Dict[str, List[Any]]:
+    """
+    Aggregates shoreline data from a GeoDataFrame, extracting contours and appending relevant metadata.
+
+    Args:
[…]
