### 2024-06-27 #6683 @cheginit
@2320sharon Thanks for addressing the issue raised by the reviewers and me diligently. We're almost there! Please do the following, so I can then move forward with recommending acceptance of the submission:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Make sure that the license listed for the archive is the same as the software license!
