### 2024-06-27 #6683 @2320sharon
@cheginit Thank you for all your help in the journal submission process. I finished the tasks you listed above, so we should be ready for submission if these are all correct.

 - CoastSeg version 1.2.9 : https://github.com/SatelliteShorelines/CoastSeg/releases/tag/v1.2.9

- [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Make sure that the license listed for the archive is the same as the software license!

- The license in the Zenodo release is a GNU General Public License v3.0 license, which matches the software license in GitHub.

- I updated the list of authors each with their ORCID, updated the title of the Zenodo release to match the title of the paper and removed any small fix contributors : https://zenodo.org/records/12555413

DOI URL: https://doi.org/10.5281/zenodo.12555413
DOI: 10.5281/zenodo.12555413
