### 2024-06-27 #6683 @cheginit
@2320sharon I think the issue is with `{#sylt width="100%"}` in your `paper.md`. Please remove those, they're not needed.
