### 2024-06-27 6acbf24
remove {#sylt width="100%"} from paper.md
