### 2024-06-27 #6683 @2320sharon
@cheginit thank you for finding that. I removed it from paper.md
