### 2024-06-27 #6683 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
