### 2024-06-28 #6683 @kthyng
Paper comments:
* fix grammer 3rd paragraph pg. 1
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "australia"; please check for others.
