### 2024-06-28 34c1cd2
fix paper.md grammer mistakes and update paper.bib to capitalize references correctly
