+++ paper/paper.bib
@@ -1,6 +1,5 @@
-
 @article{vos2019coastsat,
-  title={CoastSat: A Google Earth Engine-enabled Python toolkit to extract shorelines from publicly available satellite imagery},
+  title={{CoastSat}: {A} {Google Earth Engine}-enabled {Python} toolkit to extract shorelines from publicly available satellite imagery},
   author={Vos, K. and Splinter, K. D. and Harley, M. D. and Simmons, J. A. and Turner, I. L.},
   journal={Environmental Modelling \& Software},
   volume={122},
@@ -12,7 +11,7 @@
 }
 
 @article{luijendijk2018state,
-  title={The State of the World's Beaches},
+  title={The {State} of the {World}'s {Beaches}},
   author={Luijendijk, Arjen and Hagenaars, Gerben and Ranasinghe, Roshanka and Baart, Fedor and Donchyts, Gennadii and Aarninkhof, Stefan},
   journal={Scientific Reports},
   volume={8},
@@ -25,7 +24,7 @@
 }
 
 @article{pardopascual20121,
-title = {Automatic extraction of shorelines from Landsat TM and ETM+ multi-temporal images with subpixel precision},
+title = {Automatic extraction of shorelines from {Landsat TM} and {ETM+} multi-temporal images with subpixel precision},
 journal = {Remote Sensing of Environment},
 volume = {123},
 pages = {1-11},
@@ -45,17 +44,12 @@ pages = {S47--S56},
 year = {2019},
 publisher = {Taylor \& Francis},
 doi = {10.1080/1755876X.2018.1529714},
-URL = {
-        https://doi.org/10.1080/1755876X.2018.1529714
-},
-eprint = {
-        https://doi.org/10.1080/1755876X.2018.1529714
+URL = {https://doi.org/10.1080/1755876X.2018.1529714},
+eprint = {https://doi.org/10.1080/1755876X.2018.1529714}
 }
-}
-
 
 @article{mclean202350,
-  title={50 Years of Beach--Foredune Change on the Southeastern Coast of Australia: Bengello Beach, Moruya, NSW, 1972--2022},
+  title={50 Years of {Beach--Foredune} Change on the {Southeastern Coast of Australia}: {Bengello Beach, Moruya, NSW, 1972--2022}},
   author={McLean, R. and Thom, B. and Shen, J. and Oliver, T.},
   journal={Geomorphology},
   volume={439},
@@ -65,7 +59,6 @@ eprint = {
   url = {https://doi.org/10.1016/j.geomorph.2023.108850},
   publisher={Elsevier}
 }
-
 
 @article{vos2023benchmarking,
   title={Benchmarking satellite-derived shoreline mapping algorithms},
@@ -81,7 +74,7 @@ eprint = {
 }
 
 @article{turner2021satellite,
-  title={Satellite optical imagery in Coastal Engineering},
+  title={Satellite optical imagery in {Coastal Engineering}},
   author={Turner, I. L. and Harley, M. D. and Almar, R. and Bergsma, E. W. J.},
   journal={Coastal Engineering},
   volume={167},
@@ -93,7 +86,7 @@ eprint = {
 }
 
 @article{vos2023pacific,
-  title={Pacific shoreline erosion and accretion patterns controlled by El Ni{\~n}o/Southern Oscillation},
+  title={Pacific shoreline erosion and accretion patterns controlled by {El Ni{\~n}o/Southern Oscillation}},
   author={Vos, K. and Harley, M. D. and Turner, I. L. and Splinter, K. D.},
   journal={Nature Geoscience},
   volume={16},
@@ -117,9 +110,8 @@ eprint = {
   url={https://doi.org/10.1016/j.geomorph.2021.107707}
 }
 
-
 @article{warrick2023large,
[…]
