### 2024-06-28 #6683 @2320sharon
@kthyng thank you for looking over the coastseg paper. I fixed the grammer issue in the 3rd paragraph and I believe all the capitalization in the references now matches the original capitalization that was in the paper.bib.
