### 2024-07-01 #6683 @kthyng
I see one last typo in that 3rd paragraph:
* consisting *of* functions
