### 2024-07-01 #6683 @kthyng
Congratulations on your new publication @2320sharon! Many thanks to @cheginit and to reviewers @FlorisCalkoen and @fmemuir for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
